### 2021-12-18 #4008 @whedon
[PRE REVIEW]: TorchMetrics - Measuring Reproducibility in PyTorch
**Submitting author:** @justusschock (<a href="http://orcid.org/0000-0003-0512-3053">Justus Schock</a>)
**Repository:** <a href="https://github.com/PyTorchLightning/metrics" target ="_blank">https://github.com/PyTorchLightning/metrics</a>
**Version:** v0.6.2
