### 2021-12-18 50f1d93
Make input order of ROUGEScore and BERTScore consistent with other NLG metrics
