### 2021-12-18 d94aa36
Move comment from test_bertscore to common test inputs
