### 2021-12-18 0db755c
Apply suggestions from code review
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
