### 2021-12-18 f2e0078
Fix seed for doc + update some tests
