+++ torchmetrics/image/ms_ssim.py
@@ -51,11 +51,12 @@ class MultiScaleSSIM(Metric):
 
     Example:
         >>> from torchmetrics import MultiScaleSSIM
+        >>> torch.manual_seed(42)
         >>> preds = torch.rand([1, 1, 256, 256])
         >>> target = preds * 0.75
         >>> ssim = MultiScaleSSIM()
         >>> ssim(preds, target)
-        tensor(0.9557)
+        tensor(0.9558)
 
     References:
     [1] Multi-Scale Structural Similarity For Image Quality Assessment by Zhou Wang, Eero P. Simoncelli and Alan C.
