+++ torchmetrics/functional/image/ms_ssim.py
@@ -176,10 +176,11 @@ def ms_ssim(
 
     Example:
         >>> from torchmetrics.functional import ms_ssim
+        >>> torch.manual_seed(42)
         >>> preds = torch.rand([1, 1, 256, 256])
         >>> target = preds * 0.75
         >>> ms_ssim(preds, target)
-        tensor(0.9557)
+        tensor(0.9558)
 
     References:
     [1] Multi-Scale Structural Similarity For Image Quality Assessment by Zhou Wang, Eero P. Simoncelli and Alan C.
@@ -191,5 +192,8 @@ def ms_ssim(
         raise ValueError("Argument `betas` is expected to be a tuple of floats.")
     if normalize and normalize not in ("relu", "simple"):
         raise ValueError("Argument `normalize` to be expected either `None` or one of 'relu' or 'simple'")
+    target = target.type(preds.type())
+    print(preds.type())
     preds, target = _ssim_update(preds, target)
+    print(preds.type())
     return _ms_ssim_compute(preds, target, kernel_size, sigma, reduction, data_range, k1, k2, betas, normalize)
