### 2021-12-18 8fa8984
Clean code
