### 2021-12-19 627f8f6
Update doc example + make diff test tractable
