### 2021-12-19 60e5439
torch.manual seed passed directly to torch.rand
