### 2021-12-19 3131b5a
Add test strategy to map
