### 2021-12-19 7f51e99
Fix device type
