### 2021-12-19 27aa6c2
Add test strategy to audio metrics
