+++ tests/audio/test_si_sdr.py
@@ -17,10 +17,11 @@ from functools import partial
 import pytest
 import speechmetrics
 import torch
+from pytest_cases import parametrize_with_cases
 from torch import Tensor
 
 from tests.helpers import seed_all
-from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
+from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester, MetricTesterDDPCases
 from torchmetrics.audio import SI_SDR
 from torchmetrics.functional import si_sdr
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
@@ -77,9 +78,9 @@ speechmetrics_si_sdr_no_zero_mean = partial(speechmetrics_si_sdr, zero_mean=Fals
 class TestSISDR(MetricTester):
     atol = 1e-2
 
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
-    def test_si_sdr(self, preds, target, sk_metric, zero_mean, ddp, dist_sync_on_step):
+    def test_si_sdr(self, preds, target, sk_metric, zero_mean, ddp, dist_sync_on_step, device):
         self.run_class_metric_test(
             ddp,
             preds,
@@ -87,15 +88,18 @@ class TestSISDR(MetricTester):
             SI_SDR,
             sk_metric=partial(average_metric, metric_func=sk_metric),
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             metric_args=dict(zero_mean=zero_mean),
         )
 
-    def test_si_sdr_functional(self, preds, target, sk_metric, zero_mean):
+    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    def test_si_sdr_functional(self, preds, target, sk_metric, zero_mean, device):
         self.run_functional_metric_test(
             preds,
             target,
             si_sdr,
             sk_metric,
+            device=device,
             metric_args=dict(zero_mean=zero_mean),
         )
 
