### 2021-12-19 258da03
Add test strategy to bases
