### 2021-12-19 b4dae4d
Add test strategy to classification
