### 2021-12-19 8439b4a
Add test strategy to image
