### 2021-12-19 30b4ea7
Add test strategy to pairwise
