### 2021-12-19 2f5dc34
Add test strategy to regression
