+++ tests/regression/test_r2.py
@@ -16,10 +16,11 @@ from functools import partial
 
 import pytest
 import torch
+from pytest_cases import parametrize_with_cases
 from sklearn.metrics import r2_score as sk_r2score
 
 from tests.helpers import seed_all
-from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
+from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester, MetricTesterDDPCases
 from torchmetrics.functional import r2_score
 from torchmetrics.regression import R2Score
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
@@ -69,9 +70,9 @@ def _multi_target_sk_metric(preds, target, adjusted, multioutput):
     ],
 )
 class TestR2Score(MetricTester):
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
-    def test_r2(self, adjusted, multioutput, preds, target, sk_metric, num_outputs, ddp, dist_sync_on_step):
+    def test_r2(self, adjusted, multioutput, preds, target, sk_metric, num_outputs, ddp, dist_sync_on_step, device):
         self.run_class_metric_test(
             ddp,
             preds,
@@ -79,16 +80,19 @@ class TestR2Score(MetricTester):
             R2Score,
             partial(sk_metric, adjusted=adjusted, multioutput=multioutput),
             dist_sync_on_step,
+            device=device,
             metric_args=dict(adjusted=adjusted, multioutput=multioutput, num_outputs=num_outputs),
         )
 
-    def test_r2_functional(self, adjusted, multioutput, preds, target, sk_metric, num_outputs):
+    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    def test_r2_functional(self, adjusted, multioutput, preds, target, sk_metric, num_outputs, device):
         # todo: `num_outputs` is unused
         self.run_functional_metric_test(
             preds,
             target,
             r2_score,
             partial(sk_metric, adjusted=adjusted, multioutput=multioutput),
+            device=device,
             metric_args=dict(adjusted=adjusted, multioutput=multioutput),
         )
 

[…]
