### 2021-12-19 2dfe6c8
Add strategy to pairwise
