+++ tests/retrieval/test_map.py
@@ -12,10 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import pytest
+from pytest_cases import parametrize_with_cases
 from sklearn.metrics import average_precision_score as sk_average_precision_score
 from torch import Tensor
 
 from tests.helpers import seed_all
+from tests.helpers.testers import MetricTesterDDPCases
 from tests.retrieval.helpers import (
     RetrievalMetricTester,
     _concat_tests,
@@ -33,7 +35,7 @@ seed_all(42)
 
 
 class TestMAP(RetrievalMetricTester):
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize("ignore_index", [None, 1])  # avoid setting 0, otherwise test with all 0 targets will fail
@@ -47,6 +49,7 @@ class TestMAP(RetrievalMetricTester):
         dist_sync_on_step: bool,
         empty_target_action: str,
         ignore_index: int,
+        device: str,
     ):
         metric_args = dict(empty_target_action=empty_target_action, ignore_index=ignore_index)
 
@@ -58,10 +61,11 @@ class TestMAP(RetrievalMetricTester):
             metric_class=RetrievalMAP,
             sk_metric=sk_average_precision_score,
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             metric_args=metric_args,
         )
 
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize(**_default_metric_class_input_arguments_ignore_index)
@@ -73,6 +77,7 @@ class TestMAP(RetrievalMetricTester):
         target: Tensor,
         dist_sync_on_step: bool,
         empty_target_action: str,
+        device: str,
     ):
         metric_args = dict(empty_target_action=empty_target_action, ignore_index=-100)
 
@@ -84,16 +89,19 @@ class TestMAP(RetrievalMetricTester):
             metric_class=RetrievalMAP,
             sk_metric=sk_average_precision_score,
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             metric_args=metric_args,
         )
 
+    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
     @pytest.mark.parametrize(**_default_metric_functional_input_arguments)
-    def test_functional_metric(self, preds: Tensor, target: Tensor):
+    def test_functional_metric(self, preds: Tensor, target: Tensor, device: str):
         self.run_functional_metric_test(
             preds=preds,
             target=target,
             metric_functional=retrieval_average_precision,
             sk_metric=sk_average_precision_score,
+            device=device,
             metric_args={},
         )
 
