+++ tests/retrieval/test_r_precision.py
@@ -13,9 +13,11 @@
 # limitations under the License.
 import numpy as np
 import pytest
+from pytest_cases import parametrize_with_cases
 from torch import Tensor
 
 from tests.helpers import seed_all
+from tests.helpers.testers import MetricTesterDDPCases
 from tests.retrieval.helpers import (
     RetrievalMetricTester,
     _concat_tests,
@@ -49,7 +51,7 @@ def _r_precision(target: np.ndarray, preds: np.ndarray):
 
 
 class TestRPrecision(RetrievalMetricTester):
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize("ignore_index", [None, 1])  # avoid setting 0, otherwise test with all 0 targets will fail
@@ -63,6 +65,7 @@ class TestRPrecision(RetrievalMetricTester):
         dist_sync_on_step: bool,
         empty_target_action: str,
         ignore_index: int,
+        device: str,
     ):
         metric_args = dict(empty_target_action=empty_target_action, ignore_index=ignore_index)
 
@@ -74,10 +77,11 @@ class TestRPrecision(RetrievalMetricTester):
             metric_class=RetrievalRPrecision,
             sk_metric=_r_precision,
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             metric_args=metric_args,
         )
 
-    @pytest.mark.parametrize("ddp", [True, False])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize(**_default_metric_class_input_arguments_ignore_index)
@@ -89,6 +93,7 @@ class TestRPrecision(RetrievalMetricTester):
         target: Tensor,
         dist_sync_on_step: bool,
         empty_target_action: str,
+        device: str,
     ):
         metric_args = dict(empty_target_action=empty_target_action, ignore_index=-100)
 
@@ -100,16 +105,19 @@ class TestRPrecision(RetrievalMetricTester):
             metric_class=RetrievalRPrecision,
             sk_metric=_r_precision,
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             metric_args=metric_args,
         )
 
+    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
     @pytest.mark.parametrize(**_default_metric_functional_input_arguments)
-    def test_functional_metric(self, preds: Tensor, target: Tensor):
+    def test_functional_metric(self, preds: Tensor, target: Tensor, device: str):
         self.run_functional_metric_test(
             preds=preds,
             target=target,
             metric_functional=retrieval_r_precision,
             sk_metric=_r_precision,
+            device=device,
             metric_args={},
         )
 
