### 2021-12-19 2e9f235
Add test strategy to text
