+++ tests/text/test_wip.py
@@ -2,7 +2,9 @@ from typing import List, Union
 
 import pytest
 from jiwer import wip
+from pytest_cases import parametrize_with_cases
 
+from tests.helpers.testers import MetricTesterDDPCases
 from tests.text.helpers import INPUT_ORDER, TextTester
 from tests.text.inputs import _inputs_error_rate_batch_size_1, _inputs_error_rate_batch_size_2
 from torchmetrics.functional.text.wip import word_information_preserved
@@ -23,9 +25,9 @@ def _compute_wip_metric_jiwer(prediction: Union[str, List[str]], reference: Unio
     ],
 )
 class TestWordInfoPreserved(TextTester):
-    @pytest.mark.parametrize("ddp", [False, True])
+    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
     @pytest.mark.parametrize("dist_sync_on_step", [False, True])
-    def test_wip_class(self, ddp, dist_sync_on_step, preds, targets):
+    def test_wip_class(self, ddp, dist_sync_on_step, preds, targets, device):
 
         self.run_class_metric_test(
             ddp=ddp,
@@ -34,16 +36,19 @@ class TestWordInfoPreserved(TextTester):
             metric_class=WordInfoPreserved,
             sk_metric=_compute_wip_metric_jiwer,
             dist_sync_on_step=dist_sync_on_step,
+            device=device,
             input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
-    def test_wip_functional(self, preds, targets):
+    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    def test_wip_functional(self, preds, targets, device):
 
         self.run_functional_metric_test(
             preds,
             targets,
             metric_functional=word_information_preserved,
             sk_metric=_compute_wip_metric_jiwer,
+            device=device,
             input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
