### 2021-12-19 70bce49
Add test strategy to wrappers
