+++ torchmetrics/image/ms_ssim.py
@@ -43,8 +43,8 @@ class MultiScaleSSIM(Metric):
         betas: Exponent parameters for individual similarities and contrastive sensitivies returned by different image
         resolutions.
         normalize: When MultiScaleSSIM loss is used for training, it is desirable to use normalizes to improve the
-        training stability. This `normalize` argument is out of scope of the original implementation [1], and it is
-        adapted from https://github.com/jorge-pessoa/pytorch-msssim instead.
+            training stability. This `normalize` argument is out of scope of the original implementation [1], and it is
+            adapted from https://github.com/jorge-pessoa/pytorch-msssim instead.
 
     Return:
         Tensor with Multi-Scale SSIM score
