### 2021-12-19 5e76af4
CI: add py3.10 & default py3.8
