+++ .github/workflows/ci_test-full.yml
@@ -28,11 +28,11 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-20.04, macOS-10.15, windows-2019]
-        python-version: [3.6, 3.8, 3.9]
+        python-version: ["3.6", "3.8", "3.10"]
         requires: ['oldest', 'latest']
         exclude:
-          - {python-version: 3.6, requires: 'latest'}
-          - {python-version: 3.9, requires: 'oldest'}
+          - {python-version: "3.6", requires: 'latest'}
+          - {python-version: "3.10", requires: 'oldest'}
     env:
       PYTEST_ARTEFACT: test-results-${{ matrix.os }}-py${{ matrix.python-version }}-${{ matrix.requires }}.xml
       PYTORCH_URL: https://download.pytorch.org/whl/cpu/torch_stable.html
