+++ .github/workflows/ci_install-pkg.yml
@@ -16,7 +16,7 @@ jobs:
       - uses: actions/checkout@master
       - uses: actions/setup-python@v2
         with:
-          python-version: 3.7
+          python-version: 3.8
 
       - name: Check package
         run: |
@@ -41,7 +41,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-20.04, macOS-10.15 , windows-2019]
-        python-version: [3.7]
+        python-version: [3.8]
     timeout-minutes: 10
 
     steps:
