### 2021-12-20 c73202f
Drop some tests ; fix docs issues + add missing docs
