### 2021-12-20 #4008 @Borda
Hello, I guess that one of these authors could be good candidate for review:
- https://towardsdatascience.com/20-popular-machine-learning-metrics-part-1-classification-regression-evaluation-metrics-1ca3e282a2ce
- https://www.analyticsvidhya.com/blog/2019/08/11-important-model-evaluation-error-metrics/
- https://neptune.ai/blog/performance-metrics-in-machine-learning-complete-guide
- https://machinelearningmastery.com/metrics-evaluate-machine-learning-algorithms-python/
