+++ .github/workflows/ci_test-base.yml
@@ -17,6 +17,8 @@ jobs:
       matrix:
         os: [ubuntu-20.04, windows-2019, macOS-10.15]
         python-version: [3.7]
+    env:
+      TRANSFORMERS_CACHE: .cache/huggingface/
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
     timeout-minutes: 20
@@ -63,6 +65,12 @@ jobs:
         pip uninstall -y torchmetrics
         pip list
       shell: bash
+
+    - name: HF cache
+      uses: actions/cache@v2
+      with:
+        path: $TRANSFORMERS_CACHE
+        key: cache-transformaers
 
     - name: Test Package [only]
       run: |
