+++ .github/workflows/ci_test-full.yml
@@ -24,6 +24,8 @@ jobs:
             requires: 'latest'
           - python-version: 3.9
             requires: 'oldest'
+    env:
+      TRANSFORMERS_CACHE: .cache/huggingface/
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
     timeout-minutes: 75
@@ -87,6 +89,12 @@ jobs:
         pip uninstall -y torchmetrics
         pip list
       shell: bash
+
+    - name: HF cache
+      uses: actions/cache@v2
+      with:
+        path: $TRANSFORMERS_CACHE
+        key: cache-transformaers
 
     - name: Tests
       run: |
