### 2021-12-21 e83ede2
Apply suggestions from code review
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
