+++ .github/workflows/ci_test-conda.yml
@@ -66,7 +66,7 @@ jobs:
       uses: actions/cache@v2
       with:
         path: $TRANSFORMERS_CACHE
-        key: cache-transformaers
+        key: cache-transformers
 
     - name: Temp COCO env. fix
       if: matrix.pytorch-version == '1.4'
