+++ .github/workflows/ci_test-base.yml
@@ -70,7 +70,7 @@ jobs:
       uses: actions/cache@v2
       with:
         path: $TRANSFORMERS_CACHE
-        key: cache-transformaers
+        key: cache-transformers
 
     - name: Test Package [only]
       run: |
