### 2021-12-21 65440bc
CI: clean envs
