+++ .github/workflows/ci_test-base.yml
@@ -13,10 +13,12 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
-      # max-parallel: 6
       matrix:
         os: [ubuntu-20.04, windows-2019, macOS-10.15]
         python-version: [3.7]
+    env:
+      PYTEST_ARTEFACT: pytest-results-${{ runner.os }}-${{ matrix.python-version }}
+      PYTORCH_URL: https://download.pytorch.org/whl/cpu/torch_stable.html
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
     timeout-minutes: 20
@@ -57,9 +59,9 @@ jobs:
       run: |
         python --version
         pip --version
-        pip install --requirement requirements.txt --upgrade --find-links https://download.pytorch.org/whl/cpu/torch_stable.html
+        pip install --requirement requirements.txt --upgrade --find-links $PYTORCH_URL
         python ./requirements/adjust-versions.py requirements/image.txt
-        pip install --requirement requirements/devel.txt --upgrade --find-links https://download.pytorch.org/whl/cpu/torch_stable.html
+        pip install --requirement requirements/devel.txt --upgrade --find-links $PYTORCH_URL
         pip uninstall -y torchmetrics
         pip list
       shell: bash
@@ -67,13 +69,13 @@ jobs:
     - name: Test Package [only]
       run: |
         # NOTE: run coverage on tests does not propagare faler status for Win, https://github.com/nedbat/coveragepy/issues/1003
-        python -m pytest torchmetrics -v --cov=torchmetrics --junitxml=junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
+        python -m pytest torchmetrics -v --cov=torchmetrics --junitxml="junit/$PYTEST_ARTEFACT"
 
     - name: Upload pytest test results
       uses: actions/upload-artifact@v2
       with:
-        name: pytest-results-${{ runner.os }}-${{ matrix.python-version }}
-        path: junit/test-results-${{ runner.os }}-${{ matrix.python-version }}.xml
+        name: $PYTEST_ARTEFACT
+        path: "junit/$PYTEST_ARTEFACT"
       if: failure()
 
     - name: Statistics
