+++ .github/workflows/ci_test-full.yml
@@ -25,7 +25,7 @@ jobs:
           - python-version: 3.9
             requires: 'oldest'
     env:
-      PYTEST_ARTEFACT: test-results-${{ runner.os }}-py${{ matrix.python-version }}-${{ matrix.requires }}.xml
+      PYTEST_ARTEFACT: test-results-${{ matrix.os }}-py${{ matrix.python-version }}-${{ matrix.requires }}.xml
       PYTORCH_URL: https://download.pytorch.org/whl/cpu/torch_stable.html
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
