### 2021-12-22 c2a1277
add tuple_of_single_references
