### 2021-12-22 d739287
fix differentiability; stylistic changes
