### 2021-12-22 c917dab
update test to use TextTester correctly
