### 2021-12-22 293af54
CI: clean envs (#692)
