+++ .github/workflows/ci_test-conda.yml
@@ -19,6 +19,7 @@ jobs:
           - python-version: 3.7
             pytorch-version: '1.3'
     env:
+      PYTEST_ARTEFACT: test-conda-py${{ matrix.python-version }}-pt${{ matrix.pytorch-version }}.xml
       TRANSFORMERS_CACHE: .cache/huggingface/
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
@@ -93,13 +94,13 @@ jobs:
     - name: Testing
       run: |
         # NOTE: run coverage on tests does not propagare faler status for Win, https://github.com/nedbat/coveragepy/issues/1003
-        python -m pytest torchmetrics tests -v --durations=35 --junitxml=junit/test-conda-py${{ matrix.python-version }}-pt${{ matrix.pytorch-version }}.xml
+        python -m pytest torchmetrics tests -v --durations=50 --junitxml=$PYTEST_ARTEFACT
       shell: bash -l {0}
 
     - name: Upload pytest test results
       uses: actions/upload-artifact@master
       with:
-        name: test-conda-py${{ matrix.python-version }}-pt${{ matrix.pytorch-version }}.xml
-        path: junit/test-conda-py${{ matrix.python-version }}-pt${{ matrix.pytorch-version }}.xml
+        name: $PYTEST_ARTEFACT
+        path: $PYTEST_ARTEFACT
       # Use always() to always run this step to publish test results when there are test failures
       if: failure()
