### 2021-12-23 0f49516
Add ddp skip condition
