+++ torchmetrics/functional/text/bleu.py
@@ -57,62 +57,62 @@ def _tokenize_fn(sentence: str) -> Sequence[str]:
 
 
 def _bleu_score_update(
-    reference_corpus: Sequence[Sequence[str]],
-    translate_corpus: Sequence[str],
+    prediction_corpus: Sequence[str],
+    target_corpus: Sequence[Sequence[str]],
     numerator: Tensor,
     denominator: Tensor,
-    trans_len: Tensor,
-    ref_len: Tensor,
+    prediction_len: Tensor,
+    target_len: Tensor,
     n_gram: int = 4,
     tokenizer: Callable[[str], Sequence[str]] = _tokenize_fn,
 ) -> Tuple[Tensor, Tensor]:
     """Updates and returns variables required to compute the BLEU score.
 
     Args:
-        reference_corpus: An iterable of iterables of reference corpus
-        translate_corpus: An iterable of machine translated corpus
+        target_corpus: An iterable of iterables of reference corpus
+        prediction_corpus: An iterable of machine translated corpus
         numerator: Numerator of precision score (true positives)
         denominator: Denominator of precision score (true positives + false positives)
-        trans_len: count of words in a candidate translation
-        ref_len: count of words in a reference translation
+        prediction_len: count of words in a candidate prediction
+        target_len: count of words in a reference translation
         n_gram: gram value ranged 1 to 4
         tokenizer: A function that turns sentence into list of words
     """
-    reference_corpus_: Sequence[Sequence[Sequence[str]]] = [
-        [tokenizer(line) if line else [] for line in reference] for reference in reference_corpus
+    target_corpus_: Sequence[Sequence[Sequence[str]]] = [
+        [tokenizer(line) if line else [] for line in target] for target in target_corpus
     ]
-    translate_corpus_: Sequence[Sequence[str]] = [tokenizer(line) if line else [] for line in translate_corpus]
+    prediction_corpus_: Sequence[Sequence[str]] = [tokenizer(line) if line else [] for line in prediction_corpus]
 
-    for (translation, references) in zip(translate_corpus_, reference_corpus_):
-        trans_len += len(translation)
-        ref_len_list = [len(ref) for ref in references]
-        ref_len_diff = [abs(len(translation) - x) for x in ref_len_list]
-        ref_len += ref_len_list[ref_len_diff.index(min(ref_len_diff))]
-        translation_counter: Counter = _count_ngram(translation, n_gram)
-        reference_counter: Counter = Counter()
+    for (prediction, targets) in zip(prediction_corpus_, target_corpus_):
+        prediction_len += len(prediction)
+        target_len_list = [len(ref) for ref in targets]
+        target_len_diff = [abs(len(prediction) - x) for x in target_len_list]
+        target_len += target_len_list[target_len_diff.index(min(target_len_diff))]
+        prediction_counter: Counter = _count_ngram(prediction, n_gram)
+        target_counter: Counter = Counter()
 
-        for ref in references:
-            reference_counter |= _count_ngram(ref, n_gram)
+        for ref in targets:
+            target_counter |= _count_ngram(ref, n_gram)
 
-        ngram_counter_clip = translation_counter & reference_counter
+        ngram_counter_clip = prediction_counter & target_counter
 
         for counter_clip in ngram_counter_clip:
             numerator[len(counter_clip) - 1] += ngram_counter_clip[counter_clip]
 
-        for counter in translation_counter:
-            denominator[len(counter) - 1] += translation_counter[counter]
+        for counter in prediction_counter:
[…]
