+++ tests/classification/test_matthews_corrcoef.py
@@ -14,7 +14,6 @@
 import numpy as np
 import pytest
 import torch
-from pytest_cases import parametrize_with_cases
 from sklearn.metrics import matthews_corrcoef as sk_matthews_corrcoef
 
 from tests.classification.inputs import _input_binary, _input_binary_prob
@@ -102,7 +101,7 @@ def _sk_matthews_corrcoef_multidim_multiclass(preds, target):
     ],
 )
 class TestMatthewsCorrCoef(MetricTester):
-    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_strategy(), MetricTesterDDPCases.cases_strategy())
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     def test_matthews_corrcoef(self, preds, target, sk_metric, num_classes, ddp, dist_sync_on_step, device):
         self.run_class_metric_test(
@@ -119,7 +118,7 @@ class TestMatthewsCorrCoef(MetricTester):
             },
         )
 
-    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_device(), MetricTesterDDPCases.cases_device())
     def test_matthews_corrcoef_functional(self, preds, target, sk_metric, num_classes, device):
         self.run_functional_metric_test(
             preds,
