+++ tests/retrieval/test_ndcg.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 import numpy as np
 import pytest
-from pytest_cases import parametrize_with_cases
 from sklearn.metrics import ndcg_score
 from torch import Tensor
 
@@ -51,7 +50,7 @@ def _ndcg_at_k(target: np.ndarray, preds: np.ndarray, k: int = None):
 
 
 class TestNDCG(RetrievalMetricTester):
-    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_strategy(), MetricTesterDDPCases.cases_strategy())
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize("ignore_index", [None, 3])  # avoid setting 0, otherwise test with all 0 targets will fail
@@ -83,7 +82,7 @@ class TestNDCG(RetrievalMetricTester):
             metric_args=metric_args,
         )
 
-    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_strategy(), MetricTesterDDPCases.cases_strategy())
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     @pytest.mark.parametrize("empty_target_action", ["skip", "neg", "pos"])
     @pytest.mark.parametrize("k", [None, 1, 4, 10])
@@ -113,7 +112,7 @@ class TestNDCG(RetrievalMetricTester):
             metric_args=metric_args,
         )
 
-    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_device(), MetricTesterDDPCases.cases_device())
     @pytest.mark.parametrize(**_default_metric_functional_input_arguments_with_non_binary_target)
     @pytest.mark.parametrize("k", [None, 1, 4, 10])
     def test_functional_metric(self, preds: Tensor, target: Tensor, k: int, device: str):
