+++ tests/classification/test_roc.py
@@ -16,7 +16,6 @@ from functools import partial
 import numpy as np
 import pytest
 import torch
-from pytest_cases import parametrize_with_cases
 from sklearn.metrics import roc_curve as sk_roc_curve
 from torch import tensor
 
@@ -96,7 +95,7 @@ def _sk_roc_multilabel_multidim_prob(preds, target, num_classes=1):
     ],
 )
 class TestROC(MetricTester):
-    @parametrize_with_cases("ddp,device", cases=MetricTesterDDPCases, has_tag="strategy")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_strategy(), MetricTesterDDPCases.cases_strategy())
     @pytest.mark.parametrize("dist_sync_on_step", [True, False])
     def test_roc(self, preds, target, sk_metric, num_classes, ddp, dist_sync_on_step, device):
         self.run_class_metric_test(
@@ -110,7 +109,7 @@ class TestROC(MetricTester):
             metric_args={"num_classes": num_classes},
         )
 
-    @parametrize_with_cases("device", cases=MetricTesterDDPCases, has_tag="device")
+    @pytest.mark.parametrize(MetricTesterDDPCases.name_device(), MetricTesterDDPCases.cases_device())
     def test_roc_functional(self, preds, target, sk_metric, num_classes, device):
         self.run_functional_metric_test(
             preds,

[…]
