### 2021-12-27 2db054f
Simplify id generation
