### 2021-12-28 1e13521
Standardize TER metric
