### 2021-12-28 ddb046d
rename collections.py to metrics_collections.py to avoid clashing with python collections package
