+++ torchmetrics/__init__.py
@@ -40,7 +40,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.collections import MetricCollection  # noqa: E402
+from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.image import PSNR, SSIM  # noqa: E402
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
