+++ tests/text/test_sacre_bleu.py
@@ -31,16 +31,18 @@ if _SACREBLEU_AVAILABLE:
 TOKENIZERS = ("none", "13a", "zh", "intl", "char")
 
 
-def sacrebleu_fn(targets: Sequence[Sequence[str]], preds: Sequence[str], tokenize: str, lowercase: bool) -> Tensor:
+def sacrebleu_fn(
+    predictions: Sequence[str], targets: Sequence[Sequence[str]], tokenize: str, lowercase: bool
+) -> Tensor:
     sacrebleu_fn = BLEU(tokenize=tokenize, lowercase=lowercase)
     # Sacrebleu expects different format of input
     targets = [[target[i] for target in targets] for i in range(len(targets[0]))]
-    sacrebleu_score = sacrebleu_fn.corpus_score(preds, targets).score / 100
+    sacrebleu_score = sacrebleu_fn.corpus_score(predictions, targets).score / 100
     return tensor(sacrebleu_score)
 
 
 @pytest.mark.parametrize(
-    ["preds", "targets"],
+    ["predictions", "targets"],
     [(_inputs_multiple_references.preds, _inputs_multiple_references.targets)],
 )
 @pytest.mark.parametrize(["lowercase"], [(False,), (True,)])
@@ -49,42 +51,42 @@ def sacrebleu_fn(targets: Sequence[Sequence[str]], preds: Sequence[str], tokeniz
 class TestSacreBLEUScore(TextTester):
     @pytest.mark.parametrize("ddp", [False, True])
     @pytest.mark.parametrize("dist_sync_on_step", [False, True])
-    def test_bleu_score_class(self, ddp, dist_sync_on_step, preds, targets, tokenize, lowercase):
+    def test_bleu_score_class(self, ddp, dist_sync_on_step, predictions, targets, tokenize, lowercase):
         metric_args = {"tokenize": tokenize, "lowercase": lowercase}
         original_sacrebleu = partial(sacrebleu_fn, tokenize=tokenize, lowercase=lowercase)
 
         self.run_class_metric_test(
             ddp=ddp,
-            preds=preds,
+            preds=predictions,
             targets=targets,
             metric_class=SacreBLEUScore,
             sk_metric=original_sacrebleu,
             dist_sync_on_step=dist_sync_on_step,
             metric_args=metric_args,
-            input_order=INPUT_ORDER.TARGETS_FIRST,
+            input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
-    def test_bleu_score_functional(self, preds, targets, tokenize, lowercase):
+    def test_bleu_score_functional(self, predictions, targets, tokenize, lowercase):
         metric_args = {"tokenize": tokenize, "lowercase": lowercase}
         original_sacrebleu = partial(sacrebleu_fn, tokenize=tokenize, lowercase=lowercase)
 
         self.run_functional_metric_test(
-            preds,
+            predictions,
             targets,
             metric_functional=sacre_bleu_score,
             sk_metric=original_sacrebleu,
             metric_args=metric_args,
-            input_order=INPUT_ORDER.TARGETS_FIRST,
+            input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
-    def test_bleu_score_differentiability(self, preds, targets, tokenize, lowercase):
+    def test_bleu_score_differentiability(self, predictions, targets, tokenize, lowercase):
         metric_args = {"tokenize": tokenize, "lowercase": lowercase}
 
         self.run_differentiability_test(
-            preds=preds,
+            preds=predictions,
             targets=targets,
             metric_module=SacreBLEUScore,
             metric_functional=sacre_bleu_score,
             metric_args=metric_args,
-            input_order=INPUT_ORDER.TARGETS_FIRST,
+            input_order=INPUT_ORDER.PREDS_FIRST,
         )
