+++ torchmetrics/functional/text/sacre_bleu.py
@@ -39,6 +39,7 @@
 
 
 import re
+import warnings
 from functools import partial
 from typing import Sequence
 
@@ -321,6 +322,10 @@ def sacre_bleu_score(
         [3] Automatic Evaluation of Machine Translation Quality Using Longest Common Subsequence
         and Skip-Bigram Statistics by Chin-Yew Lin and Franz Josef Och `Machine Translation Evolution`_
     """
+    warnings.warn(
+        "Input order of preds and targets were changed to target firsts and predictions \
+                   second in v0.7. Warning will be removed in v0.8"
+    )
     if tokenize not in AVAILABLE_TOKENIZERS:
         raise ValueError(f"Argument `tokenize` expected to be one of {AVAILABLE_TOKENIZERS} but got {tokenize}.")
 
