### 2021-12-28 34f673e
rework and speed up _find_best_gt_match
