### 2021-12-28 c66de8c
fix gpu test and move inputs to gpu
