### 2021-12-28 fc482a7
Update docstrings
