### 2021-12-28 0a48f77
Update docstring + CHANGELOG.md
