+++ torchmetrics/text/sacre_bleu.py
@@ -102,8 +102,8 @@ class SacreBLEUScore(BLEUScore):
             dist_sync_fn=dist_sync_fn,
         )
         warnings.warn(
-            "Input order of preds and targets were changed to target firsts and predictions \
-                       second in v0.7. Warning will be removed in v0.8"
+            "Input order of targets and preds were changed to predictions firsts and targets \
+                    second in v0.7. Warning will be removed in v0.8"
         )
         if tokenize not in AVAILABLE_TOKENIZERS:
             raise ValueError(f"Argument `tokenize` expected to be one of {AVAILABLE_TOKENIZERS} but got {tokenize}.")
