### 2021-12-28 3eb172a
Rename collections.py to metrics_collections.py (#695)
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
