### 2021-12-30 3d1010c
fix mypy issues
