### 2021-12-30 baa4560
fix class_metric_test
