### 2022-01-03 16dfcf1
Apply suggestions from code review
Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>
