### 2022-01-03 98ab7b7
Updating docstring for binary example in AveragePrecision object
