+++ torchmetrics/functional/text/eed.py
@@ -97,23 +97,23 @@ from typing_extensions import Literal
 from torchmetrics.functional.text.helper import _validate_inputs
 
 
-def _distance_between_words(reference_word: str, hypothesis_word: str) -> int:
+def _distance_between_words(preds_word: str, target_word: str) -> int:
     """Distance measure used for substitutions/identity operation. Code adapted from
     https://github.com/rwth-i6/ExtendedEditDistance/blob/master/EED.py.
 
     Args:
-        reference_word: reference word string
-        hypothesis_word: hypothesis word string
+        preds_word: hypothesis word string
+        target_word: reference word string
 
     Return:
         0 for match, 1 for no match
     """
-    return int(reference_word != hypothesis_word)
+    return int(preds_word != target_word)
 
 
 def _eed_function(
-    ref: str,
     hyp: str,
+    ref: str,
     alpha: float = 2.0,
     rho: float = 0.3,
     deletion: float = 0.2,
@@ -123,10 +123,10 @@ def _eed_function(
     https://github.com/rwth-i6/ExtendedEditDistance/blob/master/EED.py.
 
     Args:
-        ref:
-            reference string
         hyp:
             hypothesis string
+        ref:
+            reference string
         alpha:
             optimal jump penalty, penalty for jumps between characters
         rho:
@@ -153,7 +153,7 @@ def _eed_function(
             if i > 0:
                 next_row[i] = min(
                     next_row[i - 1] + deletion,
-                    row[i - 1] + _distance_between_words(ref[w - 1], hyp[i - 1]),
+                    row[i - 1] + _distance_between_words(hyp[i - 1], ref[w - 1]),
                     row[i] + insertion,
                 )
             else:
@@ -252,27 +252,27 @@ def _eed_compute(sentence_level_scores: List[Tensor]) -> Tensor:
 
 
 def _preprocess_sentences(
-    reference_corpus: Sequence[Union[str, Sequence[str]]],
-    hypothesis_corpus: Union[str, Sequence[str]],
+    preds: Union[str, Sequence[str]],
+    target: Sequence[Union[str, Sequence[str]]],
     language: Union[Literal["en"], Literal["ja"]],
-) -> Tuple[Sequence[Union[str, Sequence[str]]], Union[str, Sequence[str]]]:
+) -> Tuple[Union[str, Sequence[str]], Sequence[Union[str, Sequence[str]]]]:
     """Proprocess strings according to language requirements.
 
     Args:
-        reference_corpus: An iterable of iterables of reference corpus.
-        hypothesis_corpus: An iterable of hypothesis corpus.
+        preds: An iterable of hypothesis corpus.
+        target: An iterable of iterables of reference corpus.
         language: Language used in sentences. Only supports English (en) and Japanese (ja) for now. Defaults to en
 
     Return:
-        Tuple of lists that contain the cleaned strings for reference_corpus and hypothesis_corpus
+        Tuple of lists that contain the cleaned strings for target and preds
 
     Raises:
         ValueError: If a different language than 'en" or 'ja' is used
-        ValueError: If length of reference_corpus not equal to length of hypothesis_corpus
+        ValueError: If length of target not equal to length of preds
         ValueError: If objects in reference and hypothesis corpus are not strings
     """
     # sanity checks
-    reference_corpus, hypothesis_corpus = _validate_inputs(reference_corpus, hypothesis_corpus)
+    target, preds = _validate_inputs(hypothesis_corpus=preds, reference_corpus=target)
 
     # preprocess string
     if language == "en":
@@ -282,15 +282,15 @@ def _preprocess_sentences(
     else:
[…]
