### 2022-01-04 ebff078
add: Refactor to avoid duplicate calculations
