### 2022-01-04 481bfe3
Fix Collection kwargs filtering
