+++ tests/bases/test_collections.py
@@ -19,6 +19,8 @@ import torch
 from tests.helpers import seed_all
 from tests.helpers.testers import DummyMetricDiff, DummyMetricSum
 from torchmetrics.metric_collections import MetricCollection
+from torchmetrics import Metric
+from torchmetrics.classification import Accuracy
 
 seed_all(42)
 
