+++ tests/wrappers/test_tracker.py
@@ -17,7 +17,7 @@ import pytest
 import torch
 
 from tests.helpers import seed_all
-from torchmetrics import Accuracy, MeanAbsoluteError, MeanSquaredError, Precision, Recall
+from torchmetrics import Accuracy, MeanAbsoluteError, MeanSquaredError, Precision, Recall, MetricCollection
 from torchmetrics.wrappers import MetricTracker
 
 seed_all(42)
@@ -48,11 +48,15 @@ def test_raises_error_if_increment_not_called(method, method_input):
 @pytest.mark.parametrize(
     "base_metric, metric_input, maximize",
     [
-        (partial(Accuracy, num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
-        (partial(Precision, num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
-        (partial(Recall, num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
-        (MeanSquaredError, (torch.randn(50), torch.randn(50)), False),
-        (MeanAbsoluteError, (torch.randn(50), torch.randn(50)), False),
+        (Accuracy(num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
+        (Precision(num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
+        (Recall(num_classes=10), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
+        (MeanSquaredError(), (torch.randn(50), torch.randn(50)), False),
+        (MeanAbsoluteError(), (torch.randn(50), torch.randn(50)), False),
+        (MetricCollection([Accuracy(num_classes=10), Precision(num_classes=10), Recall(num_classes=10)]), (torch.randint(10, (50,)), torch.randint(10, (50,))), True),
+        (MetricCollection([Accuracy(num_classes=10), Precision(num_classes=10), Recall(num_classes=10)]), (torch.randint(10, (50,)), torch.randint(10, (50,))), [True, True, True]),
+        (MetricCollection([MeanSquaredError(), MeanAbsoluteError()]), (torch.randn(50), torch.randn(50)), False),
+        (MetricCollection([MeanSquaredError(), MeanAbsoluteError()]), (torch.randn(50), torch.randn(50)), [False, False]),
     ],
 )
 def test_tracker(base_metric, metric_input, maximize):
@@ -66,11 +70,26 @@ def test_tracker(base_metric, metric_input, maximize):
             tracker(*metric_input)
 
         val = tracker.compute()
-        assert val != 0.0
+        if isinstance(val, dict):
+            for v in val.values(): 
+                assert v != 0.0
+        else:
+            assert val != 0.0
         assert tracker.n_steps == i + 1
 
     assert tracker.n_steps == 5
-    assert tracker.compute_all().shape[0] == 5
+    all_computed_val = tracker.compute_all()
+    if isinstance(all_computed_val, dict):
+        for v in all_computed_val.values():
+            assert v.shape[0] == 5
+    else:
+        assert all_computed_val == 5
+
     val, idx = tracker.best_metric(return_step=True)
-    assert val != 0.0
-    assert idx in list(range(5))
+    if isinstance(val, dict):
+        for v, i in zip(val.values(), idx.values()):
+            assert v != 0.0
+            assert i in list(range(5))   
+    else:
+        assert val != 0.0
+        assert idx in list(range(5))
