+++ torchmetrics/wrappers/tracker.py
@@ -22,9 +22,10 @@ from torchmetrics.collections import MetricCollection
 
 
 class MetricTracker(nn.ModuleList):
-    """A wrapper class that can help keeping track of a metric over time and implement useful methods. The wrapper
-    implements the standard `update`, `compute`, `reset` methods that just calls corresponding method of the
-    currently tracked metric. However, the following additional methods are provided:
+    """A wrapper class that can help keeping track of a metric or metric collection over time and implement 
+    useful methods. The wrapper implements the standard `update`, `compute`, `reset` methods that just calls 
+    corresponding method of the currently tracked metric. 
+    However, the following additional methods are provided:
 
         -``MetricTracker.n_steps``: number of metrics being tracked
 
@@ -63,28 +64,28 @@ class MetricTracker(nn.ModuleList):
         tensor([0.1120, 0.0880, 0.1260, 0.0800, 0.1020])
         
     Example (multiple metrics using MetricCollection):
-        >>> from torchmetrics import MetricTracker, MetricCollection, MeanSquaredError, R2Score
+        >>> from torchmetrics import MetricTracker, MetricCollection, MeanSquaredError, ExplainedVariance
         >>> _ = torch.manual_seed(42)
-        >>> tracker = MetricTracker(MetricCollection([MeanSquaredError(), R2Score()]), maximize=[False, True])
+        >>> tracker = MetricTracker(MetricCollection([MeanSquaredError(), ExplainedVariance()]), maximize=[False, True])
         >>> for epoch in range(5):
         ...     tracker.increment()
         ...     for batch_idx in range(5):
         ...         preds, target = torch.randn(100), torch.randn(100)
         ...         tracker.update(preds, target)
         ...     print(f"current stats={tracker.compute()}")  # doctest: +NORMALIZE_WHITESPACE
-        current stats={'MeanSquaredError': tensor(1.8218), 'MeanAbsoluteError': tensor(1.0769)}
-        current stats={'MeanSquaredError': tensor(2.0268), 'MeanAbsoluteError': tensor(1.1410)}
-        current stats={'MeanSquaredError': tensor(1.9491), 'MeanAbsoluteError': tensor(1.1089)}
-        current stats={'MeanSquaredError': tensor(1.9800), 'MeanAbsoluteError': tensor(1.1353)}
-        current stats={'MeanSquaredError': tensor(2.2481), 'MeanAbsoluteError': tensor(1.2001)}
+        current stats={'MeanSquaredError': tensor(1.8218), 'ExplainedVariance': tensor(-0.8969)}
+        current stats={'MeanSquaredError': tensor(2.0268), 'ExplainedVariance': tensor(-1.0206)}
+        current stats={'MeanSquaredError': tensor(1.9491), 'ExplainedVariance': tensor(-0.8298)}
+        current stats={'MeanSquaredError': tensor(1.9800), 'ExplainedVariance': tensor(-0.9199)}
+        current stats={'MeanSquaredError': tensor(2.2481), 'ExplainedVariance': tensor(-1.1622)}
         >>> best_res, which_epoch = tracker.best_metric(return_step=True)
         >>> best_res
-        {'MeanSquaredError': 1.8218141794204712, 'MeanAbsoluteError': 1.07692551612854}
+        {'MeanSquaredError': 1.8218144178390503, 'ExplainedVariance': -0.8297995328903198}
         >>> which_epoch
-        {'MeanSquaredError': 0, 'MeanAbsoluteError': 0}
+        {'MeanSquaredError': 0, 'ExplainedVariance': 2}
         >>> tracker.compute_all()  # doctest: +NORMALIZE_WHITESPACE
[…]
