### 2022-01-04 83976a8
use length comparison instead of object comparison
