### 2022-01-04 a66ee8a
Unify the input order for text (NLG) metrics - BLEU, SacreBLEU, TER, CHRF (#696)
* Standardize BLEU and CHRF
* Standardize TER metric
* Standardize SacreBLEU
* Add warnings for breaking change files
* Update docstring + CHANGELOG.md
* Update order and keep naming unchanged
* Apply suggestions from code review

Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
