### 2022-01-04 8dd7d91
Fix Collection kwargs filtering (#707)
* Fix Collection kwargs filtering
* Update test_collections.py
* Update CHANGELOG.md

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
