### 2022-01-04 4171140
Updated docs in PrecisionRecallCurve and AveragePrecision (#706)
* Updated documentation in PrecisionRecallCurve object to accurately reflect binary prediction values
* Updating docstring for binary example in AveragePrecision object
