+++ torchmetrics/classification/precision_recall_curve.py
@@ -53,16 +53,16 @@ class PrecisionRecallCurve(Metric):
 
     Example (binary case):
         >>> from torchmetrics import PrecisionRecallCurve
-        >>> pred = torch.tensor([0, 1, 2, 3])
+        >>> pred = torch.tensor([0, 0.1, 0.8, 0.4])
         >>> target = torch.tensor([0, 1, 1, 0])
         >>> pr_curve = PrecisionRecallCurve(pos_label=1)
         >>> precision, recall, thresholds = pr_curve(pred, target)
         >>> precision
-        tensor([0.6667, 0.5000, 0.0000, 1.0000])
+        tensor([0.6667, 0.5000, 1.0000, 1.0000])
         >>> recall
-        tensor([1.0000, 0.5000, 0.0000, 0.0000])
+        tensor([1.0000, 0.5000, 0.5000, 0.0000])
         >>> thresholds
-        tensor([1, 2, 3])
+        tensor([0.1000, 0.4000, 0.8000])
 
     Example (multiclass case):
         >>> pred = torch.tensor([[0.75, 0.05, 0.05, 0.05, 0.05],
