### 2022-01-04 97f1250
signal_distortion_ratio
