+++ torchmetrics/functional/audio/sdr.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 from typing import Optional
+from warnings import warn
 
 import torch
 
@@ -46,7 +47,7 @@ from torchmetrics.utilities import rank_zero_warn
 from torchmetrics.utilities.checks import _check_same_shape
 
 
-def sdr(
+def signal_distortion_ratio(
     preds: Tensor,
     target: Tensor,
     use_cg_iter: Optional[int] = None,
@@ -86,18 +87,18 @@ def sdr(
         sdr value of shape ``[...]``
 
     Example:
-        >>> from torchmetrics.functional.audio import sdr
+        >>> from torchmetrics.functional.audio import signal_distortion_ratio
         >>> import torch
         >>> g = torch.manual_seed(1)
         >>> preds = torch.randn(8000)
         >>> target = torch.randn(8000)
-        >>> sdr(preds, target)
+        >>> signal_distortion_ratio(preds, target)
         tensor(-12.0589)
         >>> # use with pit
         >>> from torchmetrics.functional.audio import pit
         >>> preds = torch.randn(4, 2, 8000)  # [batch, spk, time]
         >>> target = torch.randn(4, 2, 8000)
-        >>> best_metric, best_perm = pit(preds, target, sdr, 'max')
+        >>> best_metric, best_perm = pit(preds, target, signal_distortion_ratio, 'max')
         >>> best_metric
         tensor([-11.6375, -11.4358, -11.7148, -11.6325])
         >>> best_perm
@@ -192,3 +193,31 @@ def sdr(
     ratio = coh / (1 - coh)
     sdr_val = 10.0 * torch.log10(ratio)
     return sdr_val
+
+
+def sdr(
+    preds: Tensor,
+    target: Tensor,
+    use_cg_iter: Optional[int] = None,
+    filter_length: int = 512,
+    zero_mean: bool = False,
+    load_diag: Optional[float] = None,
+) -> Tensor:
+    r"""Signal to Distortion Ratio (SDR)
+
+    .. deprecated:: v0.7
+        Use :func:`torchmetrics.functional.signal_distortion_ratio`. Will be removed in v0.8.
+
+    Example:
+        >>> import torch
+        >>> g = torch.manual_seed(1)
+        >>> sdr(torch.randn(8000), torch.randn(8000))
+        tensor(-12.0589)
+    """
+    warn("`sdr` was renamed to `signal_distortion_ratio` in v0.7 and it will be removed in v0.8", DeprecationWarning)
+    return signal_distortion_ratio(preds,
+    target,
+    use_cg_iter,
+    filter_length,
+    zero_mean,
+    load_diag)
