+++ torchmetrics/functional/__init__.py
@@ -15,7 +15,7 @@ from torchmetrics.functional.audio.pit import pit, pit_permutate
 from torchmetrics.functional.audio.sdr import sdr
 from torchmetrics.functional.audio.si_sdr import si_sdr
 from torchmetrics.functional.audio.si_snr import si_snr
-from torchmetrics.functional.audio.snr import snr
+from torchmetrics.functional.audio.snr import snr, signal_noise_ratio
 from torchmetrics.functional.classification.accuracy import accuracy
 from torchmetrics.functional.classification.auc import auc
 from torchmetrics.functional.classification.auroc import auroc
@@ -129,6 +129,7 @@ __all__ = [
     "rouge_score",
     "sacre_bleu_score",
     "sdr",
+    "signal_noise_ratio",
     "si_sdr",
     "si_snr",
     "snr",
