+++ torchmetrics/functional/audio/__init__.py
@@ -15,4 +15,4 @@ from torchmetrics.functional.audio.pit import pit, pit_permutate  # noqa: F401
 from torchmetrics.functional.audio.sdr import sdr  # noqa: F401
 from torchmetrics.functional.audio.si_sdr import si_sdr  # noqa: F401
 from torchmetrics.functional.audio.si_snr import si_snr  # noqa: F401
-from torchmetrics.functional.audio.snr import snr  # noqa: F401
+from torchmetrics.functional.audio.snr import signal_noise_ratio, snr  # noqa: F401
