+++ tests/audio/test_snr.py
@@ -23,18 +23,18 @@ from torch import Tensor
 from tests.helpers import seed_all
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
 from torchmetrics.audio import SNR
-from torchmetrics.functional import snr
+from torchmetrics.functional import signal_noise_ratio
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
 
 seed_all(42)
 
-Time = 100
+TIME = 100
 
 Input = namedtuple("Input", ["preds", "target"])
 
 inputs = Input(
-    preds=torch.rand(NUM_BATCHES, BATCH_SIZE, 1, Time),
-    target=torch.rand(NUM_BATCHES, BATCH_SIZE, 1, Time),
+    preds=torch.rand(NUM_BATCHES, BATCH_SIZE, 1, TIME),
+    target=torch.rand(NUM_BATCHES, BATCH_SIZE, 1, TIME),
 )
 
 
@@ -96,14 +96,14 @@ class TestSNR(MetricTester):
         self.run_functional_metric_test(
             preds,
             target,
-            snr,
+            signal_noise_ratio,
             sk_metric,
             metric_args=dict(zero_mean=zero_mean),
         )
 
     def test_snr_differentiability(self, preds, target, sk_metric, zero_mean):
         self.run_differentiability_test(
-            preds=preds, target=target, metric_module=SNR, metric_functional=snr, metric_args={"zero_mean": zero_mean}
+            preds=preds, target=target, metric_module=SNR, metric_functional=signal_noise_ratio, metric_args={"zero_mean": zero_mean}
         )
 
     @pytest.mark.skipif(
@@ -115,7 +115,7 @@ class TestSNR(MetricTester):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires cuda")
     def test_snr_half_gpu(self, preds, target, sk_metric, zero_mean):
         self.run_precision_test_gpu(
-            preds=preds, target=target, metric_module=SNR, metric_functional=snr, metric_args={"zero_mean": zero_mean}
+            preds=preds, target=target, metric_module=SNR, metric_functional=signal_noise_ratio, metric_args={"zero_mean": zero_mean}
         )
 
 
