### 2022-01-05 8d39145
ScaleInvariantSNR
