+++ tests/audio/test_snr.py
@@ -103,7 +103,11 @@ class TestSNR(MetricTester):
 
     def test_snr_differentiability(self, preds, target, sk_metric, zero_mean):
         self.run_differentiability_test(
-            preds=preds, target=target, metric_module=SNR, metric_functional=signal_noise_ratio, metric_args={"zero_mean": zero_mean}
+            preds=preds,
+            target=target,
+            metric_module=SNR,
+            metric_functional=signal_noise_ratio,
+            metric_args={"zero_mean": zero_mean},
         )
 
     @pytest.mark.skipif(
@@ -115,7 +119,11 @@ class TestSNR(MetricTester):
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires cuda")
     def test_snr_half_gpu(self, preds, target, sk_metric, zero_mean):
         self.run_precision_test_gpu(
-            preds=preds, target=target, metric_module=SNR, metric_functional=signal_noise_ratio, metric_args={"zero_mean": zero_mean}
+            preds=preds,
+            target=target,
+            metric_module=SNR,
+            metric_functional=signal_noise_ratio,
+            metric_args={"zero_mean": zero_mean},
         )
 
 
