+++ tests/audio/test_si_snr.py
@@ -93,7 +93,12 @@ class TestSISNR(MetricTester):
         )
 
     def test_si_snr_differentiability(self, preds, target, sk_metric):
-        self.run_differentiability_test(preds=preds, target=target, metric_module=ScaleInvariantSNR, metric_functional=scale_invariant_signal_noise_ratio)
+        self.run_differentiability_test(
+            preds=preds,
+            target=target,
+            metric_module=ScaleInvariantSNR,
+            metric_functional=scale_invariant_signal_noise_ratio,
+        )
 
     @pytest.mark.skipif(
         not _TORCH_GREATER_EQUAL_1_6, reason="half support of core operations on not support before pytorch v1.6"
@@ -103,7 +108,12 @@ class TestSISNR(MetricTester):
 
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires cuda")
     def test_si_snr_half_gpu(self, preds, target, sk_metric):
-        self.run_precision_test_gpu(preds=preds, target=target, metric_module=ScaleInvariantSNR, metric_functional=scale_invariant_signal_noise_ratio)
+        self.run_precision_test_gpu(
+            preds=preds,
+            target=target,
+            metric_module=ScaleInvariantSNR,
+            metric_functional=scale_invariant_signal_noise_ratio,
+        )
 
 
 def test_error_on_different_shape(metric_class=ScaleInvariantSNR):
