+++ torchmetrics/functional/__init__.py
@@ -72,7 +72,7 @@ from torchmetrics.functional.text.mer import match_error_rate
 from torchmetrics.functional.text.rouge import rouge_score
 from torchmetrics.functional.text.sacre_bleu import sacre_bleu_score
 from torchmetrics.functional.text.squad import squad
-from torchmetrics.functional.text.ter import ter
+from torchmetrics.functional.text.ter import translation_edit_rate
 from torchmetrics.functional.text.wer import wer
 from torchmetrics.functional.text.wil import word_information_lost
 from torchmetrics.functional.text.wip import word_information_preserved
@@ -138,7 +138,7 @@ __all__ = [
     "ssim",
     "stat_scores",
     "symmetric_mean_absolute_percentage_error",
-    "ter",
+    "translation_edit_rate",
     "wer",
     "char_error_rate",
     "match_error_rate",
