+++ tests/text/test_wer.py
@@ -11,7 +11,7 @@ if _JIWER_AVAILABLE:
 else:
     compute_measures = Callable
 
-from torchmetrics.functional.text.wer import wer
+from torchmetrics.functional.text.wer import word_error_rate
 from torchmetrics.text.wer import WER
 
 
@@ -47,7 +47,7 @@ class TestWER(TextTester):
         self.run_functional_metric_test(
             preds,
             targets,
-            metric_functional=wer,
+            metric_functional=word_error_rate,
             sk_metric=_compute_wer_metric_jiwer,
             input_order=INPUT_ORDER.PREDS_FIRST,
         )
@@ -58,6 +58,6 @@ class TestWER(TextTester):
             preds=preds,
             targets=targets,
             metric_module=WER,
-            metric_functional=wer,
+            metric_functional=word_error_rate,
             input_order=INPUT_ORDER.PREDS_FIRST,
         )
