### 2022-01-05 a30dc36
bugfix: change dtype of confmat to int64
fixes slightly varying ground truth counts (issue #708)
