+++ tests/text/test_ter.py
@@ -7,7 +7,7 @@ from torch import Tensor, tensor
 from tests.text.helpers import INPUT_ORDER, TextTester
 from tests.text.inputs import _inputs_multiple_references, _inputs_single_sentence_multiple_references
 from torchmetrics.functional.text.ter import translation_edit_rate
-from torchmetrics.text.ter import TER
+from torchmetrics.text.ter import TranslationEditRate
 from torchmetrics.utilities.imports import _SACREBLEU_AVAILABLE
 
 if _SACREBLEU_AVAILABLE:
@@ -71,7 +71,7 @@ class TestTER(TextTester):
             ddp=ddp,
             preds=preds,
             targets=targets,
-            metric_class=TER,
+            metric_class=TranslationEditRate,
             sk_metric=nltk_metric,
             dist_sync_on_step=dist_sync_on_step,
             metric_args=metric_args,
@@ -113,7 +113,7 @@ class TestTER(TextTester):
         self.run_differentiability_test(
             preds=preds,
             targets=targets,
-            metric_module=TER,
+            metric_module=TranslationEditRate,
             metric_functional=translation_edit_rate,
             metric_args=metric_args,
             input_order=INPUT_ORDER.PREDS_FIRST,
@@ -127,7 +127,7 @@ def test_ter_empty_functional():
 
 
 def test_ter_empty_class():
-    ter_metric = TER()
+    ter_metric = TranslationEditRate()
     hyp = []
     ref = [[]]
     assert ter_metric(hyp, ref) == tensor(0.0)
@@ -140,7 +140,7 @@ def test_ter_empty_with_non_empty_hyp_functional():
 
 
 def test_ter_empty_with_non_empty_hyp_class():
-    ter_metric = TER()
+    ter_metric = TranslationEditRate()
     hyp = ["python"]
     ref = [[]]
     assert ter_metric(hyp, ref) == tensor(0.0)
@@ -154,7 +154,7 @@ def test_ter_return_sentence_level_score_functional():
 
 
 def test_ter_return_sentence_level_class():
-    ter_metric = TER(return_sentence_level_score=True)
+    ter_metric = TranslationEditRate(return_sentence_level_score=True)
     hyp = _inputs_single_sentence_multiple_references.preds
     ref = _inputs_single_sentence_multiple_references.targets
     _, sentence_ter = ter_metric(hyp, ref)
