+++ torchmetrics/__init__.py
@@ -67,7 +67,6 @@ from torchmetrics.retrieval import (  # noqa: E402
     RetrievalRPrecision,
 )
 from torchmetrics.text import (  # noqa: E402
-    TranslationEditRate,
     WER,
     BLEUScore,
     CharErrorRate,
@@ -75,6 +74,7 @@ from torchmetrics.text import (  # noqa: E402
     MatchErrorRate,
     SacreBLEUScore,
     SQuAD,
+    TranslationEditRate,
     WordInfoLost,
     WordInfoPreserved,
 )
