+++ torchmetrics/text/__init__.py
@@ -18,6 +18,6 @@ from torchmetrics.text.mer import MatchErrorRate  # noqa: F401
 from torchmetrics.text.sacre_bleu import SacreBLEUScore  # noqa: F401
 from torchmetrics.text.squad import SQuAD  # noqa: F401
 from torchmetrics.text.ter import TER  # noqa: F401
-from torchmetrics.text.wer import WordErrorRate, WER  # noqa: F401
+from torchmetrics.text.wer import WER, WordErrorRate  # noqa: F401
 from torchmetrics.text.wil import WordInfoLost  # noqa: F401
 from torchmetrics.text.wip import WordInfoPreserved  # noqa: F401
