### 2022-01-05 55db7c7
Apply suggestions: ms_ssim -> multiscale_ssim + typing
