### 2022-01-05 5953668
rename eed and EED to extended_edit_distance and ExtendedEditDistance, respectively
