+++ docs/source/links.rst
@@ -76,4 +76,4 @@
 .. _chrF score: https://aclanthology.org/W15-3049.pdf
 .. _chrF++ score: https://aclanthology.org/W17-4770.pdf
 .. _TER: https://aclanthology.org/2006.amta-papers.25.pdf
-.. _EED: https://aclanthology.org/W19-5359.pdf
+.. _ExtendedEditDistance: https://aclanthology.org/W19-5359.pdf

[…]
