+++ torchmetrics/functional/__init__.py
@@ -68,7 +68,7 @@ from torchmetrics.functional.text.bert import bert_score
 from torchmetrics.functional.text.bleu import bleu_score
 from torchmetrics.functional.text.cer import char_error_rate
 from torchmetrics.functional.text.chrf import chrf_score
-from torchmetrics.functional.text.eed import extended_edit_distance
+from torchmetrics.functional.text.eed import extended_edit_distance  # noqa: F401
 from torchmetrics.functional.text.mer import match_error_rate
 from torchmetrics.functional.text.rouge import rouge_score
 from torchmetrics.functional.text.sacre_bleu import sacre_bleu_score
