### 2022-01-05 67203f1
use extended_edit_distance at functional/__init__.py
