+++ torchmetrics/functional/__init__.py
@@ -92,8 +92,8 @@ __all__ = [
     "cosine_similarity",
     "tweedie_deviance_score",
     "dice_score",
-    "eed",
     "explained_variance",
+    "extended_edit_distance",
     "f1",
     "fbeta",
     "hamming_distance",
