### 2022-01-05 383f985
update functional.rst
