### 2022-01-05 0480dfc
Removed input_order from text unit tests
