+++ tests/text/test_wip.py
@@ -3,7 +3,7 @@ from typing import List, Union
 import pytest
 from jiwer import wip
 
-from tests.text.helpers import INPUT_ORDER, TextTester
+from tests.text.helpers import TextTester
 from tests.text.inputs import _inputs_error_rate_batch_size_1, _inputs_error_rate_batch_size_2
 from torchmetrics.functional.text.wip import word_information_preserved
 from torchmetrics.text.wip import WordInfoPreserved
@@ -34,7 +34,6 @@ class TestWordInfoPreserved(TextTester):
             metric_class=WordInfoPreserved,
             sk_metric=_compute_wip_metric_jiwer,
             dist_sync_on_step=dist_sync_on_step,
-            input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
     def test_wip_functional(self, preds, targets):
@@ -44,7 +43,6 @@ class TestWordInfoPreserved(TextTester):
             targets,
             metric_functional=word_information_preserved,
             sk_metric=_compute_wip_metric_jiwer,
-            input_order=INPUT_ORDER.PREDS_FIRST,
         )
 
     def test_wip_differentiability(self, preds, targets):
@@ -54,5 +52,4 @@ class TestWordInfoPreserved(TextTester):
             targets=targets,
             metric_module=WordInfoPreserved,
             metric_functional=word_information_preserved,
-            input_order=INPUT_ORDER.PREDS_FIRST,
         )
