### 2022-01-05 89f603b
Removed input_order from text unit tests (#717)
