### 2022-01-05 6e690be
Removed input_order from text unit tests (#717)
