### 2022-01-06 30e33c2
rename CorrCoef (#710)
* rename CorrCoef
* docs
