+++ torchmetrics/__init__.py
@@ -34,6 +34,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     JaccardIndex,
     KLDivergence,
     MatthewsCorrcoef,
+    MatthewsCorrCoef,
     Precision,
     PrecisionRecallCurve,
     Recall,
@@ -51,8 +52,10 @@ from torchmetrics.regression import (  # noqa: E402
     MeanSquaredError,
     MeanSquaredLogError,
     PearsonCorrcoef,
+    PearsonCorrCoef,
     R2Score,
     SpearmanCorrcoef,
+    SpearmanCorrCoef,
     SymmetricMeanAbsolutePercentageError,
     TweedieDevianceScore,
 )
@@ -106,6 +109,7 @@ __all__ = [
     "JaccardIndex",
     "KLDivergence",
     "MatthewsCorrcoef",
+    "MatthewsCorrCoef",
     "MaxMetric",
     "MeanAbsoluteError",
     "MeanAbsolutePercentageError",
@@ -119,6 +123,7 @@ __all__ = [
     "MinMetric",
     "MultioutputWrapper",
     "PearsonCorrcoef",
+    "PearsonCorrCoef",
     "PIT",
     "Precision",
     "PrecisionRecallCurve",
@@ -140,6 +145,7 @@ __all__ = [
     "SI_SNR",
     "SNR",
     "SpearmanCorrcoef",
+    "SpearmanCorrCoef",
     "Specificity",
     "SQuAD",
     "SSIM",
