### 2022-01-06 5c18b6b
refactor: ter (#713)
* translation_edit_rate
* TranslationEditRate
* docs
