### 2022-01-06 aa05557
refactor: wer (#714)
* word_error_rate
* WER

Co-authored-by: Justus Schock <12886177+justusschock@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
