+++ torchmetrics/image/kid.py
@@ -111,7 +111,7 @@ class KID(Metric):
         coef:
             Bias term in the polynomial kernel.
         reset_real_features: Whether to also reset the real features. Since in many cases the real dataset does not
-            change, the features can cached them to avoid recomputing them which is costly. Set this to ``False`` if 
+            change, the features can cached them to avoid recomputing them which is costly. Set this to ``False`` if
             your dataset does not change.
         compute_on_step:
             Forward only calls ``update()`` and return ``None`` if this is set to ``False``.
@@ -242,7 +242,7 @@ class KID(Metric):
         if reset_real_features:
             exclude_states = ()
         else:
-            exclude_states = ('real_features',)
+            exclude_states = ("real_features",)
 
         self._reset_excluded_states = exclude_states
 
@@ -289,4 +289,4 @@ class KID(Metric):
         return kid_scores.mean(), kid_scores.std(unbiased=False)
 
     def reset(self, exclude_states: Optional[Sequence[str]] = None) -> None:
-        super().reset(set([*self._reset_excluded_states, *exclude_states]))
+        super().reset({*self._reset_excluded_states, *exclude_states})
