### 2022-01-06 e6f1736
[WIP] Unify some text metrics
