+++ torchmetrics/text/wip.py
@@ -63,8 +63,8 @@ class WordInfoPreserved(Metric):
     is_differentiable = False
     higher_is_better = False
     errors: Tensor
-    target_total: Tensor
     preds_total: Tensor
+    target_total: Tensor
 
     def __init__(
         self,
@@ -84,7 +84,7 @@ class WordInfoPreserved(Metric):
         self.add_state("preds_total", tensor(0.0), dist_reduce_fx="sum")
 
     def update(self, preds: Union[str, List[str]], target: Union[str, List[str]]) -> None:  # type: ignore
-        """Store references/predictions for computing word Information Preserved scores.
+        """Store predictions/references for computing word Information Preserved scores.
 
         Args:
             preds: Transcription(s) to score as a string or list of strings
