+++ tests/text/test_bertscore.py
@@ -25,7 +25,7 @@ preds = [
     "The victim's brother said he cannot imagine anyone who would want to harm him,\"Finally, it went uphill again at "
     'him."',
 ]
-refs = [
+targets = [
     "28-Year-Old Chef Found Dead at San Francisco Mall",
     "A 28-year-old chef who had recently moved to San Francisco was found dead in the stairwell of a local mall this "
     "week.",
@@ -39,9 +39,9 @@ _METRICS = ["precision", "recall", "f1"]
 MODEL_NAME = "albert-base-v2"
 
 
-def _assert_list(preds: Any, refs: Any, threshold: float = 1e-8):
+def _assert_list(preds: Any, targets: Any, threshold: float = 1e-8):
     """Assert two lists are equal."""
-    assert np.allclose(preds, refs, atol=threshold, equal_nan=True)
+    assert np.allclose(preds, targets, atol=threshold, equal_nan=True)
 
 
 def _parse_original_bert_score(score: torch.Tensor) -> Dict[str, List[float]]:
@@ -51,21 +51,21 @@ def _parse_original_bert_score(score: torch.Tensor) -> Dict[str, List[float]]:
 
 
 preds_batched = [preds[0:2], preds[2:]]
-refs_batched = [refs[0:2], refs[2:]]
+targets_batched = [targets[0:2], targets[2:]]
 
 
 @pytest.mark.parametrize(
-    "preds,refs",
-    [(preds, refs)],
+    "preds,targets",
+    [(preds, targets)],
 )
 @pytest.mark.skipif(not _BERTSCORE_AVAILABLE, reason="test requires bert_score")
-def test_score_fn(preds, refs):
+def test_score_fn(preds, targets):
     """Tests for functional."""
-    original_score = original_bert_score(preds, refs, model_type=MODEL_NAME, num_layers=8, idf=False, batch_size=3)
+    original_score = original_bert_score(preds, targets, model_type=MODEL_NAME, num_layers=8, idf=False, batch_size=3)
     original_score = _parse_original_bert_score(original_score)
 
     metrics_score = metrics_bert_score(
-        preds, refs, model_name_or_path=MODEL_NAME, num_layers=8, idf=False, batch_size=3
+        preds, targets, model_name_or_path=MODEL_NAME, num_layers=8, idf=False, batch_size=3
     )
 
     for metric in _METRICS:
@@ -73,17 +73,17 @@ def test_score_fn(preds, refs):
 
 
 @pytest.mark.parametrize(
-    "preds,refs",
-    [(preds, refs)],
+    "preds,targets",
+    [(preds, targets)],
 )
 @pytest.mark.skipif(not _BERTSCORE_AVAILABLE, reason="test requires bert_score")
-def test_score_fn_with_idf(preds, refs):
+def test_score_fn_with_idf(preds, targets):
     """Tests for functional with IDF rescaling."""
-    original_score = original_bert_score(preds, refs, model_type=MODEL_NAME, num_layers=12, idf=True, batch_size=3)
+    original_score = original_bert_score(preds, targets, model_type=MODEL_NAME, num_layers=12, idf=True, batch_size=3)
     original_score = _parse_original_bert_score(original_score)
 
     metrics_score = metrics_bert_score(
-        preds, refs, model_name_or_path=MODEL_NAME, num_layers=12, idf=True, batch_size=3
+        preds, targets, model_name_or_path=MODEL_NAME, num_layers=12, idf=True, batch_size=3
     )
 
     for metric in _METRICS:
@@ -91,17 +91,19 @@ def test_score_fn_with_idf(preds, refs):
 
 
 @pytest.mark.parametrize(
-    "preds,refs",
-    [(preds, refs)],
+    "preds,targets",
+    [(preds, targets)],
 )
[…]
