+++ .github/workflows/ci_docker.yml
@@ -13,6 +13,10 @@ on: # Trigger the workflow on push or pull request, but only for the master bran
       - ".github/workflows/*docker*.yml"
       - "setup.py"
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+#  cancel-in-progress: true
+
 jobs:
   build-PL:
     runs-on: ubuntu-20.04
