### 2022-01-06 f6a8ca9
fast test & print info
