+++ .github/workflows/ci_test-conda.yml
@@ -85,16 +85,20 @@ jobs:
         pip --version
         python ./requirements/adjust-versions.py requirements.txt
         python ./.github/prune-packages.py requirements/image.txt torchvision
-        pip install --requirement requirements.txt --quiet
-        pip install --requirement requirements/test.txt --quiet
+        pip install --requirement requirements.txt 
+        pip install --requirement requirements/test.txt 
+        echo "pip list"
         pip list
+        echo "conda list"
         python -c "from torch import __version__ as ver; assert '.'.join(ver.split('.')[:2]) == '${{ matrix.pytorch-version }}', ver"
+        echo "from pesq import pesq"
+        python -c "from pesq import pesq"
       shell: bash -l {0}
 
     - name: Testing
       run: |
         # NOTE: run coverage on tests does not propagare faler status for Win, https://github.com/nedbat/coveragepy/issues/1003
-        python -m pytest torchmetrics tests -v --durations=50 --junitxml=$PYTEST_ARTEFACT
+        python -m pytest torchmetrics tests -v --durations=2 --junitxml=$PYTEST_ARTEFACT
       shell: bash -l {0}
 
     - name: Upload pytest test results
