+++ .github/workflows/ci_test-conda.yml
@@ -85,8 +85,8 @@ jobs:
         pip --version
         python ./requirements/adjust-versions.py requirements.txt
         python ./.github/prune-packages.py requirements/image.txt torchvision
-        pip install --requirement requirements.txt 
-        pip install --requirement requirements/test.txt 
+        pip install --requirement requirements.txt
+        pip install --requirement requirements/test.txt
         echo "pip list"
         pip list
         echo "conda list"
