+++ .github/workflows/ci_docker.yml
@@ -15,7 +15,7 @@ on: # Trigger the workflow on push or pull request, but only for the master bran
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
-#  cancel-in-progress: true
+  cancel-in-progress: true
 
 jobs:
   build-PL:
