### 2022-01-06 4e5f10d
Apply SkafteNicki's suggestions
