+++ .github/workflows/ci_docker.yml
@@ -14,7 +14,7 @@ on: # Trigger the workflow on push or pull request, but only for the master bran
       - "setup.py"
 
 concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
+  group: ${{ github.workflow }}-${{ github.ref }}-${{ github.head_ref }}
   cancel-in-progress: true
 
 jobs:
