+++ .github/workflows/ci_test-conda.yml
@@ -80,7 +80,7 @@ jobs:
         pip --version
         python ./requirements/adjust-versions.py requirements.txt
         python ./.github/prune-packages.py requirements/image.txt torchvision
-        pip install -q "numpy>=1.20.0"  # try to fix cocotools for PT 1.4 & 1.9
+        pip install -q "numpy==1.20.0"  # try to fix cocotools for PT 1.4 & 1.9
         pip install --requirement requirements.txt --quiet
         pip install --requirement requirements/test.txt --quiet
         pip list
