### 2022-01-06 54d6138
pin numpy version for Conda CI (#725)
* freeze numpy
* prune
* pin

Co-authored-by: quancs <quancs@qq.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka <jirka.borovec@seznam.cz>
