+++ .github/workflows/ci_test-conda.yml
@@ -7,12 +7,9 @@ on:  # Trigger the workflow on push or pull request, but only for the master bra
   pull_request:
     branches: [master, "release/*"]
 
-env:
-  is_release: ${{ startsWith(github.ref, 'refs/heads/release/') || startsWith(github.ref, 'refs/tags/') }}
-  is_main_branch: ${{ github.ref == 'refs/heads/main' }}
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}-${{ github.head_ref }}
-  cancel-in-progress: ${{ ! (fromJSON(env.is_main_branch) || fromJSON(env.is_release)) }}
+  cancel-in-progress: ${{ ! (github.ref == 'refs/heads/master' || startsWith(github.ref, 'refs/heads/release/') || startsWith(github.ref, 'refs/tags/')) }}
 
 jobs:
   conda:
