+++ .github/workflows/ci_docker.yml
@@ -13,12 +13,9 @@ on: # Trigger the workflow on push or pull request, but only for the master bran
       - ".github/workflows/*docker*.yml"
       - "setup.py"
 
-env:
-  is_release: ${{ startsWith(github.ref, 'refs/heads/release/') || startsWith(github.ref, 'refs/tags/') }}
-  is_main_branch: ${{ github.ref == 'refs/heads/main' }}
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}-${{ github.head_ref }}
-  cancel-in-progress: ${{ ! (fromJSON(env.is_main_branch) || fromJSON(env.is_release)) }}
+  cancel-in-progress: ${{ github.ref != 'refs/heads/master' }}
 
 jobs:
   build-PL:
