### 2022-01-07 283fbf8
Update naming conventions for cer, ter, wer, sacrebleu, rouge, squad
