+++ torchmetrics/functional/text/wer.py
@@ -22,30 +22,30 @@ from torchmetrics.functional.text.helper import _edit_distance
 
 
 def _wer_update(
-    predictions: Union[str, List[str]],
-    references: Union[str, List[str]],
+    preds: Union[str, List[str]],
+    target: Union[str, List[str]],
 ) -> Tuple[Tensor, Tensor]:
     """Update the wer score with the current set of references and predictions.
 
     Args:
-        predictions: Transcription(s) to score as a string or list of strings
-        references: Reference(s) for each speech input as a string or list of strings
+        preds: Transcription(s) to score as a string or list of strings
+        target: Reference(s) for each speech input as a string or list of strings
 
     Returns:
         Number of edit operations to get from the reference to the prediction, summed over all samples
         Number of words overall references
     """
-    if isinstance(predictions, str):
-        predictions = [predictions]
-    if isinstance(references, str):
-        references = [references]
+    if isinstance(preds, str):
+        preds = [preds]
+    if isinstance(target, str):
+        target = [target]
     errors = tensor(0, dtype=torch.float)
     total = tensor(0, dtype=torch.float)
-    for prediction, reference in zip(predictions, references):
-        prediction_tokens = prediction.split()
-        reference_tokens = reference.split()
-        errors += _edit_distance(prediction_tokens, reference_tokens)
-        total += len(reference_tokens)
+    for pred, tgt in zip(preds, target):
+        pred_tokens = pred.split()
+        tgt_tokens = tgt.split()
+        errors += _edit_distance(pred_tokens, tgt_tokens)
+        total += len(tgt_tokens)
     return errors, total
 
 
@@ -63,33 +63,33 @@ def _wer_compute(errors: Tensor, total: Tensor) -> Tensor:
 
 
 def word_error_rate(
-    predictions: Union[str, List[str]],
-    references: Union[str, List[str]],
+    preds: Union[str, List[str]],
+    target: Union[str, List[str]],
 ) -> Tensor:
     """Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system. This
     value indicates the percentage of words that were incorrectly predicted. The lower the value, the better the
     performance of the ASR system with a WER of 0 being a perfect score.
 
     Args:
-        predictions: Transcription(s) to score as a string or list of strings
-        references: Reference(s) for each speech input as a string or list of strings
+        preds: Transcription(s) to score as a string or list of strings
+        target: Reference(s) for each speech input as a string or list of strings
 
     Returns:
         Word error rate score
 
     Examples:
-        >>> predictions = ["this is the prediction", "there is an other sample"]
-        >>> references = ["this is the reference", "there is another one"]
-        >>> word_error_rate(predictions=predictions, references=references)
+        >>> preds = ["this is the prediction", "there is an other sample"]
+        >>> target = ["this is the reference", "there is another one"]
+        >>> word_error_rate(preds=preds, target=target)
         tensor(0.5000)
     """
-    errors, total = _wer_update(predictions, references)
+    errors, total = _wer_update(preds, target)
     return _wer_compute(errors, total)
 
 
 def wer(
-    predictions: Union[str, List[str]],
-    references: Union[str, List[str]],
+    preds: Union[str, List[str]],
+    target: Union[str, List[str]],
 ) -> Tensor:
     """Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system.
 
@@ -97,10 +97,10 @@ def wer(
         Use :func:`torchmetrics.fuctional.word_error_rate`. Will be removed in v0.8.
 
     Examples:
-        >>> predictions = ["this is the prediction", "there is an other sample"]
-        >>> references = ["this is the reference", "there is another one"]
[…]
