### 2022-01-07 f41c228
Remove redundant list comprehension in ter
