+++ torchmetrics/functional/text/ter.py
@@ -529,8 +529,8 @@ def _ter_update(
     target, preds = _validate_inputs(target, preds)
 
     for (pred, tgt) in zip(preds, target):
-        tgt_words_: List[List[str]] = [[word for word in _preprocess_sentence(_tgt, tokenizer).split()] for _tgt in tgt]
-        pred_words_: List[str] = [word for word in _preprocess_sentence(pred, tokenizer).split()]
+        tgt_words_: List[List[str]] = [_preprocess_sentence(_tgt, tokenizer).split() for _tgt in tgt]
+        pred_words_: List[str] = _preprocess_sentence(pred, tokenizer).split()
         num_edits, tgt_length = _compute_sentence_statistics(pred_words_, tgt_words_)
         total_num_edits += num_edits
         total_tgt_length += tgt_length
