### 2022-01-07 39b116d
Fix metrics documentation
