+++ docs/source/references/modules.rst
@@ -15,9 +15,9 @@ metrics.
     :noindex:
 
 
-*************************
-Basic Aggregation Metrics
-*************************
+*****************
+Basic Aggregation
+*****************
 
 Torchmetrics comes with a number of metrics for aggregation of basic statistics: mean, max, min etc. of
 either tensors or native python floats.
@@ -52,12 +52,12 @@ SumMetric
 .. autoclass:: torchmetrics.SumMetric
     :noindex:
 
-*************
-Audio Metrics
-*************
+*****
+Audio
+*****
 
-About Audio Metrics
-~~~~~~~~~~~~~~~~~~~
+About Audio
+~~~~~~~~~~~
 
 For the purposes of audio metrics, inputs (predictions, targets) must have the same size.
 If the input is 1D tensors the output will be a scalar. If the input is multi-dimensional with shape ``[...,time]``
@@ -116,9 +116,9 @@ STOI
     :noindex:
 
 
-**********************
-Classification Metrics
-**********************
+**************
+Classification
+**************
 
 Input types
 ~~~~~~~~~~~
@@ -359,9 +359,9 @@ StatScores
 .. autoclass:: torchmetrics.StatScores
     :noindex:
 
-*************
-Image Metrics
-*************
+*****
+Image
+*****
 
 Image quality metrics can be used to access the quality of synthetic generated images from machine
 learning algorithms such as `Generative Adverserial Networks (GANs) <https://en.wikipedia.org/wiki/Generative_adversarial_network>`_.
@@ -402,9 +402,9 @@ SSIM
 .. autoclass:: torchmetrics.SSIM
     :noindex:
 
-*****************
-Detection Metrics
-*****************
+*********
+Detection
+*********
 
 Object detection metrics can be used to evaluate the predicted detections with given groundtruth detections on images.
 
@@ -414,9 +414,9 @@ MAP
 .. autoclass:: torchmetrics.detection.map.MAP
     :noindex:
 
-******************
-Regression Metrics
-******************
+**********
+Regression
+**********
 
 CosineSimilarity
 ~~~~~~~~~~~~~~~~
