+++ docs/source/references/functional.rst
@@ -335,9 +335,9 @@ tweedie_deviance_score [func]
     :noindex:
 
 
-********
-Pairwise
-********
+****************
+Pairwise Metrics
+****************
 
 pairwise_cosine_similarity [func]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
