+++ docs/source/references/modules.rst
@@ -56,9 +56,6 @@ SumMetric
 Audio
 *****
 
-About Audio
-~~~~~~~~~~~
-
 For the purposes of audio metrics, inputs (predictions, targets) must have the same size.
 If the input is 1D tensors the output will be a scalar. If the input is multi-dimensional with shape ``[...,time]``
 the metric will be computed over the ``time`` dimension.
