### 2022-01-07 d1e31bf
Apply suggestions from code review
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
