### 2022-01-07 2f32c2d
CI: use concurrency (#724)
* use concurrency
* dockers too
* head_ref

Co-authored-by: Luca Di Liello <luca.diliello@unitn.it>
