### 2022-01-08 08d6bc4
Add warnings for BC
