### 2022-01-08 2b822d1
Change naming multiscale_ssim -> multiscale_structual_similarity_index_measure
