### 2022-01-08 5368929
Fix a typo: structual -> structural
