+++ torchmetrics/__init__.py
@@ -41,7 +41,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.image import PSNR, SSIM, MultiScaleStructualSimilarityIndexMeasure  # noqa: E402
+from torchmetrics.image import PSNR, SSIM, MultiScaleStructuralSimilarityIndexMeasure  # noqa: E402
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
@@ -123,7 +123,7 @@ __all__ = [
     "MinMaxMetric",
     "MinMetric",
     "MultioutputWrapper",
-    "MultiScaleStructualSimilarityIndexMeasure",
+    "MultiScaleStructuralSimilarityIndexMeasure",
     "PearsonCorrcoef",
     "PearsonCorrCoef",
     "PIT",
