### 2022-01-08 46cbafa
SignalDistortionRatio
