+++ torchmetrics/audio/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.audio.pit import PIT  # noqa: F401
-from torchmetrics.audio.sdr import SDR, ScaleInvariantSDR  # noqa: F401
+from torchmetrics.audio.sdr import SDR, ScaleInvariantSDR, SignalDistortionRatio  # noqa: F401
 from torchmetrics.audio.si_sdr import SI_SDR  # noqa: F401
 from torchmetrics.audio.si_snr import SI_SNR  # noqa: F401
 from torchmetrics.audio.snr import SNR  # noqa: F401
