### 2022-01-08 df6f123
ScaleInvariantSignalDistortionRatio
