+++ torchmetrics/audio/si_sdr.py
@@ -16,14 +16,14 @@ from warnings import warn
 
 from torch import Tensor
 
-from torchmetrics.audio.sdr import ScaleInvariantSDR
+from torchmetrics.audio.sdr import ScaleInvariantSignalDistortionRatio
 
 
-class SI_SDR(ScaleInvariantSDR):
+class SI_SDR(ScaleInvariantSignalDistortionRatio):
     """Scale-invariant signal-to-distortion ratio (SI-SDR).
 
     .. deprecated:: v0.7
-        Use :class:`torchmetrics.ScaleInvariantSDR`. Will be removed in v0.8.
+        Use :class:`torchmetrics.ScaleInvariantSignalDistortionRatio`. Will be removed in v0.8.
 
     Example:
         >>> import torch
@@ -40,5 +40,5 @@ class SI_SDR(ScaleInvariantSDR):
         process_group: Optional[Any] = None,
         dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
     ) -> None:
-        warn("`SI_SDR` was renamed to `ScaleInvariantSDR` in v0.7 and it will be removed in v0.8", DeprecationWarning)
+        warn("`SI_SDR` was renamed to `ScaleInvariantSignalDistortionRatio` in v0.7 and it will be removed in v0.8", DeprecationWarning)
         super().__init__(zero_mean, compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
