### 2022-01-08 be4b60a
SignalNoiseRatio
