+++ tests/audio/test_snr.py
@@ -22,7 +22,7 @@ from torch import Tensor
 
 from tests.helpers import seed_all
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
-from torchmetrics.audio import SNR
+from torchmetrics.audio import SignalNoiseRatio
 from torchmetrics.functional import signal_noise_ratio
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
 
@@ -86,7 +86,7 @@ class TestSNR(MetricTester):
             ddp,
             preds,
             target,
-            SNR,
+            SignalNoiseRatio,
             sk_metric=partial(average_metric, metric_func=sk_metric),
             dist_sync_on_step=dist_sync_on_step,
             metric_args=dict(zero_mean=zero_mean),
@@ -105,7 +105,7 @@ class TestSNR(MetricTester):
         self.run_differentiability_test(
             preds=preds,
             target=target,
-            metric_module=SNR,
+            metric_module=SignalNoiseRatio,
             metric_functional=signal_noise_ratio,
             metric_args={"zero_mean": zero_mean},
         )
@@ -121,13 +121,13 @@ class TestSNR(MetricTester):
         self.run_precision_test_gpu(
             preds=preds,
             target=target,
-            metric_module=SNR,
+            metric_module=SignalNoiseRatio,
             metric_functional=signal_noise_ratio,
             metric_args={"zero_mean": zero_mean},
         )
 
 
-def test_error_on_different_shape(metric_class=SNR):
+def test_error_on_different_shape(metric_class=SignalNoiseRatio):
     metric = metric_class()
     with pytest.raises(RuntimeError, match="Predictions and targets are expected to have the same shape"):
         metric(torch.randn(100), torch.randn(50))
