+++ README.md
@@ -268,7 +268,7 @@ We currently have implemented metrics within the following domains:
 - Audio (
   [SI_SDR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#si-sdr),
   [ScaleInvariantSNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#ScaleInvariantSNR),
-  [SNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#snr)
+  [SignalNoiseRatio](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#SignalNoiseRatio)
   and [few more](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#audio-metrics)
   )
 - Classification (
