### 2022-01-08 a6e44ba
ScaleInvariantSignalNoiseRatio
