+++ torchmetrics/functional/__init__.py
@@ -37,7 +37,7 @@ from torchmetrics.functional.classification.roc import roc
 from torchmetrics.functional.classification.specificity import specificity
 from torchmetrics.functional.classification.stat_scores import stat_scores
 from torchmetrics.functional.image.gradients import image_gradients
-from torchmetrics.functional.image.psnr import psnr
+from torchmetrics.functional.image.psnr import peak_signal_noise_ratio
 from torchmetrics.functional.image.ssim import ssim
 from torchmetrics.functional.pairwise.cosine import pairwise_cosine_similarity
 from torchmetrics.functional.pairwise.euclidean import pairwise_euclidean_distance
@@ -114,7 +114,7 @@ __all__ = [
     "precision",
     "precision_recall",
     "precision_recall_curve",
-    "psnr",
+    "peak_signal_noise_ratio",
     "r2_score",
     "recall",
     "retrieval_average_precision",
