### 2022-01-08 fcf2f7c
Revert "psnr --> peak_signal_noise_ratio"
This reverts commit 6e957726364c4721a95385203c3d13a5b870769c.
