+++ torchmetrics/functional/image/psnr.py
@@ -91,7 +91,7 @@ def _psnr_update(
     return sum_squared_error, n_obs
 
 
-def peak_signal_noise_ratio(
+def psnr(
     preds: Tensor,
     target: Tensor,
     data_range: Optional[float] = None,
@@ -125,10 +125,10 @@ def peak_signal_noise_ratio(
             If ``dim`` is not ``None`` and ``data_range`` is not provided.
 
     Example:
-        >>> from torchmetrics.functional import peak_signal_noise_ratio
+        >>> from torchmetrics.functional import psnr
         >>> pred = torch.tensor([[0.0, 1.0], [2.0, 3.0]])
         >>> target = torch.tensor([[3.0, 2.0], [1.0, 0.0]])
-        >>> peak_signal_noise_ratio(pred, target)
+        >>> psnr(pred, target)
         tensor(2.5527)
 
     .. note::
