### 2022-01-08 9bcbdf4
Refactor: SDR & SI_SDR (#711)
* signal_distortion_ratio
* scale_invariant_signal_distortion_ratio
* SignalDistortionRatio
* ScaleInvariantSignalDistortionRatio

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
