### 2022-01-08 b093b80
Apply suggestions from code review
