+++ torchmetrics/audio/snr.py
@@ -134,7 +134,6 @@ class SNR(SignalNoiseRatio):
         dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
     ) -> None:
         warn("`SNR` was renamed to `SignalNoiseRatio` in v0.7 and it will be removed in v0.8", DeprecationWarning)
-
         super().__init__(zero_mean, compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
 
 
