+++ torchmetrics/audio/si_snr.py
@@ -40,7 +40,7 @@ class SI_SNR(ScaleInvariantSignalNoiseRatio):
         dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
     ) -> None:
         warn(
-            "`SI_SDR` was renamed to `ScaleInvariantSignalNoiseRatio` in v0.7 and it will be removed in v0.8",
+            "`SI_SNR` was renamed to `ScaleInvariantSignalNoiseRatio` in v0.7 and it will be removed in v0.8",
             DeprecationWarning,
         )
         super().__init__(compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
