+++ torchmetrics/__init__.py
@@ -49,7 +49,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.image import PeakSignalNoiseRatio, SSIM  # noqa: E402
+from torchmetrics.image import SSIM, PeakSignalNoiseRatio  # noqa: E402
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
