### 2022-01-08 d11641c
Refactor: SNR & SI_SNR (#712)
* signal_noise_ratio
* scale_invariant_signal_noise_ratio
* SignalNoiseRatio
* ScaleInvariantSignalNoiseRatio

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
