+++ README.md
@@ -267,8 +267,8 @@ We currently have implemented metrics within the following domains:
 
 - Audio (
   [ScaleInvariantSignalDistortionRatio](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#ScaleInvariantSignalDistortionRatio),
-  [SI_SNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#si-snr),
-  [SNR](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#snr)
+  [ScaleInvariantSignalNoiseRatio](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#ScaleInvariantSignalNoiseRatio),
+  [SignalNoiseRatio](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#SignalNoiseRatio)
   and [few more](https://torchmetrics.readthedocs.io/en/latest/references/modules.html#audio-metrics)
   )
 - Classification (
