+++ torchmetrics/functional/__init__.py
@@ -15,7 +15,7 @@ from torchmetrics.functional.audio.pit import pit, pit_permutate
 from torchmetrics.functional.audio.sdr import scale_invariant_signal_distortion_ratio, sdr, signal_distortion_ratio
 from torchmetrics.functional.audio.si_sdr import si_sdr
 from torchmetrics.functional.audio.si_snr import si_snr
-from torchmetrics.functional.audio.snr import snr
+from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio, signal_noise_ratio, snr
 from torchmetrics.functional.classification.accuracy import accuracy
 from torchmetrics.functional.classification.auc import auc
 from torchmetrics.functional.classification.auroc import auroc
@@ -133,7 +133,9 @@ __all__ = [
     "si_sdr",
     "scale_invariant_signal_distortion_ratio",
     "si_snr",
+    "scale_invariant_signal_noise_ratio",
     "snr",
+    "signal_noise_ratio",
     "spearman_corrcoef",
     "specificity",
     "squad",
