### 2022-01-08 459b01e
Fix metrics documentation (#728)
* Fix metrics documentation
* Resolve comments
