### 2022-01-09 76e7185
modify object F1Score
