+++ torchmetrics/__init__.py
@@ -28,6 +28,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     AUC,
     AUROC,
     F1Score,
+    F1,
     ROC,
     Accuracy,
     AveragePrecision,
@@ -113,6 +114,7 @@ __all__ = [
     "CosineSimilarity",
     "TweedieDevianceScore",
     "ExplainedVariance",
+    "F1",
     "F1Score",
     "FBeta",
     "HammingDistance",
