### 2022-01-09 b644ef2
ci: gpus EU
