### 2022-01-09 db9cc3d
Apply a suggestion from the code review
