### 2022-01-09 8b330ea
Handle BC and add warnings
