+++ torchmetrics/text/chrf.py
@@ -19,6 +19,7 @@
 
 import itertools
 from typing import Any, Callable, Dict, Iterator, List, Optional, Sequence, Tuple, Union
+from warnings import warn
 
 import torch
 from torch import Tensor, tensor
@@ -143,7 +144,13 @@ class CHRFScore(Metric):
         if self.return_sentence_level_score:
             self.add_state("sentence_chrf_score", [], dist_reduce_fx="cat")
 
-    def update(self, preds: Sequence[str], target: Sequence[Sequence[str]]) -> None:  # type: ignore
+    def update(  # type: ignore
+        self,
+        preds: Sequence[str],
+        target: Sequence[Sequence[str]],
+        hypothesis_corpus: Union[None, Sequence[str]] = None,
+        reference_corpus: Union[None, Sequence[Sequence[str]]] = None,
+    ) -> None:
         """Compute Precision Scores.
 
         Args:
@@ -152,6 +159,21 @@ class CHRFScore(Metric):
             target:
                 An iterable of iterables of reference corpus.
         """
+        if hypothesis_corpus is not None:
+            warn(
+                "You are using deprecated argument `hypothesis_corpus` in v0.7 which was renamed to `preds`. "
+                " The past argument will be removed in v0.8.",
+                DeprecationWarning,
+            )
+            preds = hypothesis_corpus
+        if reference_corpus is not None:
+            warn(
+                "You are using deprecated argument `reference_corpus` in v0.7 which was renamed to `target`. "
+                " The past argument will be removed in v0.8.",
+                DeprecationWarning,
+            )
+            target = reference_corpus
+
         n_grams_dicts_tuple = _chrf_score_update(
             preds,
             target,
