+++ torchmetrics/__init__.py
@@ -37,7 +37,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     CalibrationError,
     CohenKappa,
     ConfusionMatrix,
-    FBeta,
+    FBetaScore,
     HammingDistance,
     Hinge,
     IoU,
@@ -114,7 +114,7 @@ __all__ = [
     "TweedieDevianceScore",
     "ExplainedVariance",
     "F1",
-    "FBeta",
+    "FBetaScore",
     "HammingDistance",
     "Hinge",
     "JaccardIndex",
