+++ docs/source/references/modules.rst
@@ -286,10 +286,10 @@ F1
 .. autoclass:: torchmetrics.F1
     :noindex:
 
-FBeta
+FBetaScore
 ~~~~~
 
-.. autoclass:: torchmetrics.FBeta
+.. autoclass:: torchmetrics.FBetaScore
     :noindex:
 
 HammingDistance

[…]
