### 2022-01-09 726f9bd
rename f1 score (#731)
* f1_score
* F1Score

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: Jirka <jirka.borovec@seznam.cz>
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
