+++ torchmetrics/__init__.py
@@ -37,6 +37,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     CalibrationError,
     CohenKappa,
     ConfusionMatrix,
+    F1Score,
     FBeta,
     HammingDistance,
     Hinge,
@@ -114,6 +115,7 @@ __all__ = [
     "TweedieDevianceScore",
     "ExplainedVariance",
     "F1",
+    "F1Score",
     "FBeta",
     "HammingDistance",
     "Hinge",
