### 2022-01-10 935b005
change hinge to hinge_loss
