+++ torchmetrics/functional/classification/__init__.py
@@ -21,7 +21,7 @@ from torchmetrics.functional.classification.confusion_matrix import confusion_ma
 from torchmetrics.functional.classification.dice import dice_score  # noqa: F401
 from torchmetrics.functional.classification.f_beta import f1, fbeta  # noqa: F401
 from torchmetrics.functional.classification.hamming_distance import hamming_distance  # noqa: F401
-from torchmetrics.functional.classification.hinge import hinge  # noqa: F401
+from torchmetrics.functional.classification.hinge import hinge, hinge_loss  # noqa: F401
 from torchmetrics.functional.classification.jaccard import jaccard_index  # noqa: F401
 from torchmetrics.functional.classification.kl_divergence import kl_divergence  # noqa: F401
 from torchmetrics.functional.classification.matthews_corrcoef import matthews_corrcoef  # noqa: F401
