+++ torchmetrics/functional/__init__.py
@@ -26,7 +26,7 @@ from torchmetrics.functional.classification.confusion_matrix import confusion_ma
 from torchmetrics.functional.classification.dice import dice_score
 from torchmetrics.functional.classification.f_beta import f1, fbeta
 from torchmetrics.functional.classification.hamming_distance import hamming_distance
-from torchmetrics.functional.classification.hinge import hinge
+from torchmetrics.functional.classification.hinge import hinge, hinge_loss
 from torchmetrics.functional.classification.iou import iou  # noqa: F401
 from torchmetrics.functional.classification.jaccard import jaccard_index
 from torchmetrics.functional.classification.kl_divergence import kl_divergence
@@ -96,6 +96,7 @@ __all__ = [
     "fbeta",
     "hamming_distance",
     "hinge",
+    "hinge_loss",
     "image_gradients",
     "jaccard_index",
     "kl_divergence",
