+++ tests/classification/test_hinge.py
@@ -22,7 +22,7 @@ from sklearn.preprocessing import OneHotEncoder
 from tests.classification.inputs import Input
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, NUM_CLASSES, MetricTester
 from torchmetrics.functional import hinge_loss
-from torchmetrics.functional.classification.hinge import MulticlassMode, HingeLoss
+from torchmetrics.functional.classification.hinge import HingeLoss, MulticlassMode
 
 torch.manual_seed(42)
 
