+++ torchmetrics/__init__.py
@@ -41,6 +41,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     FBeta,
     HammingDistance,
     Hinge,
+    HingeLoss,
     IoU,
     JaccardIndex,
     KLDivergence,
@@ -119,6 +120,7 @@ __all__ = [
     "FBeta",
     "HammingDistance",
     "Hinge",
+    "HingeLoss",
     "JaccardIndex",
     "KLDivergence",
     "MatthewsCorrcoef",
