### 2022-01-10 5c12358
hingeLoss correct function
