+++ tests/classification/test_hinge.py
@@ -21,8 +21,9 @@ from sklearn.preprocessing import OneHotEncoder
 
 from tests.classification.inputs import Input
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, NUM_CLASSES, MetricTester
-from torchmetrics.functional.classification.hinge import hinge_loss
-from torchmetrics.classification.hinge import HingeLoss, MulticlassMode
+from torchmetrics import HingeLoss
+from torchmetrics.functional import hinge_loss
+from torchmetrics.functional.classification.hinge import MulticlassMode
 
 torch.manual_seed(42)
 
