+++ CHANGELOG.md
@@ -76,9 +76,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   * `SNR` -> `SignalNoiseRatio`
   * `SI_SNR` -> `ScaleInvariantSignalNoiseRatio`
 
+
 - Renamed F-score metrics: ([#731](https://github.com/PyTorchLightning/metrics/pull/731))
   * `torchmetrics.functional.f1` -> `torchmetrics.functional.f1_score`
   * `torchmetrics.F1` -> `torchmetrics.F1Score`
+
+
+- Renamed Hinge metric: ([#734](https://github.com/PyTorchLightning/metrics/pull/734))
+  * `torchmetrics.functional.hinge` -> `torchmetrics.functional.hinge_loss`
+  * `torchmetrics.Hinge` -> `torchmetrics.HingeLoss`
+
 
 ### Removed
 
