### 2022-01-10 4fd42ba
enumerate instead of range
