### 2022-01-10 8f0a15e
Revert "fbeta --> fbeta_score FBeta --> FBetaScore"
This reverts commit f77774c8
