+++ tests/classification/test_f_beta.py
@@ -17,7 +17,7 @@ from typing import Callable, Optional
 import numpy as np
 import pytest
 import torch
-from sklearn.metrics import f1_score, fbeta
+from sklearn.metrics import f1_score, fbeta_score
 from torch import Tensor
 
 from tests.classification.inputs import _input_binary, _input_binary_logits, _input_binary_prob
@@ -32,8 +32,8 @@ from tests.classification.inputs import _input_multilabel_logits as _input_mlb_l
 from tests.classification.inputs import _input_multilabel_prob as _input_mlb_prob
 from tests.helpers import seed_all
 from tests.helpers.testers import NUM_BATCHES, NUM_CLASSES, THRESHOLD, MetricTester
-from torchmetrics import F1, FBetaScore, Metric
-from torchmetrics.functional import f1, fbeta_score
+from torchmetrics import F1, FBeta, Metric
+from torchmetrics.functional import f1, fbeta
 from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.enums import AverageMethod
 
@@ -92,7 +92,7 @@ def _sk_fbeta_f1_multidim_multiclass(
 @pytest.mark.parametrize(
     "metric_class, metric_fn",
     [
-        (partial(FBetaScore, beta=2.0), partial(fbeta_score, beta=2.0)),
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
         (F1, f1),
     ],
 )
@@ -128,7 +128,7 @@ def test_wrong_params(metric_class, metric_fn, average, mdmc_average, num_classe
 @pytest.mark.parametrize(
     "metric_class, metric_fn",
     [
-        (partial(FBetaScore, beta=2.0), partial(fbeta_score, beta=2.0)),
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
         (F1, f1),
     ],
 )
@@ -150,7 +150,7 @@ def test_zero_division(metric_class, metric_fn):
 @pytest.mark.parametrize(
     "metric_class, metric_fn",
     [
-        (partial(FBetaScore, beta=2.0), partial(fbeta_score, beta=2.0)),
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
         (F1, f1),
     ],
 )
@@ -178,9 +178,7 @@ def test_no_support(metric_class, metric_fn):
     assert result_cl == result_fn == 0
 
 
-@pytest.mark.parametrize(
-    "metric_class, metric_fn", [(partial(FBetaScore, beta=2.0), partial(fbeta_score, beta=2.0)), (F1, f1)]
-)
+@pytest.mark.parametrize("metric_class, metric_fn", [(partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)), (F1, f1)])
 @pytest.mark.parametrize(
     "ignore_index, expected", [(None, torch.tensor([1.0, np.nan])), (0, torch.tensor([np.nan, np.nan]))]
 )
@@ -204,10 +202,7 @@ def test_class_not_present(metric_class, metric_fn, ignore_index, expected):
 
 @pytest.mark.parametrize(
     "metric_class, metric_fn, sk_fn",
-    [
-        (partial(FBetaScore, beta=2.0), partial(fbeta_score, beta=2.0), partial(fbeta_score, beta=2.0)),
-        (F1, f1, f1_score),
-    ],
+    [(partial(FBeta, beta=2.0), partial(fbeta, beta=2.0), partial(fbeta_score, beta=2.0)), (F1, f1, f1_score)],
 )
[…]
