+++ torchmetrics/__init__.py
@@ -37,7 +37,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     CalibrationError,
     CohenKappa,
     ConfusionMatrix,
-    FBetaScore,
+    FBeta,
     HammingDistance,
     Hinge,
     IoU,
@@ -114,7 +114,7 @@ __all__ = [
     "TweedieDevianceScore",
     "ExplainedVariance",
     "F1",
-    "FBetaScore",
+    "FBeta",
     "HammingDistance",
     "Hinge",
     "JaccardIndex",
