### 2022-01-10 45806f7
compute on device
