### 2022-01-10 9288222
Add `MultiScaleStructuralSimilarityIndexMeasure` (#679)
* [WIP] Add MS-SSIM

* Fix docs and a mypy issue

* fix badge

* Apply some SkafteNicki's suggestions from code review

* Fix docstring: MS_SSIM -> MultiScaleSSIM

* Fix missed MS_SSIM in a docstring

* flake8

* Add tests

* Update doc example

* Fix seed for doc + update some tests

* Clean code

* Update doc example + make diff test tractable

* torch.manual seed passed directly to torch.rand

* Apply suggestions from code review

* Drop some tests ; fix docs issues + add missing docs

* Update paper.md (#690)

* ci: rename oldest

* C
[…]
