### 2022-01-10 94dec52
Add `MultiScaleStructuralSimilarityIndexMeasure` (#679)
* Apply suggestions from code review
* Change naming multiscale_ssim -> multiscale_structual_similarity_index_measure

Co-authored-by: Jirka <jirka.borovec@seznam.cz>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>
