+++ torchmetrics/__init__.py
@@ -52,7 +52,11 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.image import SSIM, PeakSignalNoiseRatio  # noqa: E402
+from torchmetrics.image import (  # noqa: E402
+    PSNR,
+    SSIM,
+    PeakSignalNoiseRatio,
+)
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
@@ -140,6 +144,7 @@ __all__ = [
     "PIT",
     "Precision",
     "PrecisionRecallCurve",
+    "PSNR",
     "PeakSignalNoiseRatio",
     "R2Score",
     "Recall",
