### 2022-01-10 ab91460
Remove exception
