+++ torchmetrics/functional/text/wip.py
@@ -78,13 +78,19 @@ def word_information_preserved(
     Args:
         preds: Transcription(s) to score as a string or list of strings
         total: Reference(s) for each speech input as a string or list of strings
+        predictions:
+            Transcription(s) to score as a string or list of strings
+            This argument is deprecated in v0.7 and will be removed in v0.8. Use `preds` instead.
+        references:
+            Reference(s) for each speech input as a string or list of strings
+            This argument is deprecated in v0.7 and will be removed in v0.8. Use `target` instead.
     Returns:
         Word Information preserved rate
     Examples:
         >>> from torchmetrics.functional import word_information_preserved
         >>> preds = ["this is the prediction", "there is an other sample"]
         >>> target = ["this is the reference", "there is another one"]
-        >>> word_information_preserved(preds=preds, target=target)
+        >>> word_information_preserved(preds, target)
         tensor(0.3472)
     """
     if predictions is not None:
