+++ requirements/devel.txt
@@ -4,5 +4,16 @@
 # add the testing dependencies
 -r test.txt
 
+# add extra requirements
+-r image.txt
+-r text.txt
+-r detection.txt
+-r audio.txt
+
+# add extra testing
+-r image_test.txt
+-r text_test.txt
+-r audio_test.txt
+
 # add the integration dependencies
 -r integrate.txt
