### 2022-01-10 a714acf
Replace prec score loop
