### 2022-01-10 317182c
rename `hinge` to `hinge_loss` (#734)
* change hinge to hinge_loss
* HingeLoss

Co-authored-by: Jirka <jirka.borovec@seznam.cz>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>
