+++ .deepsource.toml
@@ -4,7 +4,7 @@ test_patterns = ["tests/**"]
 
 [[analyzers]]
 name = "test-coverage"
-enabled = true
+enabled = false
 
 [[analyzers]]
 name = "shell"
