+++ torchmetrics/image/inception.py
@@ -127,10 +127,9 @@ class IS(Metric):
 
         if isinstance(feature, (str, int)):
             if not _TORCH_FIDELITY_AVAILABLE:
-                raise ValueError(
+                raise ModuleNotFoundError(
                     "IS metric requires that Torch-fidelity is installed."
-                    "Either install as `pip install torchmetrics[image]`"
-                    " or `pip install torch-fidelity`"
+                    " Either install as `pip install torchmetrics[image]` or `pip install torch-fidelity`."
                 )
             valid_int_input = ("logits_unbiased", 64, 192, 768, 2048)
             if feature not in valid_int_input:
