### 2022-01-10 cbeaad6
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
