+++ torchmetrics/functional/audio/sdr.py
@@ -128,7 +128,7 @@ def signal_distortion_ratio(
     """
     if not _FAST_BSS_EVAL_AVAILABLE:
         raise ModuleNotFoundError(
-            "SDR metric requires that fast-bss-eval is installed."
+            "SDR metric requires that `fast-bss-eval` is installed."
             " Either install as `pip install torchmetrics[audio]` or `pip install fast-bss-eval`."
         )
     _check_same_shape(preds, target)
