+++ torchmetrics/functional/text/sacre_bleu.py
@@ -337,7 +337,7 @@ def sacre_bleu_score(
         raise ValueError(f"Corpus has different size {len(translate_corpus)} != {len(reference_corpus)}")
     if tokenize == "intl" and not _REGEX_AVAILABLE:
         raise ModuleNotFoundError(
-            "`'intl'` tokenization requires `regex` installed."
+            "`'intl'` tokenization requires that `regex` is installed."
             " Use `pip install regex` or `pip install torchmetrics[text]`."
         )
 
