+++ torchmetrics/text/rouge.py
@@ -102,7 +102,7 @@ class ROUGEScore(Metric):
         if use_stemmer or "rougeLsum" in rouge_keys:
             if not _NLTK_AVAILABLE:
                 raise ModuleNotFoundError(
-                    "Stemmer and/or `rougeLsum` requires that nltk is installed. Use `pip install nltk`."
+                    "Stemmer and/or `rougeLsum` requires that `nltk` is installed. Use `pip install nltk`."
                 )
             import nltk
 
