+++ requirements/devel.txt
@@ -1,5 +1,5 @@
 # use mandatory dependencies
--r ../requirements.txt
+#-r ../requirements.txt
 
 # add the testing dependencies
 -r test.txt
