### 2022-01-10 3c3395c
Apply suggestions from code review
