### 2022-01-10 9f1404c
Unify functional metrics names #729
To unify `pit` to `permutation_invariant_training`
