+++ torchmetrics/__init__.py
@@ -14,7 +14,7 @@ _PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
 from torchmetrics import functional  # noqa: E402
 from torchmetrics.aggregation import CatMetric, MaxMetric, MeanMetric, MinMetric, SumMetric  # noqa: E402
 from torchmetrics.audio import (  # noqa: E402
-    PIT,
+    permutation_invariant_training,
     SDR,
     SI_SDR,
     SI_SNR,
@@ -140,7 +140,7 @@ __all__ = [
     "MultiScaleStructuralSimilarityIndexMeasure",
     "PearsonCorrcoef",
     "PearsonCorrCoef",
-    "PIT",
+    "permutation_invariant_training",
     "Precision",
     "PrecisionRecallCurve",
     "PSNR",
