+++ tests/audio/test_pit.py
@@ -23,8 +23,8 @@ from torch import Tensor
 
 from tests.helpers import seed_all
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
-from torchmetrics.audio import PIT
-from torchmetrics.functional import pit, scale_invariant_signal_distortion_ratio, snr
+from torchmetrics.audio import permutation_invariant_training
+from torchmetrics.functional import permutation_invariant_training, scale_invariant_signal_distortion_ratio, snr
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
 
 seed_all(42)
@@ -71,7 +71,7 @@ def naive_implementation_pit_scipy(
         for e in range(spk_num):
             metric_mtx[:, t, e] = metric_func(preds[:, e, ...], target[:, t, ...])
 
-    # pit_r = PIT(metric_func, eval_func)(preds, target)
+    # pit_r = permutation_invariant_training(metric_func, eval_func)(preds, target)
     metric_mtx = metric_mtx.detach().cpu().numpy()
     best_metrics = []
     best_perms = []
@@ -121,7 +121,7 @@ class TestPIT(MetricTester):
             ddp,
             preds,
             target,
-            PIT,
+            permutation_invariant_training,
             sk_metric=partial(_average_metric, metric_func=sk_metric),
             dist_sync_on_step=dist_sync_on_step,
             metric_args=dict(metric_func=metric_func, eval_func=eval_func),
@@ -131,19 +131,19 @@ class TestPIT(MetricTester):
         self.run_functional_metric_test(
             preds=preds,
             target=target,
-            metric_functional=pit,
+            metric_functional=permutation_invariant_training,
             sk_metric=sk_metric,
             metric_args=dict(metric_func=metric_func, eval_func=eval_func),
         )
 
     def test_pit_differentiability(self, preds, target, sk_metric, metric_func, eval_func):
         def pit_diff(preds, target, metric_func, eval_func):
-            return pit(preds, target, metric_func, eval_func)[0]
+            return permutation_invariant_training(preds, target, metric_func, eval_func)[0]
 
         self.run_differentiability_test(
             preds=preds,
             target=target,
-            metric_module=PIT,
+            metric_module=permutation_invariant_training,
             metric_functional=pit_diff,
             metric_args={"metric_func": metric_func, "eval_func": eval_func},
         )
@@ -159,32 +159,32 @@ class TestPIT(MetricTester):
         self.run_precision_test_gpu(
             preds=preds,
             target=target,
-            metric_module=PIT,
-            metric_functional=partial(pit, metric_func=metric_func, eval_func=eval_func),
+            metric_module=permutation_invariant_training,
+            metric_functional=partial(permutation_invariant_training, metric_func=metric_func, eval_func=eval_func),
             metric_args={"metric_func": metric_func, "eval_func": eval_func},
         )
 
 
 def test_error_on_different_shape() -> None:
-    metric = PIT(snr, "max")
+    metric = permutation_invariant_training(snr, "max")
     with pytest.raises(RuntimeError, match="Predictions and targets are expected to have the same shape"):
         metric(torch.randn(3, 3, 10), torch.randn(3, 2, 10))
 
 
 def test_error_on_wrong_eval_func() -> None:
-    metric = PIT(snr, "xxx")
+    metric = permutation_invariant_training(snr, "xxx")
     with pytest.raises(ValueError, match='eval_func can only be "max" or "min"'):
         metric(torch.randn(3, 3, 10), torch.randn(3, 3, 10))
 
 
 def test_error_on_wrong_shape() -> None:
-    metric = PIT(snr, "max")
[…]
