### 2022-01-10 8357b04
cleaning ModuleNotFoundError (#736)
* use ModuleNotFoundError
* Apply suggestions from code review
* cleaning

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
