+++ torchmetrics/functional/audio/__init__.py
@@ -11,8 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from torchmetrics.functional.audio.pit import PermutationInvariantTraining  # noqa: F401
-from torchmetrics.audio.sdr import SDR, ScaleInvariantSignalDistortionRatio, SignalDistortionRatio  # noqa: F401
-from torchmetrics.audio.si_sdr import SI_SDR  # noqa: F401
-from torchmetrics.audio.si_snr import SI_SNR  # noqa: F401
-from torchmetrics.audio.snr import SNR, ScaleInvariantSignalNoiseRatio, SignalNoiseRatio  # noqa: F401
+from torchmetrics.functional.audio.pit import permutation_invariant_training, pit_permutate  # noqa: F401
+from torchmetrics.functional.audio.sdr import (  # noqa: F401
+    scale_invariant_signal_distortion_ratio,
+    sdr,
+    signal_distortion_ratio,
+)
+from torchmetrics.functional.audio.si_sdr import si_sdr  # noqa: F401
+from torchmetrics.functional.audio.si_snr import si_snr  # noqa: F401
+from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio, signal_noise_ratio, snr  # noqa: F401
