+++ torchmetrics/__init__.py
@@ -14,6 +14,7 @@ _PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
 from torchmetrics import functional  # noqa: E402
 from torchmetrics.aggregation import CatMetric, MaxMetric, MeanMetric, MinMetric, SumMetric  # noqa: E402
 from torchmetrics.audio import (  # noqa: E402
+    PermutationInvariantTraining,
     SDR,
     SI_SDR,
     SI_SNR,
@@ -22,7 +23,6 @@ from torchmetrics.audio import (  # noqa: E402
     ScaleInvariantSignalNoiseRatio,
     SignalDistortionRatio,
     SignalNoiseRatio,
-    permutation_invariant_training,
 )
 from torchmetrics.classification import (  # noqa: E402, F401
     AUC,
