+++ .github/workflows/ci_test-conda.yml
@@ -85,8 +85,8 @@ jobs:
         python ./requirements/adjust-versions.py requirements.txt
         python ./.github/prune-packages.py requirements/image.txt torchvision
         pip install -q "numpy==1.20.0"  # try to fix cocotools for PT 1.4 & 1.9
-        pip install .[all] --quiet
-        pip uninstall -y torchmetrics
+        pip install --requirement requirements.txt --quiet
+        pip install --requirement requirements/devel.txt --quiet
         pip list
         python -c "from torch import __version__ as ver; assert '.'.join(ver.split('.')[:2]) == '${{ matrix.pytorch-version }}', ver"
       shell: bash -l {0}
