+++ requirements/devel.txt
@@ -16,4 +16,4 @@
 -r audio_test.txt
 
 # add the integration dependencies
--r integrate.txt
+#-r integrate.txt
