### 2022-01-10 71e82c1
Set preds, target = None
