+++ torchmetrics/text/sacre_bleu.py
@@ -117,8 +117,8 @@ class SacreBLEUScore(BLEUScore):
 
     def update(  # type: ignore
         self,
-        preds: Sequence[str],
-        target: Sequence[Sequence[str]],
+        preds: Union[None, Sequence[str]] = None,
+        target: Union[None, Sequence[Sequence[str]]] = None,
         translate_corpus: Union[None, Sequence[str]] = None,
         reference_corpus: Union[None, Sequence[Sequence[str]]] = None,
     ) -> None:
@@ -134,20 +134,27 @@ class SacreBLEUScore(BLEUScore):
                 An iterable of iterables of reference corpus
                 This argument is deprecated in v0.7 and will be removed in v0.8. Use `target` instead.
         """
+        if preds is None and translate_corpus is None:
+            raise ValueError("Either `preds` or `translate_corpus` must be provided.")
+        if target is None and reference_corpus is None:
+            raise ValueError("Either `target` or `reference_corpus` must be provided.")
+
         if translate_corpus is not None:
             warn(
                 "You are using deprecated argument `translate_corpus` in v0.7 which was renamed to `preds`. "
                 " The past argument will be removed in v0.8.",
                 DeprecationWarning,
             )
-            preds = translate_corpus
+            warn("If you specify both `preds` and `translate_corpus`, only `preds` is considered.")
+            preds = preds or translate_corpus
         if reference_corpus is not None:
             warn(
                 "You are using deprecated argument `reference_corpus` in v0.7 which was renamed to `target`. "
                 " The past argument will be removed in v0.8.",
                 DeprecationWarning,
             )
-            target = reference_corpus
+            warn("If you specify both `target` and `reference_corpus`, only `target` is considered.")
+            target = target or reference_corpus
 
         self.preds_len, self.target_len = _bleu_score_update(
             preds,
