### 2022-01-10 27783a0
Fix auc flattening
