### 2022-01-10 04e8724
Rename instances of `manhatten` to `manhattan`
