+++ torchmetrics/functional/__init__.py
@@ -43,7 +43,7 @@ from torchmetrics.functional.image.ssim import ssim
 from torchmetrics.functional.pairwise.cosine import pairwise_cosine_similarity
 from torchmetrics.functional.pairwise.euclidean import pairwise_euclidean_distance
 from torchmetrics.functional.pairwise.linear import pairwise_linear_similarity
-from torchmetrics.functional.pairwise.manhatten import pairwise_manhatten_distance
+from torchmetrics.functional.pairwise.manhattan import pairwise_manhattan_distance
 from torchmetrics.functional.regression.cosine_similarity import cosine_similarity
 from torchmetrics.functional.regression.explained_variance import explained_variance
 from torchmetrics.functional.regression.mean_absolute_error import mean_absolute_error
@@ -114,7 +114,7 @@ __all__ = [
     "pairwise_cosine_similarity",
     "pairwise_euclidean_distance",
     "pairwise_linear_similarity",
-    "pairwise_manhatten_distance",
+    "pairwise_manhattan_distance",
     "pearson_corrcoef",
     "permutation_invariant_training",
     "pit",
