### 2022-01-10 5d8cc5e
Revert typo correction in CHANGELOG.md
