### 2022-01-10 3bd4fb0
bugfix: change dtype of confmat to int64 (#715)
* bugfix: change dtype of confmat to int64

fixes slightly varying ground truth counts (issue #708)

* doctests

* docs

* changelog

Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: Jirka <jirka.borovec@seznam.cz>
Co-authored-by: SkafteNicki <skaftenicki@gmail.com>
