+++ integrations/test_lightning.py
@@ -202,7 +202,8 @@ def test_metrics_reset(tmpdir):
 
 
 def test_metric_lightning_log(tmpdir):
-    """ Test logging a metric object and that the metric state gets reset after each epoch."""
+    """Test logging a metric object and that the metric state gets reset after each epoch."""
+
     class TestModel(BoringModel):
         def __init__(self):
             super().__init__()
@@ -218,10 +219,10 @@ def test_metric_lightning_log(tmpdir):
             self.metric_step(x.sum())
             self.sum += x.sum()
             self.log("sum_step", self.metric_step, on_epoch=True, on_step=False)
-            return {'loss': self.step(x), 'data': x}
+            return {"loss": self.step(x), "data": x}
 
         def training_epoch_end(self, outs):
-            self.log("sum_epoch", self.metric_epoch(torch.stack([o['data'] for o in outs]).sum()))
+            self.log("sum_epoch", self.metric_epoch(torch.stack([o["data"] for o in outs]).sum()))
 
     model = TestModel()
     model.val_dataloader = None
@@ -242,9 +243,7 @@ def test_metric_lightning_log(tmpdir):
 
 
 def test_metric_collection_lightning_log(tmpdir):
-
     class TestModel(BoringModel):
-
         def __init__(self):
             super().__init__()
             self.metric = MetricCollection([SumMetric(), DiffMetric()])
@@ -256,12 +255,12 @@ def test_metric_collection_lightning_log(tmpdir):
             metric_vals = self.metric(x.sum())
             self.sum += x.sum()
             self.diff -= x.sum()
-            self.log_dict({f'{k}_step': v for k, v in metric_vals.items()})
+            self.log_dict({f"{k}_step": v for k, v in metric_vals.items()})
             return self.step(x)
 
         def training_epoch_end(self, outputs):
             metric_vals = self.metric.compute()
-            self.log_dict({f'{k}_epoch': v for k, v in metric_vals.items()})
+            self.log_dict({f"{k}_epoch": v for k, v in metric_vals.items()})
 
     model = TestModel()
     model.val_dataloader = None
