### 2022-01-10 9390fc6
PSNR backward compatibility.
