+++ torchmetrics/functional/image/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 
 from torchmetrics.functional.image.gradients import image_gradients  # noqa: F401
-from torchmetrics.functional.image.psnr import peak_signal_noise_ratio  # noqa: F401
+from torchmetrics.functional.image.psnr import peak_signal_noise_ratio,psnr  # noqa: F401
