+++ torchmetrics/functional/image/psnr.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Optional, Tuple, Union
-
+from warnings import warn
 import torch
 from torch import Tensor, tensor
 
@@ -147,4 +147,67 @@ def peak_signal_noise_ratio(
     else:
         data_range = tensor(float(data_range))
     sum_squared_error, n_obs = _psnr_update(preds, target, dim=dim)
+    return _psnr_compute(sum_squared_error, n_obs, data_range, base=base, reduction=reduction)
+
+
+def psnr(
+    preds: Tensor,
+    target: Tensor,
+    data_range: Optional[float] = None,
+    base: float = 10.0,
+    reduction: str = "elementwise_mean",
+    dim: Optional[Union[int, Tuple[int, ...]]] = None,
+) -> Tensor:
+    """Computes the peak signal-to-noise ratio.
+        . deprecated:: v0.7
+        Use :function:torchmetrics.functional.psnr. Will be removed in v0.8.
+    Args:
+        preds: estimated signal
+        target: groun truth signal
+        data_range:
+            the range of the data. If None, it is determined from the data (max - min). ``data_range`` must be given
+            when ``dim`` is not None.
+        base: a base of a logarithm to use (default: 10)
+        reduction: a method to reduce metric score over labels.
+
+            - ``'elementwise_mean'``: takes the mean (default)
+            - ``'sum'``: takes the sum
+            - ``'none'``: no reduction will be applied
+
+        dim:
+            Dimensions to reduce PSNR scores over provided as either an integer or a list of integers. Default is
+            None meaning scores will be reduced across all dimensions.
+    Return:
+        Tensor with PSNR score
+
+    Raises:
+        ValueError:
+            If ``dim`` is not ``None`` and ``data_range`` is not provided.
+
+    Example:
+        >>> from torchmetrics.functional import psnr
+        >>> pred = torch.tensor([[0.0, 1.0], [2.0, 3.0]])
+        >>> target = torch.tensor([[3.0, 2.0], [1.0, 0.0]])
+        >>> psnr(pred, target)
+        tensor(2.5527)
+
+    .. note::
+        Half precision is only support on GPU for this metric
+    """
+    if dim is None and reduction != "elementwise_mean":
+        rank_zero_warn(f"The `reduction={reduction}` will not have any effect when `dim` is None.")
+
+    if data_range is None:
+        if dim is not None:
+            # Maybe we could use `torch.amax(target, dim=dim) - torch.amin(target, dim=dim)` in PyTorch 1.7 to calculate
+            # `data_range` in the future.
+            raise ValueError("The `data_range` must be given when `dim` is not None.")
+
+        data_range = target.max() - target.min()
+    else:
+        data_range = tensor(float(data_range))
+    sum_squared_error, n_obs = _psnr_update(preds, target, dim=dim)
+    warn(
+        "`psnr` was renamed to `peak_signal_noise_ratio` in v0.7 and it will be removed in v0.8",
+        DeprecationWarning,)
     return _psnr_compute(sum_squared_error, n_obs, data_range, base=base, reduction=reduction)
