+++ torchmetrics/image/__init__.py
@@ -11,5 +11,5 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from torchmetrics.image.psnr import PeakSignalNoiseRatio  # noqa: F401
+from torchmetrics.image.psnr import PeakSignalNoiseRatio,PSNR  # noqa: F401
 from torchmetrics.image.ssim import SSIM  # noqa: F401
