+++ torchmetrics/__init__.py
@@ -39,6 +39,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     ConfusionMatrix,
     F1Score,
     FBeta,
+    FBetaScore,
     HammingDistance,
     Hinge,
     HingeLoss,
@@ -118,6 +119,7 @@ __all__ = [
     "F1",
     "F1Score",
     "FBeta",
+    "FBetaScore",
     "HammingDistance",
     "Hinge",
     "HingeLoss",
