### 2022-01-10 7b6736d
restructure require files (#735)
* restructure requirements
* update setup
