+++ .github/workflows/ci_test-full.yml
@@ -86,7 +86,6 @@ jobs:
         python ./requirements/adjust-versions.py requirements.txt
         python ./requirements/adjust-versions.py requirements/image.txt
         pip install --requirement requirements/devel.txt --upgrade --find-links $PYTORCH_URL
-        pip uninstall -y torchmetrics
         pip list
       shell: bash
 
@@ -130,4 +129,7 @@ jobs:
 
     - name: Integrations
       run: |
+        pip install -r requirements/integrate.txt --quiet --upgrade-strategy only-if-needed --find-links $PYTORCH_URL
+        pip uninstall -y torchmetrics
         python -m pytest integrations -v --durations=25
+      shell: bash
