+++ azure-pipelines.yml
@@ -57,7 +57,6 @@ jobs:
         # python -m pip install "pip==20.1"
         pip install --requirement ./requirements/devel.txt --upgrade-strategy only-if-needed
         pip install mkl-service==2.4.0  # needed for the gpu multiprocessing
-        pip uninstall -y torchmetrics
         pip list
       displayName: 'Install dependencies'
 
@@ -95,5 +94,7 @@ jobs:
     #  condition: succeededOrFailed()
 
     - bash: |
+        pip install -r requirements/integrate.txt --quiet --upgrade-strategy only-if-needed
+        pip uninstall -y torchmetrics
         python -m pytest integrations -v --durations=25
       displayName: 'Integrations'
