### 2022-01-10 2ebc20f
Add ignore statements for mypy
