+++ torchmetrics/text/bleu.py
@@ -137,8 +137,8 @@ class BLEUScore(Metric):
             target = target or reference_corpus
 
         self.preds_len, self.target_len = _bleu_score_update(
-            preds,
-            target,
+            preds,  # type: ignore
+            target,  # type: ignore
             self.numerator,
             self.denominator,
             self.preds_len,
