### 2022-01-10 dc26e33
Update None arguments for BC
