+++ torchmetrics/functional/text/mer.py
@@ -64,8 +64,8 @@ def _mer_compute(errors: Tensor, total: Tensor) -> Tensor:
 
 
 def match_error_rate(
-    preds: Union[str, List[str]],
-    target: Union[str, List[str]],
+    preds: Union[None, str, List[str]] = None,
+    target: Union[None, str, List[str]] = None,
     predictions: Union[None, str, List[str]] = None,
     references: Union[None, str, List[str]] = None,
 ) -> Tensor:
@@ -86,6 +86,11 @@ def match_error_rate(
         >>> match_error_rate(preds=preds, target=target)
         tensor(0.4444)
     """
+    if preds is None and predictions is None:
+        raise ValueError("Either `preds` or `predictions` must be provided.")
+    if target is None and references is None:
+        raise ValueError("Either `target` or `references` must be provided.")
+
     if predictions is not None:
         warn(
             "You are using deprecated argument `predictions` in v0.7 which was renamed to `preds`. "
@@ -101,5 +106,8 @@ def match_error_rate(
         )
         target = references
 
-    errors, total = _mer_update(preds, target)
+    errors, total = _mer_update(
+        preds,  # type: ignore
+        target,  # type: ignore
+    )
     return _mer_compute(errors, total)

[…]
