### 2022-01-10 0d80cd2
Update docstring and add comments for deprecation
