+++ torchmetrics/functional/text/cer.py
@@ -63,8 +63,8 @@ def _cer_compute(errors: Tensor, total: Tensor) -> Tensor:
 def char_error_rate(
     preds: Union[None, str, List[str]] = None,
     target: Union[None, str, List[str]] = None,
-    predictions: Union[None, str, List[str]] = None,
-    references: Union[None, str, List[str]] = None,
+    predictions: Union[None, str, List[str]] = None,  # ToDo: remove in v0.8
+    references: Union[None, str, List[str]] = None,  # ToDo: remove in v0.8
 ) -> Tensor:
     """character error rate is a common metric of the performance of an automatic speech recognition system. This
     value indicates the percentage of characters that were incorrectly predicted. The lower the value, the better the
@@ -72,6 +72,12 @@ def char_error_rate(
     Args:
         preds: Transcription(s) to score as a string or list of strings
         target: Reference(s) for each speech input as a string or list of strings
+        predictions:
+            .. deprecated:: v0.7
+                This argument is deprecated in favor of  `preds` and will be removed in v0.8.
+            references:
+            .. deprecated:: v0.7
+                This argument is deprecated in favor of  `target` and will be removed in v0.8.
     Returns:
         Character error rate score
     Examples:
