+++ torchmetrics/functional/text/wer.py
@@ -65,8 +65,8 @@ def _wer_compute(errors: Tensor, total: Tensor) -> Tensor:
 def word_error_rate(
     preds: Union[None, str, List[str]] = None,
     target: Union[None, str, List[str]] = None,
-    predictions: Union[None, str, List[str]] = None,
-    references: Union[None, str, List[str]] = None,
+    predictions: Union[None, str, List[str]] = None,  # ToDo: remove in v0.8
+    references: Union[None, str, List[str]] = None,  # ToDo: remove in v0.8
 ) -> Tensor:
     """Word error rate (WER_) is a common metric of the performance of an automatic speech recognition system. This
     value indicates the percentage of words that were incorrectly predicted. The lower the value, the better the
@@ -75,7 +75,12 @@ def word_error_rate(
     Args:
         preds: Transcription(s) to score as a string or list of strings
         target: Reference(s) for each speech input as a string or list of strings
-
+        predictions:
+            .. deprecated:: v0.7
+                This argument is deprecated in favor of  `preds` and will be removed in v0.8.
+            references:
+            .. deprecated:: v0.7
+                This argument is deprecated in favor of  `target` and will be removed in v0.8.
     Returns:
         Word error rate score
 
@@ -113,7 +118,7 @@ def word_error_rate(
     return _wer_compute(errors, total)
 
 
-def wer(
+def wer(  # ToDo: remove in v0.8
     preds: Union[None, str, List[str]] = None,
     target: Union[None, str, List[str]] = None,
     predictions: Union[None, str, List[str]] = None,
