### 2022-01-10 ab391e2
Fix indentation of docstring
