+++ torchmetrics/functional/audio/pit.py
@@ -104,8 +104,8 @@ def _find_best_perm_by_exhuastive_method(
 def permutation_invariant_training(
     preds: torch.Tensor, target: torch.Tensor, metric_func: Callable, eval_func: str = "max", **kwargs: Dict[str, Any]
 ) -> Tuple[Tensor, Tensor]:
-    """Permutation invariant training (PIT). The ``permutation_invariant_training`` implements the famous Permutation
-    Invariant Training method.
+    """Permutation invariant training (PIT). The ``permutation_invariant_training`` implements the famous
+    Permutation Invariant Training method.
 
     [1] in speech separation field in order to calculate audio metrics in a permutation invariant way.
 
