### 2022-01-11 faffa86
Apply suggestions from code review
Co-authored-by: Ethan Harris <ewah1g13@soton.ac.uk>
