### 2022-01-11 d36fffc
Unify r/fbeta/fbeta_score (#740)
* fbeta/fbeta_score
* Apply suggestions from code review

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
