+++ torchmetrics/classification/matthews_corrcoef.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, Callable, Optional
-from warnings import warn
 
 import torch
+from deprecate import deprecated, void
 from torch import Tensor
 
 from torchmetrics.functional.classification.matthews_corrcoef import (
@@ -126,6 +126,7 @@ class MatthewsCorrcoef(MatthewsCorrCoef):
         Renamed in favor of :class:`torchmetrics.MatthewsCorrCoef`. Will be removed in v0.8.
     """
 
+    @deprecated(target=MatthewsCorrCoef, deprecated_in="0.7", remove_in="0.8")
     def __init__(
         self,
         num_classes: int,
@@ -135,8 +136,4 @@ class MatthewsCorrcoef(MatthewsCorrCoef):
         process_group: Optional[Any] = None,
         dist_sync_fn: Callable = None,
     ) -> None:
-        warn(
-            "`MatthewsCorrcoef` was renamed to `MatthewsCorrCoef` in v0.7 and it will be removed in v0.8",
-            DeprecationWarning,
-        )
-        super().__init__(num_classes, threshold, compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
+        void(num_classes, threshold, compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
