+++ torchmetrics/text/ter.py
@@ -118,9 +118,12 @@ class TranslationEditRate(Metric):
             target:
                 An iterable of iterables of reference corpus.
             hypothesis_corpus:
+
                 .. deprecated:: v0.7
                     This argument is deprecated in favor of  `preds` and will be removed in v0.8.
+
             reference_corpus:
+
                 .. deprecated:: v0.7
                     This argument is deprecated in favor of  `target` and will be removed in v0.8.
         """
