+++ torchmetrics/text/mer.py
@@ -96,9 +96,12 @@ class MatchErrorRate(Metric):
             preds: Transcription(s) to score as a string or list of strings
             target: Reference(s) for each speech input as a string or list of strings
             predictions:
+
                 .. deprecated:: v0.7
                     This argument is deprecated in favor of  `preds` and will be removed in v0.8.
+
             references:
+
                 .. deprecated:: v0.7
                     This argument is deprecated in favor of  `target` and will be removed in v0.8.
         """
