+++ torchmetrics/text/squad.py
@@ -123,6 +123,7 @@ class SQuAD(Metric):
                         'title': 'train test'
                     }
             targets:
+
                 .. deprecated:: v0.7
                     This argument is deprecated in favor of  `target` and will be removed in v0.8.
 
