### 2022-01-12 ae2afd1
Add newline in docstring for cer
