+++ torchmetrics/functional/text/cer.py
@@ -72,13 +72,15 @@ def char_error_rate(
     Args:
         preds: Transcription(s) to score as a string or list of strings
         target: Reference(s) for each speech input as a string or list of strings
-        .. deprecated:: v0.7
+    Returns:
+        Character error rate score
+
+    .. deprecated:: v0.7
+        Args:
             predictions:
                 This argument is deprecated in favor of  `preds` and will be removed in v0.8.
             references:
                 This argument is deprecated in favor of  `target` and will be removed in v0.8.
-    Returns:
-        Character error rate score
     Examples:
         >>> preds = ["this is the prediction", "there is an other sample"]
         >>> target = ["this is the reference", "there is another one"]
