### 2022-01-12 d9502f5
Update docstrings for deprecated arguments based on v2
