+++ torchmetrics/functional/text/ter.py
@@ -575,16 +575,6 @@ def translation_edit_rate(
             An iterable of hypothesis corpus.
         target:
             An iterable of iterables of reference corpus.
-        hypothesis_corpus:
-
-            .. deprecated:: v0.7
-                This argument is deprecated in favor of  `preds` and will be removed in v0.8.
-
-        reference_corpus:
-
-            .. deprecated:: v0.7
-                This argument is deprecated in favor of  `target` and will be removed in v0.8.
-
         normalize:
             An indication whether a general tokenization to be applied.
         no_punctuation:
@@ -599,6 +589,13 @@ def translation_edit_rate(
     Return:
         A corpus-level translation edit rate (TER).
         (Optionally) A list of sentence-level translation_edit_rate (TER) if `return_sentence_level_score=True`.
+
+    .. deprecated:: v0.7
+        Args:
+            hypothesis_corpus:
+                This argument is deprecated in favor of  `preds` and will be removed in v0.8.
+            reference_corpus:
+                This argument is deprecated in favor of  `target` and will be removed in v0.8.
 
     Example:
         >>> preds = ['the cat is on the mat']
