### 2022-01-12 abba3ea
rename `permutation_invariant_training` (#737)
* unify `pit` to `permutation_invariant_training`
* To remain PIT, pit
* Apply suggestions from code review

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
