+++ torchmetrics/functional/pairwise/__init__.py
@@ -15,3 +15,4 @@ from torchmetrics.functional.pairwise.cosine import pairwise_cosine_similarity  
 from torchmetrics.functional.pairwise.euclidean import pairwise_euclidean_distance  # noqa: F401
 from torchmetrics.functional.pairwise.linear import pairwise_linear_similarity  # noqa: F401
 from torchmetrics.functional.pairwise.manhattan import pairwise_manhattan_distance  # noqa: F401
+from torchmetrics.functional.pairwise.manhatten import pairwise_manhatten_distance  # noqa: F401
