### 2022-01-12 4ceed45
structural_similarity_index_measure
