### 2022-01-12 3a0ec57
StructuralSimilarityIndexMeasure
