+++ torchmetrics/__init__.py
@@ -55,7 +55,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
-from torchmetrics.image import PSNR, SSIM, MultiScaleStructuralSimilarityIndexMeasure  # noqa: E402
+from torchmetrics.image import PSNR, SSIM, StructuralSimilarityIndexMeasure, MultiScaleStructuralSimilarityIndexMeasure  # noqa: E402
 from torchmetrics.metric import Metric  # noqa: E402
 from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
@@ -175,6 +175,7 @@ __all__ = [
     "Specificity",
     "SQuAD",
     "SSIM",
+    "StructuralSimilarityIndexMeasure",
     "StatScores",
     "SumMetric",
     "SymmetricMeanAbsolutePercentageError",
