+++ tests/image/test_ssim.py
@@ -21,7 +21,7 @@ from skimage.metrics import structural_similarity
 from tests.helpers import seed_all
 from tests.helpers.testers import BATCH_SIZE, NUM_BATCHES, MetricTester
 from torchmetrics.functional import structural_similarity_index_measure
-from torchmetrics.image import SSIM
+from torchmetrics.image import StructuralSimilarityIndexMeasure
 
 seed_all(42)
 
@@ -79,7 +79,7 @@ class TestSSIM(MetricTester):
             ddp,
             preds,
             target,
-            SSIM,
+            StructuralSimilarityIndexMeasure,
             partial(_sk_ssim, data_range=1.0, multichannel=multichannel, kernel_size=kernel_size),
             metric_args={"data_range": 1.0, "kernel_size": (kernel_size, kernel_size)},
             dist_sync_on_step=dist_sync_on_step,
@@ -97,11 +97,11 @@ class TestSSIM(MetricTester):
     # SSIM half + cpu does not work due to missing support in torch.log
     @pytest.mark.xfail(reason="SSIM metric does not support cpu + half precision")
     def test_ssim_half_cpu(self, preds, target, multichannel, kernel_size):
-        self.run_precision_test_cpu(preds, target, SSIM, structural_similarity_index_measure, {"data_range": 1.0})
+        self.run_precision_test_cpu(preds, target, StructuralSimilarityIndexMeasure, structural_similarity_index_measure, {"data_range": 1.0})
 
     @pytest.mark.skipif(not torch.cuda.is_available(), reason="test requires cuda")
     def test_ssim_half_gpu(self, preds, target, multichannel, kernel_size):
-        self.run_precision_test_gpu(preds, target, SSIM, structural_similarity_index_measure, {"data_range": 1.0})
+        self.run_precision_test_gpu(preds, target, StructuralSimilarityIndexMeasure, structural_similarity_index_measure, {"data_range": 1.0})
 
 
 @pytest.mark.parametrize(
