### 2022-01-12 2e9fd18
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
