### 2022-01-12 dcc56b5
simplify regression modules
