+++ tests/regression/test_mean_error.py
@@ -33,7 +33,7 @@ from torchmetrics.functional import (
     mean_squared_error,
     mean_squared_log_error,
 )
-from torchmetrics.functional.regression.symmetric_mean_absolute_percentage_error import (
+from torchmetrics.functional.regression.symmetric_mape import (
     symmetric_mean_absolute_percentage_error,
 )
 from torchmetrics.regression import (
@@ -42,7 +42,7 @@ from torchmetrics.regression import (
     MeanSquaredError,
     MeanSquaredLogError,
 )
-from torchmetrics.regression.symmetric_mean_absolute_percentage_error import SymmetricMeanAbsolutePercentageError
+from torchmetrics.regression.symmetric_mape import SymmetricMeanAbsolutePercentageError
 from torchmetrics.utilities.imports import _TORCH_GREATER_EQUAL_1_6
 
 seed_all(42)

[…]
