+++ torchmetrics/regression/log_mse.py
@@ -16,10 +16,7 @@ from typing import Any, Callable, Optional
 import torch
 from torch import Tensor, tensor
 
-from torchmetrics.functional.regression.log_mse import (
-    _mean_squared_log_error_compute,
-    _mean_squared_log_error_update,
-)
+from torchmetrics.functional.regression.log_mse import _mean_squared_log_error_compute, _mean_squared_log_error_update
 from torchmetrics.metric import Metric
 
 
