+++ torchmetrics/functional/image/ssim.py
@@ -11,12 +11,13 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import List, Literal, Optional, Sequence, Tuple, Union
+from typing import List, Optional, Sequence, Tuple, Union
 from warnings import warn
 
 import torch
 from torch import Tensor
 from torch.nn import functional as F
+from typing_extensions import Literal
 
 from torchmetrics.utilities.checks import _check_same_shape
 from torchmetrics.utilities.distributed import reduce
