### 2022-01-12 93e4311
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
