### 2022-01-12 c8a3463
Apply suggestions from code review
