+++ docs/source/pages/overview.rst
@@ -168,7 +168,7 @@ the following limitations:
   but they are also listed below:
 
   - :ref:`references/modules:PSNR` and :ref:`references/functional:psnr [func]`
-  - :ref:`references/modules:SSIM` and :ref:`references/functional:ssim [func]`
+  - :ref:`references/modules:StructuralSimilarityIndexMeasure` and :ref:`references/functional:structural_similarity_index_measure [func]`
   - :ref:`references/modules:KLDivergence` and :ref:`references/functional:kl_divergence [func]`
 
 You can always check the precision/dtype of the metric by checking the `.dtype` property.
