### 2022-01-12 e632c1d
Apply some suggestions from code review
