+++ torchmetrics/text/sacre_bleu.py
@@ -124,11 +124,7 @@ class SacreBLEUScore(BLEUScore):
         deprecated_in="0.7",
         remove_in="0.8",
     )
-    def update(  # type: ignore
-        self,
-        preds: Sequence[str],
-        target: Sequence[Sequence[str]],
-    ) -> None:
+    def update(self, preds: Sequence[str], target: Sequence[Sequence[str]]) -> None:  # type: ignore
         """Compute Precision Scores.
 
         Args:
