### 2022-01-12 ba41a57
Remove deprecation handling in MER, TER and SQUAD
