+++ torchmetrics/functional/text/ter.py
@@ -36,7 +36,6 @@
 import re
 from functools import lru_cache
 from typing import Dict, Iterator, List, Optional, Sequence, Tuple, Union
-from warnings import warn
 
 from torch import Tensor, tensor
 
@@ -555,10 +554,8 @@ def _ter_compute(total_num_edits: Tensor, total_tgt_length: Tensor) -> Tensor:
 
 
 def translation_edit_rate(
-    preds: Union[None, str, Sequence[str]] = None,
-    target: Union[None, Sequence[Union[str, Sequence[str]]]] = None,
-    hypothesis_corpus: Union[None, str, Sequence[str]] = None,  # ToDo: remove in v0.8
-    reference_corpus: Union[None, Sequence[Union[str, Sequence[str]]]] = None,  # ToDo: remove in v0.8
+    preds: Union[str, Sequence[str]],
+    target: Sequence[Union[str, Sequence[str]]],
     normalize: bool = False,
     no_punctuation: bool = False,
     lowercase: bool = True,
@@ -590,13 +587,6 @@ def translation_edit_rate(
         A corpus-level translation edit rate (TER).
         (Optionally) A list of sentence-level translation_edit_rate (TER) if `return_sentence_level_score=True`.
 
-    .. deprecated:: v0.7
-        Args:
-            hypothesis_corpus:
-                This argument is deprecated in favor of  `preds` and will be removed in v0.8.
-            reference_corpus:
-                This argument is deprecated in favor of  `target` and will be removed in v0.8.
-
     Example:
         >>> preds = ['the cat is on the mat']
         >>> target = [['there is a cat on the mat', 'a cat is on the mat']]
@@ -616,27 +606,6 @@ def translation_edit_rate(
     if not isinstance(asian_support, bool):
         raise ValueError(f"Expected argument `asian_support` to be of type boolean but got {asian_support}.")
 
-    if preds is None and hypothesis_corpus is None:
-        raise ValueError("Either `preds` or `hypothesis_corpus` must be provided.")
-    if target is None and reference_corpus is None:
-        raise ValueError("Either `target` or `reference_corpus` must be provided.")
-
-    if hypothesis_corpus is not None:
-        warn(
-            "You are using deprecated argument `hypothesis_corpus` in v0.7 which was renamed to `preds`. "
-            " The past argument will be removed in v0.8.",
-            DeprecationWarning,
-        )
-        preds = hypothesis_corpus
-
-    if reference_corpus is not None:
-        warn(
-            "You are using deprecated argument `reference_corpus` in v0.7 which was renamed to `target`. "
-            " The past argument will be removed in v0.8.",
-            DeprecationWarning,
-        )
-        target = reference_corpus
-
     tokenizer: _TercomTokenizer = _TercomTokenizer(normalize, no_punctuation, lowercase, asian_support)
 
     total_num_edits = tensor(0.0)
@@ -644,8 +613,8 @@ def translation_edit_rate(
     sentence_ter: Optional[List[Tensor]] = [] if return_sentence_level_score else None
 
     total_num_edits, total_tgt_length, sentence_ter = _ter_update(
-        preds,  # type: ignore
-        target,  # type: ignore
+        preds,
+        target,
         tokenizer,
         total_num_edits,
         total_tgt_length,
