+++ torchmetrics/functional/text/squad.py
@@ -18,7 +18,6 @@ import re
 import string
 from collections import Counter
 from typing import Any, Callable, Dict, List, Tuple, Union
-from warnings import warn
 
 from torch import Tensor, tensor
 
@@ -200,11 +199,7 @@ def _squad_compute(f1: Tensor, exact_match: Tensor, total: Tensor) -> Dict[str, 
     return {"exact_match": exact_match, "f1": f1}
 
 
-def squad(
-    preds: PREDS_TYPE,
-    target: Union[None, TARGETS_TYPE] = None,
-    targets: Union[None, TARGETS_TYPE] = None,  # ToDo: remove in v0.8
-) -> Dict[str, Tensor]:
+def squad(preds: PREDS_TYPE, target: TARGETS_TYPE) -> Dict[str, Tensor]:
     """Calculate `SQuAD Metric`_ .
 
     Args:
@@ -245,11 +240,6 @@ def squad(
     Return:
         Dictionary containing the F1 score, Exact match score for the batch.
 
-    .. deprecated:: v0.7
-        Args:
-            targets:
-                This argument is deprecated in favor of  `target` and will be removed in v0.8.
-
     Example:
         >>> from torchmetrics.functional.text.squad import squad
         >>> preds = [{"prediction_text": "1976", "id": "56e10a3be3433e1400422b22"}]
@@ -265,17 +255,6 @@ def squad(
         [1] SQuAD: 100,000+ Questions for Machine Comprehension of Text by Pranav Rajpurkar, Jian Zhang, Konstantin
         Lopyrev, Percy Liang `SQuAD Metric`_ .
     """
-    if target is None and targets is None:
-        raise ValueError("Either `target` or `targets` must be provided.")
-
-    if targets is not None:
-        warn(
-            "You are using deprecated argument `targets` in v0.7 which was renamed to `target`. "
-            " The past argument will be removed in v0.8.",
-            DeprecationWarning,
-        )
-        target = targets
-
-    preds_dict, target_dict = _squad_input_check(preds, target)  # type: ignore
+    preds_dict, target_dict = _squad_input_check(preds, target)
     f1, exact_match, total = _squad_update(preds_dict, target_dict)
     return _squad_compute(f1, exact_match, total)
