+++ torchmetrics/text/squad.py
@@ -11,8 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Dict, Optional, Union
-from warnings import warn
+from typing import Any, Callable, Dict, Optional
 
 import torch
 from torch import Tensor
@@ -81,12 +80,7 @@ class SQuAD(Metric):
         self.add_state(name="exact_match", default=torch.tensor(0, dtype=torch.float), dist_reduce_fx="sum")
         self.add_state(name="total", default=torch.tensor(0, dtype=torch.int), dist_reduce_fx="sum")
 
-    def update(  # type: ignore
-        self,
-        preds: PREDS_TYPE,
-        target: Union[None, TARGETS_TYPE] = None,
-        targets: Union[None, TARGETS_TYPE] = None,  # ToDo: remove in v0.8
-    ) -> None:
+    def update(self, preds: PREDS_TYPE, target: TARGETS_TYPE) -> None:  # type: ignore
         """Compute F1 Score and Exact Match for a collection of predictions and references.
 
         Args:
@@ -123,28 +117,12 @@ class SQuAD(Metric):
                         'title': 'train test'
                     }
 
-        .. deprecated:: v0.7
-            Args:
-                targets:
-                    This argument is deprecated in favor of  `target` and will be removed in v0.8.
-
         Raises:
             KeyError:
                 If the required keys are missing in either predictions or targets.
         """
-        if target is None and targets is None:
-            raise ValueError("Either `target` or `targets` must be provided.")
-
-        if targets is not None:
-            warn(
-                "You are using deprecated argument `targets` in v0.7 which was renamed to `target`. "
-                " The past argument will be removed in v0.8.",
-                DeprecationWarning,
-            )
-            target = targets
-
-        preds_dict, target_dict = _squad_input_check(preds, target)  # type: ignore
-        f1_score, exact_match, total = _squad_update(preds_dict, target_dict)  # type: ignore
+        preds_dict, target_dict = _squad_input_check(preds, target)
+        f1_score, exact_match, total = _squad_update(preds_dict, target_dict)
         self.f1_score += f1_score
         self.exact_match += exact_match
         self.total += total

[…]
