### 2022-01-12 47fd3b9
Drop deprecated warning where not needed + add deprecated info to doc
