+++ torchmetrics/functional/text/wip.py
@@ -11,10 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import logging
 from typing import List, Tuple, Union
 
-from deprecate import deprecated
 from torch import Tensor, tensor
 
 from torchmetrics.functional.text.helper import _edit_distance
@@ -66,13 +64,6 @@ def _wip_compute(errors: Tensor, target_total: Tensor, preds_total: Tensor) -> T
     return (errors / target_total) * (errors / preds_total)
 
 
-@deprecated(
-    args_mapping={"predictions": "preds", "references": "target"},
-    target=True,
-    stream=logging.warning,
-    deprecated_in="0.7",
-    remove_in="0.8",
-)
 def word_information_preserved(
     preds: Union[str, List[str]],
     target: Union[str, List[str]],
