### 2022-01-12 4c87069
simplify regression modules (#748)
