### 2022-01-12 b4ceaea
rename `peak_signal_noise_ratio` (#732)
* psnr --> peak_signal_noise_ratio
* PSNR --> PeakSignalNoiseRatio
* backward compatibility
* docs

Co-authored-by: Jirka <jirka.borovec@seznam.cz>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
