### 2022-01-12 83345d7
rename SSIM (#747)
* structural_similarity_index_measure
* StructuralSimilarityIndexMeasure
* back compatible
* docs
* chlog

* ms_ssim

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
