### 2022-01-12 d3cab93
refactor: pyDeprecate (#745)
* deprecate: class
* deprecate: function
* cleaning
