+++ torchmetrics/image/ssim.py
@@ -12,9 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, List, Optional, Sequence, Tuple
-from warnings import warn
 
 import torch
+from deprecate import deprecated, void
 from torch import Tensor
 from typing_extensions import Literal
 
@@ -122,6 +122,7 @@ class SSIM(StructuralSimilarityIndexMeasure):
         tensor(0.9219)
     """
 
+    @deprecated(target=StructuralSimilarityIndexMeasure, deprecated_in="0.7", remove_in="0.8")
     def __init__(
         self,
         kernel_size: Sequence[int] = (11, 11),
@@ -134,13 +135,7 @@ class SSIM(StructuralSimilarityIndexMeasure):
         dist_sync_on_step: bool = False,
         process_group: Optional[Any] = None,
     ) -> None:
-        warn(
-            "`SSIM` was renamed to `StructuralSimilarityIndexMeasure` in v0.7 and it will be removed in v0.8",
-            DeprecationWarning,
-        )
-        super().__init__(
-            kernel_size, sigma, reduction, data_range, k1, k2, compute_on_step, dist_sync_on_step, process_group
-        )
+        void(kernel_size, sigma, reduction, data_range, k1, k2, compute_on_step, dist_sync_on_step, process_group)
 
 
 class MultiScaleStructuralSimilarityIndexMeasure(Metric):
