+++ requirements/integrate.txt
@@ -1,1 +1,1 @@
-pytorch-lightning>=1.1
+pytorch-lightning>=1.2
