### 2022-01-12 0a0d80a
Switch to default stream method
