+++ torchmetrics/text/bleu.py
@@ -16,7 +16,6 @@
 # Authors: torchtext authors and @sluks
 # Date: 2020-07-18
 # Link: https://pytorch.org/text/_modules/torchtext/data/metrics.html#bleu_score
-from functools import partial
 from typing import Any, Callable, Optional, Sequence
 from warnings import warn
 
@@ -100,7 +99,6 @@ class BLEUScore(Metric):
     @deprecated(
         args_mapping={"translate_corpus": "preds", "reference_corpus": "target"},
         target=True,
-        stream=partial(warn, category=FutureWarning),
         deprecated_in="0.7",
         remove_in="0.8",
     )
