+++ torchmetrics/text/sacre_bleu.py
@@ -17,7 +17,6 @@
 # Authors: torchtext authors and @sluks
 # Date: 2020-07-18
 # Link: https://pytorch.org/text/_modules/torchtext/data/metrics.html#bleu_score
-from functools import partial
 from typing import Any, Callable, Optional, Sequence
 from warnings import warn
 
@@ -120,7 +119,6 @@ class SacreBLEUScore(BLEUScore):
     @deprecated(
         args_mapping={"translate_corpus": "preds", "reference_corpus": "target"},
         target=True,
-        stream=partial(warn, category=FutureWarning),
         deprecated_in="0.7",
         remove_in="0.8",
     )
