### 2022-01-13 77babc9
Unify `preds, target` input arguments for `text` metrics [1of2] bert, bleu, chrf, sacre_bleu, wip, wil (#723)
* Unify some text metrics
* Unify some other stuff
* Handle BC and add warnings
* Use deprecate package
* Apply some suggestions from code review

Co-authored-by: Scott Cronin <j.scott.cronin@gmail.com>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>
