+++ torchmetrics/functional/text/bert.py
@@ -14,11 +14,12 @@
 import csv
 import math
 import urllib
-import warnings
 from collections import Counter, defaultdict
 from typing import Any, Callable, Dict, List, Optional, Set, Tuple, Union
+from warnings import warn
 
 import torch
+from deprecate import deprecated
 from torch import Tensor
 from torch.utils.data import DataLoader, Dataset
 
@@ -350,24 +351,24 @@ def _get_scaled_precision_or_recall(cos_sim: Tensor, metric: str, idf_scale: Ten
 
 
 def _get_precision_recall_f1(
-    pred_embeddings: Tensor, ref_embeddings: Tensor, pred_idf_scale: Tensor, ref_idf_scale: Tensor
+    preds_embeddings: Tensor, target_embeddings: Tensor, preds_idf_scale: Tensor, target_idf_scale: Tensor
 ) -> Tuple[Tensor, Tensor, Tensor]:
     """Calculate precision, recall and F1 score over candidate and reference sentences.
 
     Args:
-        pred_embeddings: Embeddings of candidate sentenecs.
-        ref_embeddings: Embeddings of reference sentences.
-        pred_idf_scale: An IDF scale factor for candidate sentences.
-        ref_idf_scale: An IDF scale factor for reference sentences.
+        preds_embeddings: Embeddings of candidate sentenecs.
+        target_embeddings: Embeddings of reference sentences.
+        preds_idf_scale: An IDF scale factor for candidate sentences.
+        target_idf_scale: An IDF scale factor for reference sentences.
 
     Return:
         Tensors containing precision, recall and F1 score, respectively.
     """
     # Dimensions: b = batch_size, l = num_layers, p = predictions_seq_len, r = references_seq_len, d = bert_dim
-    cos_sim = torch.einsum("blpd, blrd -> blpr", pred_embeddings, ref_embeddings)
+    cos_sim = torch.einsum("blpd, blrd -> blpr", preds_embeddings, target_embeddings)
     # Final metrics shape = (batch_size * num_layers | batch_size)
-    precision = _get_scaled_precision_or_recall(cos_sim, "precision", pred_idf_scale)
-    recall = _get_scaled_precision_or_recall(cos_sim, "recall", ref_idf_scale)
+    precision = _get_scaled_precision_or_recall(cos_sim, "precision", preds_idf_scale)
+    recall = _get_scaled_precision_or_recall(cos_sim, "recall", target_idf_scale)
 
     f1_score = 2 * precision * recall / (precision + recall)
     f1_score = f1_score.masked_fill(torch.isnan(f1_score), 0.0)
@@ -426,7 +427,7 @@ def _load_baseline(
         baseline = _read_csv_from_url(baseline_url)
     else:
         baseline = None
-        warnings.warn("Baseline was not successfully loaded. No baseline is going to be used.")
+        warn("Baseline was not successfully loaded. No baseline is going to be used.")
 
     return baseline
 
@@ -449,9 +450,15 @@ def _rescale_metrics_with_baseline(
     return all_metrics[..., 0], all_metrics[..., 1], all_metrics[..., 2]
 
 
+@deprecated(
+    args_mapping={"predictions": "preds", "references": "target"},
+    target=True,
+    deprecated_in="0.7",
+    remove_in="0.8",
+)
 def bert_score(
-    predictions: Union[List[str], Dict[str, Tensor]],
-    references: Union[List[str], Dict[str, Tensor]],
+    preds: Union[List[str], Dict[str, Tensor]],
+    target: Union[List[str], Dict[str, Tensor]],
     model_name_or_path: Optional[str] = None,
     num_layers: Optional[int] = None,
     all_layers: bool = False,
[…]

[…]
