### 2022-01-13 9094985
Unify `preds, target` input arguments for `text` metrics [2of2] cer, ter, wer, mer, rouge, squad (#727)
* Update naming conventions for cer, ter, wer, sacrebleu, rouge, squad
* Add warnings for BC

Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
Co-authored-by: Jirka Borovec <Borda@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
