+++ torchmetrics/__about__.py
@@ -1,4 +1,4 @@
-__version__ = "0.7.0dev"
+__version__ = "0.7.0rc0"
 __author__ = "PyTorchLightning et al."
 __author_email__ = "name@pytorchlightning.ai"
 __license__ = "Apache-2.0"
