### 2022-01-13 bd70c40
Changing deprecation warnings to FutureWarning
