+++ torchmetrics/functional/image/psnr.py
@@ -18,6 +18,7 @@ from deprecate import deprecated, void
 from torch import Tensor, tensor
 
 from torchmetrics.utilities import rank_zero_warn, reduce
+from torchmetrics.utilities.warn import _future_warning
 
 
 def _psnr_compute(
@@ -151,7 +152,7 @@ def peak_signal_noise_ratio(
     return _psnr_compute(sum_squared_error, n_obs, data_range, base=base, reduction=reduction)
 
 
-@deprecated(target=peak_signal_noise_ratio, deprecated_in="0.7", remove_in="0.8")
+@deprecated(target=peak_signal_noise_ratio, deprecated_in="0.7", remove_in="0.8", stream=_future_warning)
 def psnr(
     preds: Tensor,
     target: Tensor,
