+++ torchmetrics/__init__.py
@@ -55,6 +55,7 @@ from torchmetrics.classification import (  # noqa: E402, F401
     Specificity,
     StatScores,
 )
+from torchmetrics.collections import MetricCollection  # noqa: E402
 from torchmetrics.image import (  # noqa: E402
     PSNR,
     SSIM,
@@ -63,7 +64,6 @@ from torchmetrics.image import (  # noqa: E402
     StructuralSimilarityIndexMeasure,
 )
 from torchmetrics.metric import Metric  # noqa: E402
-from torchmetrics.metric_collections import MetricCollection  # noqa: E402
 from torchmetrics.regression import (  # noqa: E402
     CosineSimilarity,
     ExplainedVariance,
