+++ torchmetrics/utilities/imports.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 """Import utilities."""
 import operator
+from collections import OrderedDict  # noqa: F401
 from importlib import import_module
 from importlib.util import find_spec
 from typing import Callable, Optional
