### 2022-01-13 4dbdc77
Changing default stream in deprecation warnings to FutureWarning (#749)
* Changing deprecation warnings to FutureWarning
* move to prints
* clean

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka <jirka.borovec@seznam.cz>
