+++ torchmetrics/functional/audio/si_sdr.py
@@ -17,9 +17,12 @@ from deprecate import deprecated, void
 from torch import Tensor
 
 from torchmetrics.functional.audio import scale_invariant_signal_distortion_ratio
+from torchmetrics.utilities import _future_warning
 
 
-@deprecated(target=scale_invariant_signal_distortion_ratio, deprecated_in="0.7", remove_in="0.8")
+@deprecated(
+    target=scale_invariant_signal_distortion_ratio, deprecated_in="0.7", remove_in="0.8", stream=_future_warning
+)
 def si_sdr(preds: Tensor, target: Tensor, zero_mean: bool = False) -> Tensor:
     """Calculates Scale-invariant signal-to-distortion ratio (SI-SDR) metric.
 
