+++ tests/image/test_fid.py
@@ -18,7 +18,7 @@ import torch
 from scipy.linalg import sqrtm as scipy_sqrtm
 from torch.utils.data import Dataset
 
-from torchmetrics.image.fid import FID, sqrtm
+from torchmetrics.image.fid import FrechetInceptionDistance, sqrtm
 from torchmetrics.utilities.imports import _TORCH_FIDELITY_AVAILABLE
 
 torch.manual_seed(42)
@@ -47,7 +47,7 @@ def test_no_train():
     class MyModel(torch.nn.Module):
         def __init__(self):
             super().__init__()
-            self.metric = FID()
+            self.metric = FrechetInceptionDistance()
 
         def forward(self, x):
             return x
@@ -61,7 +61,7 @@ def test_no_train():
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
 def test_fid_pickle():
     """Assert that we can initialize the metric and pickle it."""
-    metric = FID()
+    metric = FrechetInceptionDistance()
     assert metric
 
     # verify metrics work after being loaded from pickled state
@@ -76,21 +76,21 @@ def test_fid_raises_errors_and_warnings():
         match="Metric `FID` will save all extracted features in buffer."
         " For large datasets this may lead to large memory footprint.",
     ):
-        _ = FID()
+        _ = FrechetInceptionDistance()
 
     if _TORCH_FIDELITY_AVAILABLE:
         with pytest.raises(ValueError, match="Integer input to argument `feature` must be one of .*"):
-            _ = FID(feature=2)
+            _ = FrechetInceptionDistance(feature=2)
     else:
         with pytest.raises(
             ModuleNotFoundError,
             match="FID metric requires that `Torch-fidelity` is installed."
             " Either install as `pip install torchmetrics[image-quality]` or `pip install torch-fidelity`.",
         ):
-            _ = FID()
+            _ = FrechetInceptionDistance()
 
     with pytest.raises(TypeError, match="Got unknown input to argument `feature`"):
-        _ = FID(feature=[1, 2])
+        _ = FrechetInceptionDistance(feature=[1, 2])
 
 
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
@@ -98,7 +98,7 @@ def test_fid_raises_errors_and_warnings():
 def test_fid_same_input(feature):
     """if real and fake are update on the same data the fid score should be
     0."""
-    metric = FID(feature=feature)
+    metric = FrechetInceptionDistance(feature=feature)
 
     for _ in range(2):
         img = torch.randint(0, 255, (10, 3, 299, 299), dtype=torch.uint8)
@@ -128,7 +128,7 @@ def test_compare_fid(tmpdir, feature=2048):
     """check that the hole pipeline give the same result as torch-fidelity."""
     from torch_fidelity import calculate_metrics
 
-    metric = FID(feature=feature).cuda()
+    metric = FrechetInceptionDistance(feature=feature).cuda()
 
     # Generate some synthetic data
     img1 = torch.randint(0, 180, (100, 3, 299, 299), dtype=torch.uint8)
