+++ torchmetrics/image/kid.py
@@ -14,12 +14,13 @@
 from typing import Any, Callable, List, Optional, Tuple, Union
 
 import torch
+from deprecate import deprecated, void
 from torch import Tensor
 from torch.nn import Module
 
+from torchmetrics import Metric
 from torchmetrics.image.fid import NoTrainInceptionV3
-from torchmetrics.metric import Metric
-from torchmetrics.utilities import rank_zero_warn
+from torchmetrics.utilities import _future_warning, rank_zero_warn
 from torchmetrics.utilities.data import dim_zero_cat
 from torchmetrics.utilities.imports import _TORCH_FIDELITY_AVAILABLE
 
@@ -62,7 +63,7 @@ def poly_mmd(
     return maximum_mean_discrepancy(k_11, k_12, k_22)
 
 
-class KID(Metric):
+class KernelInceptionDistance(Metric):
     r"""
     Calculates Kernel Inception Distance (KID) which is used to access the quality of generated images. Given by
 
@@ -150,8 +151,8 @@ class KID(Metric):
     Example:
         >>> import torch
         >>> _ = torch.manual_seed(123)
-        >>> from torchmetrics.image.kid import KID
-        >>> kid = KID(subset_size=50)  # doctest: +SKIP
+        >>> from torchmetrics.image.kid import KernelInceptionDistance
+        >>> kid = KernelInceptionDistance(subset_size=50)  # doctest: +SKIP
         >>> # generate two slightly overlapping image intensity distributions
         >>> imgs_dist1 = torch.randint(0, 200, (100, 3, 299, 299), dtype=torch.uint8)  # doctest: +SKIP
         >>> imgs_dist2 = torch.randint(100, 255, (100, 3, 299, 299), dtype=torch.uint8)  # doctest: +SKIP
@@ -187,7 +188,7 @@ class KID(Metric):
         )
 
         rank_zero_warn(
-            "Metric `KID` will save all extracted features in buffer."
+            "Metric `Kernel Inception Distance` will save all extracted features in buffer."
             " For large datasets this may lead to large memory footprint.",
             UserWarning,
         )
@@ -195,7 +196,7 @@ class KID(Metric):
         if isinstance(feature, (str, int)):
             if not _TORCH_FIDELITY_AVAILABLE:
                 raise ModuleNotFoundError(
-                    "KID metric requires that `Torch-fidelity` is installed."
+                    "Kernel Inception Distance metric requires that `Torch-fidelity` is installed."
                     " Either install as `pip install torchmetrics[image]` or `pip install torch-fidelity`."
                 )
             valid_int_input = ("logits_unbiased", 64, 192, 768, 2048)
@@ -275,3 +276,52 @@ class KID(Metric):
             kid_scores_.append(o)
         kid_scores = torch.stack(kid_scores_)
         return kid_scores.mean(), kid_scores.std(unbiased=False)
+
+
+class KID(KernelInceptionDistance):
+    r"""
+    Calculates Kernel Inception Distance (KID) which is used to access the quality of generated images.
+
+    .. deprecated:: v0.7
+        Use :class:`torchmetrics.image.KernelInceptionDistance`. Will be removed in v0.8.
+
+    Example:
+        >>> import torch
+        >>> _ = torch.manual_seed(123)
+        >>> kid = KID(subset_size=50)  # doctest: +SKIP
+        >>> # generate two slightly overlapping image intensity distributions
+        >>> imgs_dist1 = torch.randint(0, 200, (100, 3, 299, 299), dtype=torch.uint8)  # doctest: +SKIP
[…]
