### 2022-01-14 9c43891
Update integrations/test_lightning.py
