### 2022-01-14 6968f08
use own sum/diff
