+++ torchmetrics/functional/audio/__init__.py
@@ -20,3 +20,7 @@ from torchmetrics.functional.audio.sdr import (  # noqa: F401
 from torchmetrics.functional.audio.si_sdr import si_sdr  # noqa: F401
 from torchmetrics.functional.audio.si_snr import si_snr  # noqa: F401
 from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio, signal_noise_ratio, snr  # noqa: F401
+from torchmetrics.utilities.imports import _PESQ_AVAILABLE
+
+if _PESQ_AVAILABLE:
+    from torchmetrics.functional.audio.pesq import perceptual_evaluation_speech_quality
