+++ torchmetrics/audio/pesq.py
@@ -13,15 +13,17 @@
 # limitations under the License.
 from typing import Any, Callable, Optional
 
+from deprecate import deprecated, void
 from torch import Tensor, tensor
 
 from torchmetrics.functional.audio.pesq import perceptual_evaluation_speech_quality
 from torchmetrics.metric import Metric
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.imports import _PESQ_AVAILABLE
 
 
-class PESQ(Metric):
-    """PESQ (Perceptual Evaluation of Speech Quality)
+class PerceptualEvaluationSpeechQuality(Metric):
+    """Perceptual Evaluation of Speech Quality (PESQ)
 
     This is a wrapper for the pesq package [1]. . Note that input will be moved to `cpu`
     to perform the metric calculation.
@@ -62,15 +64,15 @@ class PESQ(Metric):
             If ``mode`` is not either ``"wb"`` or ``"nb"``
 
     Example:
-        >>> from torchmetrics.audio.perceptual_evaluation_speech_quality import PESQ
+        >>> from torchmetrics.audio.pesq import PerceptualEvaluationSpeechQuality
         >>> import torch
         >>> g = torch.manual_seed(1)
         >>> preds = torch.randn(8000)
         >>> target = torch.randn(8000)
-        >>> nb_pesq = PESQ(8000, 'nb')
+        >>> nb_pesq = PerceptualEvaluationSpeechQuality(8000, 'nb')
         >>> nb_pesq(preds, target)
         tensor(2.2076)
-        >>> wb_pesq = PESQ(16000, 'wb')
+        >>> wb_pesq = PerceptualEvaluationSpeechQuality(16000, 'wb')
         >>> wb_pesq(preds, target)
         tensor(1.7359)
 
@@ -100,7 +102,7 @@ class PESQ(Metric):
         )
         if not _PESQ_AVAILABLE:
             raise ModuleNotFoundError(
-                "PESQ metric requires that pesq is installed."
+                "PerceptualEvaluationSpeechQuality metric requires that pesq is installed."
                 " Either install as `pip install torchmetrics[audio]` or `pip install pesq`."
             )
         if fs not in (8000, 16000):
@@ -128,3 +130,42 @@ class PESQ(Metric):
     def compute(self) -> Tensor:
         """Computes average PESQ."""
         return self.sum_pesq / self.total
+
+
+class PESQ(PerceptualEvaluationSpeechQuality):
+    """Perceptual Evaluation of Speech Quality (PESQ).
+
+    .. deprecated:: v0.7
+        Use :class:`torchmetrics.audio.PermutationInvariantTraining`. Will be removed in v0.8.
+
+    Example:
+        >>> import torch
+        >>> g = torch.manual_seed(1)
+        >>> preds = torch.randn(8000)
+        >>> target = torch.randn(8000)
+        >>> nb_pesq = PESQ(8000, 'nb')
+        >>> nb_pesq(preds, target)
+        tensor(2.2076)
+        >>> wb_pesq = PESQ(16000, 'wb')
+        >>> wb_pesq(preds, target)
+        tensor(1.7359)
+    """
+
+    @deprecated(target=PerceptualEvaluationSpeechQuality, deprecated_in="0.7", remove_in="0.8", stream=_future_warning)
+    def __init__(
+        self,
+        fs: int,
+        mode: str,
+        compute_on_step: bool = True,
+        dist_sync_on_step: bool = False,
+        process_group: Optional[Any] = None,
+        dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
+    ) -> None:
+        void(
+            fs,
+        mode,
+        compute_on_step,
+                                dist_sync_on_step,
+                                                          process_group,
+                                                                                         dist_sync_fn
+        )
