+++ torchmetrics/functional/text/__init__.py
@@ -22,7 +22,6 @@ from torchmetrics.functional.text.ter import translation_edit_rate  # noqa: F401
 from torchmetrics.functional.text.wer import wer, word_error_rate  # noqa: F401
 from torchmetrics.functional.text.wil import word_information_lost  # noqa: F401
 from torchmetrics.functional.text.wip import word_information_preserved  # noqa: F401
-
 from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 if _TRANSFORMERS_AVAILABLE:
