+++ torchmetrics/functional/text/__init__.py
@@ -25,4 +25,4 @@ from torchmetrics.functional.text.wip import word_information_preserved  # noqa:
 from torchmetrics.utilities.imports import _TRANSFORMERS_AVAILABLE
 
 if _TRANSFORMERS_AVAILABLE:
-    from torchmetrics.functional.text.bert import bert_score
+    from torchmetrics.functional.text.bert import bert_score  # noqa: F401
