### 2022-01-14 b195047
InceptionScore
