+++ torchmetrics/image/inception.py
@@ -14,16 +14,17 @@
 from typing import Any, Callable, List, Optional, Tuple, Union
 
 import torch
+from deprecate import deprecated, void
 from torch import Tensor
 
 from torchmetrics.image.fid import NoTrainInceptionV3
 from torchmetrics.metric import Metric
-from torchmetrics.utilities import rank_zero_warn
+from torchmetrics.utilities import _future_warning, rank_zero_warn
 from torchmetrics.utilities.data import dim_zero_cat
 from torchmetrics.utilities.imports import _TORCH_FIDELITY_AVAILABLE
 
 
-class IS(Metric):
+class InceptionScore(Metric):
     r"""
     Calculates the Inception Score (IS) which is used to access how realistic generated images are.
     It is defined as
@@ -91,8 +92,8 @@ class IS(Metric):
     Example:
         >>> import torch
         >>> _ = torch.manual_seed(123)
-        >>> from torchmetrics.image.inception import IS
-        >>> inception = IS()  # doctest: +SKIP
+        >>> from torchmetrics.image.inception import InceptionScore
+        >>> inception = InceptionScore()  # doctest: +SKIP
         >>> # generate some images
         >>> imgs = torch.randint(0, 255, (100, 3, 299, 299), dtype=torch.uint8)  # doctest: +SKIP
         >>> inception.update(imgs)  # doctest: +SKIP
@@ -120,7 +121,7 @@ class IS(Metric):
         )
 
         rank_zero_warn(
-            "Metric `IS` will save all extracted features in buffer."
+            "Metric `InceptionScore` will save all extracted features in buffer."
             " For large datasets this may lead to large memory footprint.",
             UserWarning,
         )
@@ -128,7 +129,7 @@ class IS(Metric):
         if isinstance(feature, (str, int)):
             if not _TORCH_FIDELITY_AVAILABLE:
                 raise ModuleNotFoundError(
-                    "IS metric requires that `Torch-fidelity` is installed."
+                    "InceptionScore metric requires that `Torch-fidelity` is installed."
                     " Either install as `pip install torchmetrics[image]` or `pip install torch-fidelity`."
                 )
             valid_int_input = ("logits_unbiased", 64, 192, 768, 2048)
@@ -177,3 +178,40 @@ class IS(Metric):
 
         # return mean and std
         return kl.mean(), kl.std()
+
+
+class IS(InceptionScore):
+    r"""
+    Calculates the Inception Score (IS) which is used to access how realistic generated images are.
+
+    .. deprecated:: v0.7
+        Use :class:`torchmetrics.image.InceptionScore`. Will be removed in v0.8.
+
+    Example:
+        >>> import torch
+        >>> _ = torch.manual_seed(123)
+        >>> inception = IS()  # doctest: +SKIP
+        >>> # generate some images
+        >>> imgs = torch.randint(0, 255, (100, 3, 299, 299), dtype=torch.uint8)  # doctest: +SKIP
+        >>> inception.update(imgs)  # doctest: +SKIP
+        >>> inception.compute()  # doctest: +SKIP
+        (tensor(1.0569), tensor(0.0113))
+
+    """
+
+    @deprecated(target=InceptionScore, deprecated_in="0.7", remove_in="0.8", stream=_future_warning)
+    def __init__(
+        self,
+        feature: Union[str, int, torch.nn.Module] = "logits_unbiased",
+        splits: int = 10,
+        compute_on_step: bool = False,
+        dist_sync_on_step: bool = False,
+        process_group: Optional[Any] = None,
+        dist_sync_fn: Callable[[Tensor], List[Tensor]] = None,
+    ) -> None:
+        void(feature,
+        splits,
+        compute_on_step,
+        dist_sync_on_step,
+        process_group,
+        dist_sync_fn)
