+++ tests/image/test_inception.py
@@ -38,7 +38,9 @@ def test_no_train():
     model = MyModel()
     model.train()
     assert model.training
-    assert not model.metric.inception.training, "InceptionScore metric was changed to training mode which should not happen"
+    assert (
+        not model.metric.inception.training
+    ), "InceptionScore metric was changed to training mode which should not happen"
 
 
 @pytest.mark.skipif(not _TORCH_FIDELITY_AVAILABLE, reason="test requires torch-fidelity")
