+++ CHANGELOG.md
@@ -70,6 +70,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Renamed image SSIM metric: ([#747](https://github.com/PyTorchLightning/metrics/pull/747))
   * `torchmetrics.functional.ssim` -> `torchmetrics.functional.scale_invariant_signal_noise_ratio`
   * `torchmetrics.SSIM` -> `torchmetrics.StructuralSimilarityIndexMeasure`
+- Renamed Fidelity & LPIPS image metric: ([#752](https://github.com/PyTorchLightning/metrics/pull/752))
+  * `torchmetrics.image.FID` -> `torchmetrics.image.FrechetInceptionDistance`
+  * `torchmetrics.image.KID` -> `torchmetrics.image.KernelInceptionDistance`
+  * `torchmetrics.image.LPIPS` -> `torchmetrics.image.LearnedPerceptualImagePatchSimilarity`
 
 ### Removed
 
