### 2022-01-14 3eb8393
ShortTermObjectiveIntelligibility
