+++ torchmetrics/audio/__init__.py
@@ -16,3 +16,7 @@ from torchmetrics.audio.sdr import SDR, ScaleInvariantSignalDistortionRatio, Sig
 from torchmetrics.audio.si_sdr import SI_SDR  # noqa: F401
 from torchmetrics.audio.si_snr import SI_SNR  # noqa: F401
 from torchmetrics.audio.snr import SNR, ScaleInvariantSignalNoiseRatio, SignalNoiseRatio  # noqa: F401
+from torchmetrics.utilities.imports import _PYSTOI_AVAILABLE
+
+if _PYSTOI_AVAILABLE:
+    from torchmetrics.audio.stoi import ShortTermObjectiveIntelligibility  # noqa: F401
