### 2022-01-14 dff9fc1
MeanAveragePrecision
