+++ torchmetrics/detection/map.py
@@ -15,9 +15,11 @@ import logging
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple
 
 import torch
+from deprecate import deprecated, void
 from torch import IntTensor, Size, Tensor
 
 from torchmetrics.metric import Metric
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.imports import _TORCHVISION_AVAILABLE, _TORCHVISION_GREATER_EQUAL_0_8
 
 if _TORCHVISION_AVAILABLE and _TORCHVISION_GREATER_EQUAL_0_8:
@@ -129,7 +131,7 @@ def _fix_empty_tensors(boxes: Tensor) -> Tensor:
     return boxes
 
 
-class MAP(Metric):
+class MeanAveragePrecision(Metric):
     r"""
     Computes the `Mean-Average-Precision (mAP) and Mean-Average-Recall (mAR)
     <https://jonathan-hui.medium.com/map-mean-average-precision-for-object-detection-45c121a31173>`_
@@ -181,7 +183,7 @@ class MAP(Metric):
 
     Example:
         >>> import torch
-        >>> from torchmetrics.detection.map import MAP
+        >>> from torchmetrics.detection.map import MeanAveragePrecision
         >>> preds = [
         ...   dict(
         ...     boxes=torch.Tensor([[258.0, 41.0, 606.0, 285.0]]),
@@ -195,7 +197,7 @@ class MAP(Metric):
         ...     labels=torch.IntTensor([0]),
         ...   )
         ... ]
-        >>> metric = MAP()  # doctest: +SKIP
+        >>> metric = MeanAveragePrecision()  # doctest: +SKIP
         >>> metric.update(preds, target)  # doctest: +SKIP
         >>> from pprint import pprint
         >>> pprint(metric.compute())  # doctest: +SKIP
@@ -243,7 +245,7 @@ class MAP(Metric):
 
         if not (_TORCHVISION_AVAILABLE and _TORCHVISION_GREATER_EQUAL_0_8):
             raise ModuleNotFoundError(
-                "`MAP` metric requires that `torchvision` version 0.8.0 or newer is installed."
+                "`MeanAveragePrecision` metric requires that `torchvision` version 0.8.0 or newer is installed."
                 " Please install with `pip install torchvision` or `pip install torchmetrics[detection]`."
             )
 
@@ -430,7 +432,7 @@ class MAP(Metric):
         if torch.numel(ious) > 0:
             for idx_iou, t in enumerate(self.iou_thresholds):
                 for idx_det in range(nb_det):
-                    m = MAP._find_best_gt_match(t, nb_gt, gt_matches, idx_iou, gt_ignore, ious, idx_det)
+                    m = MeanAveragePrecision._find_best_gt_match(t, nb_gt, gt_matches, idx_iou, gt_ignore, ious, idx_det)
                     if m != -1:
                         det_ignore[idx_iou, idx_det] = gt_ignore[m]
                         det_matches[idx_iou, idx_det] = True
@@ -573,7 +575,7 @@ class MAP(Metric):
         for idx_cls in range(nb_classes):
             for idx_bbox_area in range(nb_bbox_areas):
                 for idx_max_det_thrs, max_det in enumerate(self.max_detection_thresholds):
-                    recall, precision, scores = MAP.__calculate_recall_precision_scores(
+                    recall, precision, scores = MeanAveragePrecision.__calculate_recall_precision_scores(
                         recall,
                         precision,
                         scores,
@@ -731,3 +733,76 @@ class MAP(Metric):
         metrics.map_per_class = map_per_class_values
[…]
