+++ torchmetrics/retrieval/average_precision.py
@@ -11,29 +11,27 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Optional
-
 from torch import Tensor, tensor
 
-from torchmetrics.functional.retrieval.ndcg import retrieval_normalized_dcg
-from torchmetrics.retrieval.retrieval_metric import RetrievalMetric
+from torchmetrics.functional.retrieval.average_precision import retrieval_average_precision
+from torchmetrics.retrieval.metric_base import RetrievalMetric
 
 
-class RetrievalNormalizedDCG(RetrievalMetric):
-    """Computes `Normalized Discounted Cumulative Gain`_.
+class RetrievalMAP(RetrievalMetric):
+    """Computes `Mean Average Precision`_.
 
-    Works with binary or positive integer target data. Accepts float predictions from a model output.
+    Works with binary target data. Accepts float predictions from a model output.
 
-    Forward accepts:
+    Forward accepts
 
     - ``preds`` (float tensor): ``(N, ...)``
-    - ``target`` (long, int, bool or float tensor): ``(N, ...)``
+    - ``target`` (long or bool tensor): ``(N, ...)``
     - ``indexes`` (long tensor): ``(N, ...)``
 
     ``indexes``, ``preds`` and ``target`` must have the same dimension.
     ``indexes`` indicate to which query a prediction belongs.
-    Predictions will be first grouped by ``indexes`` and then `Normalized Discounted Cumulative Gain`
-    will be computed as the mean of the `Normalized Discounted Cumulative Gain` over each query.
+    Predictions will be first grouped by ``indexes`` and then `MAP` will be computed as the mean
+    of the `Average Precisions` over each query.
 
     Args:
         empty_target_action:
@@ -46,7 +44,6 @@ class RetrievalNormalizedDCG(RetrievalMetric):
 
         ignore_index:
             Ignore predictions where the target is equal to this number.
-        k: consider only the top k elements for each query (default: `None`, which considers them all)
         compute_on_step:
             Forward only calls ``update()`` and return None if this is set to False.
         dist_sync_on_step:
@@ -63,44 +60,18 @@ class RetrievalNormalizedDCG(RetrievalMetric):
             If ``empty_target_action`` is not one of ``error``, ``skip``, ``neg`` or ``pos``.
         ValueError:
             If ``ignore_index`` is not `None` or an integer.
-        ValueError:
-            If ``k`` parameter is not `None` or an integer larger than 0.
 
     Example:
-        >>> from torchmetrics import RetrievalNormalizedDCG
+        >>> from torchmetrics import RetrievalMAP
         >>> indexes = tensor([0, 0, 0, 1, 1, 1, 1])
         >>> preds = tensor([0.2, 0.3, 0.5, 0.1, 0.3, 0.5, 0.2])
         >>> target = tensor([False, False, True, False, True, False, True])
-        >>> ndcg = RetrievalNormalizedDCG()
-        >>> ndcg(preds, target, indexes=indexes)
-        tensor(0.8467)
+        >>> rmap = RetrievalMAP()
+        >>> rmap(preds, target, indexes=indexes)
+        tensor(0.7917)
     """
 
     higher_is_better = True
 
-    def __init__(
-        self,
-        empty_target_action: str = "neg",
-        ignore_index: Optional[int] = None,
-        k: Optional[int] = None,
-        compute_on_step: bool = True,
-        dist_sync_on_step: bool = False,
-        process_group: Optional[Any] = None,
-        dist_sync_fn: Callable = None,
-    ) -> None:
-        super().__init__(
[…]
