+++ torchmetrics/retrieval/hit_rate.py
@@ -13,16 +13,14 @@
 # limitations under the License.
 from typing import Any, Callable, Optional
 
-import torch
 from torch import Tensor, tensor
 
-from torchmetrics.functional.retrieval.fall_out import retrieval_fall_out
-from torchmetrics.retrieval.retrieval_metric import RetrievalMetric
-from torchmetrics.utilities.data import get_group_indexes
+from torchmetrics.functional.retrieval.hit_rate import retrieval_hit_rate
+from torchmetrics.retrieval.metric_base import RetrievalMetric
 
 
-class RetrievalFallOut(RetrievalMetric):
-    """Computes `Fall-out`_.
+class RetrievalHitRate(RetrievalMetric):
+    """Computes `IR HitRate`.
 
     Works with binary target data. Accepts float predictions from a model output.
 
@@ -34,12 +32,12 @@ class RetrievalFallOut(RetrievalMetric):
 
     ``indexes``, ``preds`` and ``target`` must have the same dimension.
     ``indexes`` indicate to which query a prediction belongs.
-    Predictions will be first grouped by ``indexes`` and then `Fall-out` will be computed as the mean
-    of the `Fall-out` over each query.
+    Predictions will be first grouped by ``indexes`` and then the `Hit Rate` will be computed as the mean
+    of the `Hit Rate` over each query.
 
     Args:
         empty_target_action:
-            Specify what to do with queries that do not have at least a negative ``target``. Choose from:
+            Specify what to do with queries that do not have at least a positive ``target``. Choose from:
 
             - ``'neg'``: those queries count as ``0.0`` (default)
             - ``'pos'``: those queries count as ``1.0``
@@ -69,20 +67,20 @@ class RetrievalFallOut(RetrievalMetric):
             If ``k`` parameter is not `None` or an integer larger than 0.
 
     Example:
-        >>> from torchmetrics import RetrievalFallOut
+        >>> from torchmetrics import RetrievalHitRate
         >>> indexes = tensor([0, 0, 0, 1, 1, 1, 1])
         >>> preds = tensor([0.2, 0.3, 0.5, 0.1, 0.3, 0.5, 0.2])
-        >>> target = tensor([False, False, True, False, True, False, True])
-        >>> fo = RetrievalFallOut(k=2)
-        >>> fo(preds, target, indexes=indexes)
+        >>> target = tensor([True, False, False, False, True, False, True])
+        >>> hr2 = RetrievalHitRate(k=2)
+        >>> hr2(preds, target, indexes=indexes)
         tensor(0.5000)
     """
 
-    higher_is_better = False
+    higher_is_better = True
 
     def __init__(
         self,
-        empty_target_action: str = "pos",
+        empty_target_action: str = "neg",
         ignore_index: Optional[int] = None,
         k: Optional[int] = None,
         compute_on_step: bool = True,
@@ -103,36 +101,5 @@ class RetrievalFallOut(RetrievalMetric):
             raise ValueError("`k` has to be a positive integer or None")
         self.k = k
 
-    def compute(self) -> Tensor:
-        """First concat state `indexes`, `preds` and `target` since they were stored as lists.
-
-        After that, compute list of groups that will help in keeping together predictions about the same query. Finally,
-        for each group compute the `_metric` if the number of negative targets is at least 1, otherwise behave as
-        specified by `self.empty_target_action`.
-        """
-        indexes = torch.cat(self.indexes, dim=0)
-        preds = torch.cat(self.preds, dim=0)
-        target = torch.cat(self.target, dim=0)
-
-        res = []
-        groups = get_group_indexes(indexes)
-
-        for group in groups:
-            mini_preds = preds[group]
-            mini_target = target[group]
-
-            if not (1 - mini_target).sum():
[…]

[…]
