+++ torchmetrics/retrieval/__init__.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from torchmetrics.retrieval.average_precision import RetrievalMAP  # noqa: F401
-from torchmetrics.retrieval.rreciprocal_rank import RetrievalMRR  # noqa: F401
 from torchmetrics.retrieval.fall_out import RetrievalFallOut  # noqa: F401
 from torchmetrics.retrieval.hit_rate import RetrievalHitRate  # noqa: F401
 from torchmetrics.retrieval.metric_base import RetrievalMetric  # noqa: F401
@@ -20,3 +19,4 @@ from torchmetrics.retrieval.ndcg import RetrievalNormalizedDCG  # noqa: F401
 from torchmetrics.retrieval.precision import RetrievalPrecision  # noqa: F401
 from torchmetrics.retrieval.r_precision import RetrievalRPrecision  # noqa: F401
 from torchmetrics.retrieval.recall import RetrievalRecall  # noqa: F401
+from torchmetrics.retrieval.rreciprocal_rank import RetrievalMRR  # noqa: F401
