+++ torchmetrics/utilities/imports.py
@@ -87,6 +87,7 @@ _TORCHVISION_AVAILABLE: bool = _module_available("torchvision")
 _TORCHVISION_GREATER_EQUAL_0_8: Optional[bool] = _compare_version("torchvision", operator.ge, "0.8.0")
 _TQDM_AVAILABLE: bool = _module_available("tqdm")
 _TRANSFORMERS_AVAILABLE: bool = _module_available("transformers")
+_TRANSFORMERS_AUTO_AVAILABLE = _module_available("transformers.models.auto")
 _PESQ_AVAILABLE: bool = _module_available("pesq")
 _SACREBLEU_AVAILABLE: bool = _module_available("sacrebleu")
 _REGEX_AVAILABLE: bool = _module_available("regex")
