+++ torchmetrics/detection/__init__.py
@@ -15,4 +15,4 @@
 from torchmetrics.utilities.imports import _TORCHVISION_AVAILABLE, _TORCHVISION_GREATER_EQUAL_0_8
 
 if _TORCHVISION_AVAILABLE and _TORCHVISION_GREATER_EQUAL_0_8:
-    from torchmetrics.detection.map import MeanAveragePrecision
+    from torchmetrics.detection.map import MeanAveragePrecision  # noqa: F401
