### 2022-01-14 053b10c
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
