### 2022-01-14 feeab60
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
