+++ torchmetrics/retrieval/__init__.py
@@ -19,4 +19,4 @@ from torchmetrics.retrieval.ndcg import RetrievalNormalizedDCG  # noqa: F401
 from torchmetrics.retrieval.precision import RetrievalPrecision  # noqa: F401
 from torchmetrics.retrieval.r_precision import RetrievalRPrecision  # noqa: F401
 from torchmetrics.retrieval.recall import RetrievalRecall  # noqa: F401
-from torchmetrics.retrieval.rreciprocal_rank import RetrievalMRR  # noqa: F401
+from torchmetrics.retrieval.reciprocal_rank import RetrievalMRR  # noqa: F401
