### 2022-01-14 06d820b
rename PESQ (#751)
* perceptual_evaluation_speech_quality
* PerceptualEvaluationSpeechQuality
* chlog

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
