### 2022-01-14 d044d33
rename STOI (#753)
* ShortTermObjectiveIntelligibility
* docs
