### 2022-01-14 1d1b7d5
rename some Image metrics (#752)
* FrechetInceptionDistance
* KernelInceptionDistance
* InceptionScore
* LearnedPerceptualImagePatchSimilarity
* docs & chlog
