### 2022-01-14 32d3cc8
rename MeanAveragePrecision (#754)
* MeanAveragePrecision
* chlog & fmt

Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
