### 2022-01-14 7595120
unify retrieval names (#755)
* unify retrieval module's names between func and class
* base

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>
