### 2022-01-14 07cca4d
enable more PL integrations (#739)
* enable more Pl integrations
* if: success()
* try 1.1
* try 1.2
* try 1.3
* fix imports
* use own sum/diff
* Apply suggestions from code review

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
Co-authored-by: Nicki Skafte Detlefsen <skaftenicki@gmail.com>
