### 2022-01-14 5b3141e
Refactor stoi functional
