### 2022-01-14 43e18c6
short_term_objective_intelligibility
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
