### 2022-01-14 7deadff
Rename`stoi` functional (#758)
* short_term_objective_intelligibility
* changelog

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jirka <jirka.borovec@seznam.cz>
