+++ torchmetrics/functional/audio/stoi.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 import numpy as np
 import torch
+from deprecate import deprecated, void
 
 from torchmetrics.utilities.imports import _PYSTOI_AVAILABLE
 
@@ -22,10 +23,13 @@ else:
     stoi_backend = None
 from torch import Tensor
 
+from torchmetrics.utilities import _future_warning
 from torchmetrics.utilities.checks import _check_same_shape
 
 
-def stoi(preds: Tensor, target: Tensor, fs: int, extended: bool = False, keep_same_device: bool = False) -> Tensor:
+def short_term_objective_intelligibility(
+    preds: Tensor, target: Tensor, fs: int, extended: bool = False, keep_same_device: bool = False
+) -> Tensor:
     r"""STOI (Short Term Objective Intelligibility, see [2,3]), a wrapper for the pystoi package [1].
     Note that input will be moved to `cpu` to perform the metric calculation.
 
@@ -59,12 +63,12 @@ def stoi(preds: Tensor, target: Tensor, fs: int, extended: bool = False, keep_sa
             If ``pystoi`` package is not installed
 
     Example:
-        >>> from torchmetrics.functional.audio.stoi import stoi
+        >>> from torchmetrics.functional.audio.stoi import short_term_objective_intelligibility
         >>> import torch
         >>> g = torch.manual_seed(1)
         >>> preds = torch.randn(8000)
         >>> target = torch.randn(8000)
-        >>> stoi(preds, target, 8000).float()
+        >>> short_term_objective_intelligibility(preds, target, 8000).float()
         tensor(-0.0100)
 
     References:
@@ -103,3 +107,22 @@ def stoi(preds: Tensor, target: Tensor, fs: int, extended: bool = False, keep_sa
         stoi_val = stoi_val.to(preds.device)
 
     return stoi_val
+
+
+@deprecated(target=short_term_objective_intelligibility, deprecated_in="0.7", remove_in="0.8", stream=_future_warning)
+def stoi(preds: Tensor, target: Tensor, fs: int, mode: str, keep_same_device: bool = False) -> Tensor:
+    r"""STOI (Short Term Objective Intelligibility)
+
+    .. deprecated:: v0.7
+        Use :func:`torchmetrics.functional.audio.short_term_objective_intelligibility`. Will be removed in v0.8.
+
+    Example:
+        >>> from torchmetrics.functional.audio.stoi import stoi
+        >>> import torch
+        >>> g = torch.manual_seed(1)
+        >>> preds = torch.randn(8000)
+        >>> target = torch.randn(8000)
+        >>> stoi(preds, target, 8000).float()
+        tensor(-0.0100)
+    """
+    return void(preds, target, fs, mode, keep_same_device)
