+++ torchmetrics/detection/map.py
@@ -24,6 +24,7 @@ from torchmetrics.utilities.imports import _TORCHVISION_AVAILABLE, _TORCHVISION_
 
 if _TORCHVISION_AVAILABLE and _TORCHVISION_GREATER_EQUAL_0_8:
     from torchvision.ops import box_area, box_convert, box_iou
+    __doctest_skip__ = ["MeanAveragePrecision", "MAP"]
 else:
     box_convert = box_iou = box_area = None
 
@@ -197,10 +198,10 @@ class MeanAveragePrecision(Metric):
         ...     labels=torch.IntTensor([0]),
         ...   )
         ... ]
-        >>> metric = MeanAveragePrecision()  # doctest: +SKIP
-        >>> metric.update(preds, target)  # doctest: +SKIP
+        >>> metric = MeanAveragePrecision()
+        >>> metric.update(preds, target)
         >>> from pprint import pprint
-        >>> pprint(metric.compute())  # doctest: +SKIP
+        >>> pprint(metric.compute())
         {'map': tensor(0.6000),
          'map_50': tensor(1.),
          'map_75': tensor(1.),
@@ -761,10 +762,10 @@ class MAP(MeanAveragePrecision):
         ...     labels=torch.IntTensor([0]),
         ...   )
         ... ]
-        >>> metric = MAP()  # doctest: +SKIP
-        >>> metric.update(preds, target)  # doctest: +SKIP
+        >>> metric = MAP()
+        >>> metric.update(preds, target)
         >>> from pprint import pprint
-        >>> pprint(metric.compute())  # doctest: +SKIP
+        >>> pprint(metric.compute())
         {'map': tensor(0.6000),
          'map_50': tensor(1.),
          'map_75': tensor(1.),
