+++ torchmetrics/functional/text/rouge.py
@@ -191,23 +191,24 @@ def _rouge_score_update(
         >>> target = "Is your name John".split()
         >>> from pprint import pprint
         >>> score = _rouge_score_update(preds, target, rouge_keys_values=[1, 2, 3, 'L'], accumulate='best')
-        >>> pprint(score)
+        >>> pprint(score)  # doctest: +NORMALIZE_WHITESPACE
         {1: [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(1.), 'precision': tensor(1.), 'recall': tensor(1.)}],
-        2: [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)}],
-        3: [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)}],
-        'L': [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
-            {'fmeasure': tensor(1.), 'precision': tensor(1.), 'recall': tensor(1.)}]}
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)}],
+         2: [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)}],
+         3: [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+             {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)}],
+         'L': [{'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
+               {'fmeasure': tensor(0.), 'precision': tensor(0.), 'recall': tensor(0.)},
[…]
