+++ torchmetrics/image/__init__.py
@@ -11,7 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from torchmetrics.image.inception import InceptionScore  # noqa: F401
+from torchmetrics.image.inception import IS, InceptionScore  # noqa: F401
+from torchmetrics.image.kid import KID, KernelInceptionDistance  # noqa: F401
 from torchmetrics.image.psnr import PSNR, PeakSignalNoiseRatio  # noqa: F401
 from torchmetrics.image.ssim import (  # noqa: F401
     SSIM,
@@ -21,7 +22,7 @@ from torchmetrics.image.ssim import (  # noqa: F401
 from torchmetrics.utilities.imports import _LPIPS_AVAILABLE, _TORCH_FIDELITY_AVAILABLE
 
 if _TORCH_FIDELITY_AVAILABLE:
-    from torchmetrics.image.fid import FrechetInceptionDistance  # noqa: F401
+    from torchmetrics.image.fid import FID, FrechetInceptionDistance  # noqa: F401
 
 if _LPIPS_AVAILABLE:
-    from torchmetrics.image.lpip import LearnedPerceptualImagePatchSimilarity  # noqa: F401
+    from torchmetrics.image.lpip import LPIPS, LearnedPerceptualImagePatchSimilarity  # noqa: F401
