### 2022-01-14 7a6c514
Apply suggestions from code review
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
