+++ torchmetrics/text/bert.py
@@ -118,17 +118,18 @@ class BERTScore(Metric):
         >>> preds = ["hello there", "general kenobi"]
         >>> target = ["hello there", "master kenobi"]
         >>> bert = BERTScore(model_name_or_path="roberta-base")
-        >>> bert(preds, target)
-        {'precision': [0.99..., 0.99...],
-         'recall': [0.99..., 0.99...],
-         'f1': [0.99..., 0.99...]}
+        >>> bert(preds, target)  # doctest: +NORMALIZE_WHITESPACE +ELLIPSES
+        {'precision': [0.999..., 0.999., 0.987..., 0.987...],
+         'recall': [0.999..., 0.999..., 0.987..., 0.987...],
+         'f1': [0.999..., 0.999..., 0.987..., 0.987...]}
+
     """
 
     higher_is_better = True
 
     def __init__(
         self,
-        model_name_or_path: Optional[str] = None,
+        model_name_or_path: str,
         num_layers: Optional[int] = None,
         all_layers: bool = False,
         model: Optional[torch.nn.Module] = None,
