+++ torchmetrics/text/bert.py
@@ -122,7 +122,6 @@ class BERTScore(Metric):
         {'precision': [0.999..., 0.999., 0.987..., 0.987...],
          'recall': [0.999..., 0.999..., 0.987..., 0.987...],
          'f1': [0.999..., 0.999..., 0.987..., 0.987...]}
-
     """
 
     higher_is_better = True
