+++ .github/workflows/ci_test-base.yml
@@ -19,7 +19,6 @@ jobs:
     env:
       PYTEST_ARTEFACT: pytest-results-${{ matrix.os }}-${{ matrix.python-version }}
       PYTORCH_URL: https://download.pytorch.org/whl/cpu/torch_stable.html
-      TRANSFORMERS_CACHE: .cache/huggingface/
 
     # Timeout: https://stackoverflow.com/a/59076067/4521646
     timeout-minutes: 20
@@ -65,12 +64,6 @@ jobs:
         pip uninstall -y torchmetrics
         pip list
       shell: bash
-
-    - name: HF cache
-      uses: actions/cache@v2
-      with:
-        path: $TRANSFORMERS_CACHE
-        key: cache-transformers
 
     - name: Test Package [only]
       run: |
