### 2022-01-15 7c2b56e
Fix BERTScore default values + peculiar module behaviour
