+++ setup.cfg
@@ -4,9 +4,9 @@ norecursedirs =
     .github
     dist
     build
+doctest_plus = enabled
 addopts =
     --strict
-    --doctest-modules
     --color=yes
 
 [coverage:run]
