+++ setup.cfg
@@ -4,10 +4,15 @@ norecursedirs =
     .github
     dist
     build
-doctest_plus = enabled
+#doctest_plus = enabled
 addopts =
     --strict
+    --doctest-plus
     --color=yes
+doctest_optionflags =
+    NORMALIZE_WHITESPACE
+    ELLIPSIS
+    FLOAT_CMP
 
 [coverage:run]
 parallel = True
