### 2022-01-16 b84a110
Apply suggestions from code review
Co-authored-by: Rohit Gupta <rohitgr1998@gmail.com>
