### 2022-01-16 b1d0a7a
Fix BERTScore default values + peculiar module behaviour (#760)
