### 2022-01-16 c0e9062
Apply suggestions from code review
Co-authored-by: Daniel Stancl <46073029+stancld@users.noreply.github.com>
