### 2022-01-26 #4008 @richrobe
Hi @taless474!

I would be happy to review this work for JOSS, but I'm not that familiar with pytorch. However, a colleague of mine, @reneraab, is very familiar with pytorch, but has not yet submitted or reviewed for JOSS, so he's not that familiar with the reviewing process. Would it be possible to do the review together?
