### 2022-01-26 #4008 @Borda
> but I'm not that familiar with pytorch

I would say that `pyTorch` is rather a tool than the core that paper aims at... :rabbit:
