### 2022-01-27 #4008 @taless474
> Hi @taless474!
>
> I would be happy to review this work for JOSS, but I'm not that familiar with pytorch. However, a colleague of mine, @reneraab, is very familiar with pytorch, but has not yet submitted or reviewed for JOSS, so he's not that familiar with the reviewing process. Would it be possible to do the review together?

@richrobe, thank you for your response. I believe we can have 3 reviewers for this work, so I added @reneraab. @reneraab, thank you for offering to review. Please ask us if you have any questions.
