### 2022-01-27 #4101 @whedon
[REVIEW]: TorchMetrics - Measuring Reproducibility in PyTorch
**Submitting author:** @justusschock (<a href="http://orcid.org/0000-0003-0512-3053">Justus Schock</a>)
**Repository:** <a href="https://github.com/PyTorchLightning/metrics" target ="_blank">https://github.com/PyTorchLightning/metrics</a>
**Version:** v0.7.2
