### 2022-01-28 PyTorchLightning/metrics#807 @inpefess
The authors affiliation info is incomplete in paper.md
For more details see https://github.com/openjournals/joss-reviews/issues/4101#issuecomment-1022812990
