### 2022-01-28 #4101 @inpefess
'Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.' --> https://github.com/PyTorchLightning/metrics/issues/808
