### 2022-01-28 PyTorchLightning/metrics#809 @inpefess
Missing DOI in paper.md
For more details please look [here](https://github.com/openjournals/joss-reviews/issues/4101#issuecomment-1022813164) and [here](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax)
