### 2022-01-28 PyTorchLightning/metrics#808 @Borda
I think the problem comes from PyTorch dependency as they have not officially released wheels for python 3.10 so for this python version you would need to compile PyTorch from the source... the binaries for python 3.10 are expected for PT 1.11

see: https://github.com/pytorch/pytorch/issues/66424
