### 2022-01-28 #4101 @Borda
> 'Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.' --> [PyTorchLightning/metrics#808](https://github.com/PyTorchLightning/metrics/issues/808)

yes, the problem seems to be related to PyTorch as they do not provide binaries for python 3.10 yet, and moreover, they suggest using their package warehouse... on the other hand, all shall be well working in the Conda environment >> https://anaconda.org/conda-forge/torchmetrics

cc: @justusschock
