### 2022-01-28 PyTorchLightning/metrics#808 @inpefess
So, `torchmetrics` doesn't support Python 3.10 officially, although there is a badge claiming that on the front page. PyTorchLightning doesn't claim 3.10 support, for example, and `torch` documentation recommends using 3.6, 3.7, or 3.8.

I suggest either dropping the Python 3.10 support claim or adding specific 3.10 installation instructions to the documentation as a form of weak support.
