### 2022-01-28 PyTorchLightning/metrics#808 @justusschock
@inpefess you are correct. For now, we certainly need to remove the badge.
