### 2022-01-28 PyTorchLightning/metrics#808 @justusschock
@inpefess I opened a PR to fix this, the badge will only update after the next release. However, we want to wait with that one for all reviews. Are you fine with that? Can you install it with python <3.10?
