### 2022-01-28 #4101 @justusschock
@taless474 is there a way to have authors without affiliations (@maximsch2 chose not to have an affiliation here)?
