### 2022-01-28 PyTorchLightning/metrics#807 @justusschock
Hi @inpefess ,

Yes the author @maximsch2 chose not to have an affiliation. I asked the editor how to handle that (see https://github.com/openjournals/joss-reviews/issues/4101#issuecomment-1024121296)
