### 2022-01-28 PyTorchLightning/metrics#809 @justusschock
Hi, Thanks a lot, I did look into this before and the missing doi is completely unrelated to the software it links it to.

The doi seems to be part of an encyclopaedia explaining the term `keras`. The software is cited as suggested by the authors [here](https://keras.io/getting_started/faq/#how-should-i-cite-keras)
