### 2022-01-28 PyTorchLightning/metrics#808 @Borda
> @inpefess you are correct. For now, we certainly need to remove the badge.

`torchmetrics` does but Pytorch does not provide wheels for it so when PT releases we would have it but if we change the tag we can't add it back so it won't be correct in two months' time that TM does not support 3.10 when it does...
