### 2022-01-28 PyTorchLightning/metrics#808 @inpefess
> @inpefess I opened a PR to fix this, the badge will only update after the next release. However, we want to wait with that one for all reviews. Are you fine with that? Can you install it with python <3.10?

Sure, I've already done that! Thank you.

> ok, I'll add docs for building TM with PT from source as this is not limitation of TM but PT

Yes, I think it's not a question of tags but the documentation. If it mentions that for Python 3.10, one should use Anaconda, no one will probably try doing something like I did. But I noticed the badge, executed the first line from the docs, and failed.
