### 2022-01-28 #4101 @taless474
@justusschock yes, authors without affiliations are allowed
