### 2022-01-28 #4101 @danielskatz
No, they are not. See affiliation 3 in https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography for how to handle this.
