### 2022-01-29 #4101 @taless474
Thank you @danielskatz
@justusschock I think they can go with "Independent Researcher" as shown above
