### 2022-01-31 PyTorchLightning/metrics#807 @justusschock
Fixed by #817
