### 2022-02-01 #4101 @Borda
FYI, both changes https://github.com/PyTorchLightning/metrics/pull/810 and https://github.com/PyTorchLightning/metrics/pull/817 were cherry-picked and will be released in the next `v0.7.1` in meantime you can install the stable from branch `release/0.7.x`
