### 2022-02-01 #4101 @reneraab
The software paper gives a good introduction to TorchMetrics and contains all the required sections. However, overall it may be made slightly more accessible for non-specialist audiences. Since this is my first JOSS review, I can't quite judge how accessible JOSS papers typically are and would leave this judgment to the other reviewers. In terms of language I would suggest the following changes:

line 16 in PDF (56 in paper.md): Change "an code and data section" to "a code and data section"
line 16-17 (56): "that links to both official and community code to papers". Remove "to" before both (or rephrase the sentence).
line 17 (56): Change "on the paper code made publicly accessible" to "on the paper code to be made publicly accessible"
line 19 (56): Change "non opensource work" to "closed-source work" or "proprietary work"

line 54 (96): "trained on hundreds of devices GPUs or TPUs". Either remove "GPUs or TPUs" or add "such as" before "GPUs": "trained on hundreds of devices such as GPUs or TPUs".

line 60 (98): The first sentence introduces "a functional interface" and the next sentence uses "they" to refer to the functions that this interface provides. This is not immediately obvious language-wise. Consider rephrasing the second sentence to "This interface provides simple Python functions that"
line 60 (98): Same sentence: capitalize "python functions" to "Python functions.
