### 2022-02-01 #4101 @Borda
@reneraab thank you for your kind suggestions, all shall be implemented in https://github.com/PyTorchLightning/metrics/pull/819
