### 2022-02-09 #4101 @Borda
> @Borda Would you please take care of the missing DOI?

we are already generating DOI for each release since `v0.7.0` with Zenodo as it is referred in [docs](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content): https://zenodo.org/account/settings/github/repository/PyTorchLightning/metrics
Do you find it sufficient, otherwise I'll sync with @justusschock to complete it :rabbit:
