### 2022-02-09 #4101 @taless474
>  [docs](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content): https://zenodo.org/account/settings/github/repository/PyTorchLightning/metrics
This link doesn't work for me; says Permission Required. Would you please modify it?
