### 2022-02-09 #4101 @Borda
> > [docs](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content): https://zenodo.org/account/settings/github/repository/PyTorchLightning/metrics
> > This link doesn't work for me; says Permission Required. Would you please modify it?

I have exported the badge [![DOI](https://zenodo.org/badge/323721661.svg)](https://zenodo.org/badge/latestdoi/323721661), mind it try now? and we add it also to readme
