### 2022-02-09 #4101 @justusschock
> @Borda Would you please take care of the missing DOI?

@taless please have a look on https://github.com/PyTorchLightning/metrics/issues/809 about why the doi on keras is wrongly stated as missing

> Good! Would you please add the reference for PyTorch too?

Sorry, this must have been lost during fixing other review comments. Addressed in https://github.com/PyTorchLightning/metrics/pull/830
