### 2022-02-09 #4101 @Borda
> Sorry, this must have been lost during fixing other review comments. Addressed in [PyTorchLightning/metrics#830](https://github.com/PyTorchLightning/metrics/pull/830)

@taless474 it is merged, so shall be fixed now :rabbit:
