### 2022-02-10 #4101 @taless474
@justusschock just one thing, the zenodo archive has a different title and authors than the paper. Typically, these are the same for a JOSS-related paper. Note that you can manually change the metadata of the zenodo archive.
