### 2022-02-10 #4101 @Borda
The zendo points to the latest release so it seems that the title will always change according to the release name which is rather dependent on the particular release content... also for having the last edit from https://github.com/PyTorchLightning/metrics/pull/830 in a release we would need to make one more `v0.7.2` which we can then name exactly as the paper title... :rabbit:
