### 2022-02-10 #4101 @kthyng
Hi @justusschock! A few items to wrap up:

1. version and archive should be updated now ✅
2. paper:
- paragraph after the code on pg 2.: the references should be inline not parenthetical, so remove the [] around them.
- can you check your references for correct capitalization? I'm guessing Heusel et al should have "nash" capitalized since isn't that a name? You can preserve capitalization by putting {} around characters in your .bib file.
