### 2022-02-11 #4101 @justusschock
Hi @kthyng , Thanks a lot!

With your points you are completely correct and they have been addressed in https://github.com/PyTorchLightning/metrics/pull/836 .

Anything else we need to do on this?
